% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{a01data}
\alias{a01data}
\title{Data for an alpha (0,1) design}
\format{
A data frame with 60 rows and 4 columns.
}
\source{
International Potato Center, sweetpotato experimental data.
}
\usage{
a01data
}
\description{
This data set contains data for an alpha (0,1) design.
}
