% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_1env.R
\docType{data}
\name{alpha01data}
\alias{alpha01data}
\title{Data for an alpha (0,1) design}
\format{A data frame.}
\source{
International Potato Center, sweetpotato experimental data.
}
\usage{
alpha01data
}
\description{
This data set contains data for al alpha (0,1) design.
}
