% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ammi_gxe.R
\name{ammi.gxe}
\alias{ammi.gxe}
\title{AMMI or GGE with data from an interaction means matrix}
\usage{
ammi.gxe(
  int.mean,
  trait.name = NULL,
  method = c("ammi", "gge"),
  f = 0.5,
  aov.model = NULL,
  nrep = NULL
)
}
\arguments{
\item{int.mean}{GxE means matrix, genotypes in rows, environments in columns.}

\item{trait.name}{The name of the column for the trait to analyze.}

\item{method}{\code{"ammi"} or \code{"gge"}.}

\item{f}{Scaling factor, defaults to 0.5.}

\item{aov.model}{Analysis of variance model.}

\item{nrep}{Number of replications.}
}
\value{
It returns an object of class \code{ammi} with the overall, genotype,
environment and interaction means, the interaction effects matrix, the
first and second PC values for genotypes and environments, and a table
with the contribution of each PC. ANOVA table and significance of PCs are
included only if \code{method = "ammi"}, and \code{aov.model} and \code{nrep}
are specified.
}
\description{
This function runs AMMI (Gollob, H. R., 1968) or GGE (Yan , W. et al., 2000)
with data from an interaction means matrix.
}
\details{
Significance of PCs are evaluated only with \code{method = "ammi"}
and if \code{aov.model} and \code{nrep} are specified.
}
\examples{
# Compute GxE means
int.mean <- tapply(met8x12$y, list(met8x12$geno, met8x12$env), mean, na.rm = TRUE)
# Run AMMI with GxE means matrix
model.ammi <- ammi.gxe(int.mean, trait.name = "y")
model.ammi
# Run GGE with GxE means matrix
model.gge <- ammi.gxe(int.mean, trait.name = "y", method = "gge")
model.gge
}
\references{
Gollob, H. R. (1968). A Statistical Model which combines Features of Factor Analytic
and Analysis of Variance Techniques, Psychometrika, Vol 33(1): 73-114.

Yan, W. et al. (2000). Cultivar evaluation and mega-environment investigation based on the GGE
biplot, Crop Sci., Vol 40: 597-605.
}
\seealso{
\code{svd}
}
\author{
Raul Eyzaguirre.
}
