% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ammi.R
\name{ammi}
\alias{ammi}
\title{AMMI or GGE with data at plot level}
\usage{
ammi(dfr, y, geno, env, rep, method = c("ammi", "gge"), f = 0.5, maxp = 0.1)
}
\arguments{
\item{dfr}{The name of the data frame.}

\item{y}{The name of the column for the variable to analyze.}

\item{geno}{The name of the column that identifies the genotypes.}

\item{env}{The name of the column that identifies the environments.}

\item{rep}{The name of the column that identifies the replications or blocks. A RCBD is assumed.}

\item{method}{\code{"ammi"} or \code{"gge"}.}

\item{f}{Scaling factor, defaults to 0.5.}

\item{maxp}{Maximum allowed proportion of missing values to estimate, default is 10\%.}
}
\value{
It returns an object of class \code{ammi} with the overall, genotype,
environment and interaction means, the interaction effects matrix, the
first and second PC values for genotypes and environments, the ANOVA table,
and a table with the contribution of each PC. Significance of PCs are
included only if \code{method = "ammi"} and the data are balanced
(with missing values estimated if possible).
}
\description{
This function runs AMMI (Gollob, H. R., 1968) or GGE (Yan , W. et al., 2000)
with data at plot level.
}
\details{
Significance of PCs are evaluated only with \code{method = "ammi"}
and if the data are balanced (with missing values estimated if possible).
}
\examples{
model.ammi <- ammi(met8x12, "y", "geno", "env", "rep")
model.ammi
model.gge <- ammi(met8x12, "y", "geno", "env", "rep", "gge")
model.gge
}
\references{
Gollob, H. R. (1968). A Statistical Model which combines Features of Factor Analytic
and Analysis of Variance Techniques, Psychometrika, Vol 33(1): 73-114.

Yan, W. et al. (2000). Cultivar evaluation and mega-environment investigation based
on the GGE biplot, Crop Sci., Vol 40: 597-605.
}
\seealso{
\code{svd}
}
\author{
Raul Eyzaguirre.
}
