% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/AMMI.R
\name{ammigxe}
\alias{ammigxe}
\title{AMMI or GGE with data from an interaction means matrix}
\usage{
ammigxe(int.mean, trait = NULL, rep.num = NULL, rdf = NULL, rms = NULL,
  method = "AMMI", f = 0.5, biplot = 1, biplot1 = "effects",
  title = NULL, xlab = NULL, color = c("darkorange", "black", "gray"),
  ...)
}
\arguments{
\item{int.mean}{GxE means matrix, genotypes in rows, environments in columns.}

\item{trait}{Name of the trait.}

\item{rep.num}{Number of replications.}

\item{rdf}{Residual degrees of freedom.}

\item{rms}{Residual mean square.}

\item{method}{AMMI or GGE.}

\item{f}{Scaling factor, defaults to 0.5.}

\item{biplot}{1 for the trait-PC1 biplot and 2 for the PC1-PC2 biplot.}

\item{biplot1}{Choose "effects" or "means" for biplot1.}

\item{title}{Main title for biplot1 or biplot2.}

\item{xlab}{Xlab for biplot1.}

\item{color}{Color for lines, symbols and/or labels for environments, genotypes and axes.}

\item{...}{Additional graphic parameters.}
}
\value{
It returns the first and second PC values for genotypes and environments,
a table with the contribution of each PC, a dispersion plot of means or effects
against the first PC, or a dispersion plot of PC1 against PC2.
Significance of PCs are included in the contributions table only if \code{rep.num},
\code{rms} and \code{rdf} are specified and method is set to AMMI.
}
\description{
This function runs AMMI (Gollob, H. R., 1968) or GGE biplot (Yan , W. et al., 2000)
with data from an interaction means matrix.
}
\details{
Significance of PCs are evaluated only with \code{method = "AMMI"} and if
\code{rep.num}, \code{rms} and \code{rdf} are specified.
}
\examples{
# The data
head(met8x12)
str(met8x12)

# Compute GxE means
int.mean <- tapply(met8x12$y, list(met8x12$geno, met8x12$env), mean, na.rm = TRUE)

# Run AMMI with GxE means matrix, biplot2
ammigxe(int.mean, trait = "y", biplot = 2)

# Run GGE with GxE means matrix, biplot2
ammigxe(int.mean, trait = "y", method = "GGE", biplot = 2)
}
\author{
Raul Eyzaguirre
}
\references{
Gollob, H. R. (1968). A Statistical Model which combines Features of Factor Analytic
and Analysis of Variance Techniques, Psychometrika, Vol 33(1): 73-114.

Yan, W. et al. (2000). Cultivar evaluation and mega-environment investigation based on the GGE
biplot, Crop Sci., Vol 40: 597-605.
}
\seealso{
\code{svd}
}

