% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov_f.R
\name{aov.f}
\alias{aov.f}
\title{ANOVA for a factorial experiment}
\usage{
aov.f(dfr, y, factors, rep = NULL, maxp = 0.1)
}
\arguments{
\item{dfr}{The name of the data frame.}

\item{y}{The name of the column for the variable to analyze.}

\item{factors}{The names of the columns that identify the factors.}

\item{rep}{The name of the column that identifies the replications or blocks,
default is \code{NULL} for a CRD.}

\item{maxp}{Maximum allowed proportion of missing values to estimate, default is 10\%.}
}
\value{
It returns the ANOVA table.
}
\description{
Fit an analysis of variance model for a factorial experiment with a CRD or RCBD.
}
\examples{
aov.f(asc, "asc.dw", c("geno", "treat"))
}
\author{
Raul Eyzaguirre.
}
