% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov_lxt.R
\name{aov.lxt}
\alias{aov.lxt}
\title{Analysis for a Line x Tester genetic design.}
\usage{
aov.lxt(trait, line, tester, rep, dfr)
}
\arguments{
\item{trait}{The trait to analyze.}

\item{line}{The lines.}

\item{tester}{The testers.}

\item{rep}{The replication.}

\item{dfr}{The name of the data frame.}
}
\value{
It returns the analysis for the line by tester genetic design.
}
\description{
This is a modified version of the \code{"lineXtester"} function of package
\code{"agricolae"} for the analysis of a Lina x Tester genetic design.
}
\details{
Data must correspond with a line by tester design and must be balanced.
If there are missing values they are not estimated and results can be dubious.
}
\examples{
aov.lxt("yield", "line", "tester", "rep", lxt)
}
\author{
Raul Eyzaguirre.
}
