% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov_met.R
\name{aov.met}
\alias{aov.met}
\title{ANOVA for MET with a RCBD}
\usage{
aov.met(trait, geno, env, rep, dfr, maxp = 0.1)
}
\arguments{
\item{trait}{The name of the column for the trait to analyze.}

\item{geno}{The name of the column that identifies the genotypes.}

\item{env}{The name of the column that identifies the environments.}

\item{rep}{The name of the column that identifies the replications or blocks.}

\item{dfr}{The name of the data frame.}

\item{maxp}{Maximum allowed proportion of missing values to estimate, default is 10\%.}
}
\value{
It returns the ANOVA table.
}
\description{
Fit an analysis of variance model for a multi environment trial (MET) with a RCBD
in each environment.
}
\details{
If data is unbalanced, missing values are estimated up to an specified maximum
proportion, 10\% by default. Genotypes and environments are considered as fixed
factors while the blocks are considered as random and nested into the environments.
}
\examples{
aov.met("y", "geno", "env", "rep", met8x12)
}
\author{
Raul Eyzaguirre.
}
