% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov_rcbd.R
\name{aov.rcbd}
\alias{aov.rcbd}
\title{ANOVA for a RCBD}
\usage{
aov.rcbd(dfr, y, geno, rep, maxp = 0.1)
}
\arguments{
\item{dfr}{The name of the data frame.}

\item{y}{The name of the column for the variable to analyze.}

\item{geno}{The name of the column that identifies the genotypes.}

\item{rep}{The name of the column that identifies the replications.}

\item{maxp}{Maximum allowed proportion of missing values to estimate, default is 10\%.}
}
\value{
It returns an ANOVA table.
}
\description{
Fit an analysis of variance model for a RCBD.
}
\details{
If data is unbalanced, missing values are estimated up to an specified maximum
proportion, 10\% by default.
}
\examples{
# Get a copy with some missing values for trw and run ANOVA
tmp <- pjpz09
tmp[c(10, 20, 30), "trw"] <- NA
aov.rcbd(tmp, "trw", "geno", "rep")
}
\author{
Raul Eyzaguirre.
}
