% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aovmet.R
\name{aovmet}
\alias{aovmet}
\title{ANOVA for MET with a RCBD}
\usage{
aovmet(trait, geno, env, rep, data, maxp = 0.1, anova = TRUE)
}
\arguments{
\item{trait}{The trait to analyze.}

\item{geno}{The genotypes.}

\item{env}{The environments.}

\item{rep}{The replications or blocks.}

\item{data}{The name of the data frame containing the data.}

\item{maxp}{Maximum allowed proportion of missing values to estimate, default is 10\%.}

\item{anova}{Logical, if TRUE the ANOVA table is shown.}
}
\value{
If \code{anova} is {TRUE} it returns and shows the ANOVA table.
If \code{anova} is {FALSE} it returns the ANOVA table and some other components as the
estimated missing values, but nothing is shown.
}
\description{
Fit an analysis of variance model for a multi environment trial (MET) with a RCBD
in each environment.
}
\details{
If data is unbalanced, missing values are estimated up to an specified maximum
proportion, 10\% by default. Genotypes and environments are considered as fixed
factors while the blocks are considered as random and nested into the environments.
}
\examples{
# The data
head(met8x12)
str(met8x12)

# Run ANOVA for MET
aovmet("y", "geno", "env", "rep", met8x12)
}
\author{
Raul Eyzaguirre
}

