% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audpc.R
\name{audpc}
\alias{audpc}
\title{Area Under the Disease Progress Curve}
\usage{
audpc(evals, dates, dfr)
}
\arguments{
\item{evals}{The names of the columns for the evaluations.}

\item{dates}{A vector with the dates of evaluation.}

\item{dfr}{The name of the data frame.}
}
\value{
It returns the original data frame with added columns for AUDPC and rAUDPC.
}
\description{
Computes the absolute (AUDPC) or relative (rAUDPC) area under the disease progress curve.
}
\details{
Evaluations are subjective ranging from 0 (no disease) to 100 (full disease).
Dates should follow the format YYYY-MM-DD.
\code{audpc} is expressed in days x % and \code{raudpc} as a proportion.
}
\examples{
head(lbb3c3)
evals <- c("lb1", "lb2", "lb3", "lb4", "lb5", "lb6")
dates <- c("2017-11-20", "2017-11-27", "2017-12-04",
           "2017-12-11", "2017-12-18", "2017-12-25")
tmp <- audpc(evals, dates, lbb3c3)
head(tmp)              
}
\author{
Raul Eyzaguirre.
}
