% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm2bb.R
\name{bm2ptb}
\alias{bm2ptb}
\title{Biomart to potatobase format}
\usage{
bm2ptb(file.name)
}
\arguments{
\item{file.name}{The name of the biomart Excel fieldbook.}
}
\value{
It returns a design file and a data file ready to upload in potatobase.
}
\description{
Converts biomart Excel files to potatobase formats.
}
\details{
All labels (lower or upper case) listed in function \code{check.names.pt}
are converted to CO variable numbers.
}
\author{
Raul Eyzaguirre.
}
