% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_2fcr.R
\name{cd.2fcr}
\alias{cd.2fcr}
\title{Completely Randomized Design for a two-factor factorial}
\usage{
cd.2fcr(A, B, nrep, nc)
}
\arguments{
\item{A}{The levels of factor A.}

\item{B}{The levels of factor B.}

\item{nrep}{Number of replications.}

\item{nc}{Number of columns.}
}
\value{
It returns the fieldbook and fieldplan.
}
\description{
This function creates the fieldbook and fieldplan for a two-factor
factorial with a CRD.
}
\details{
The treatments are randomly allocated on a field following a CRD.
}
\examples{
A <- paste("a", 1:5, sep = "")
B <- paste("b", 1:3, sep = "")
cd.2fcr(A, B, 3, 12)
cd.2fcr(A, B, 2, 7)
}
\author{
Raul Eyzaguirre.
}
