% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_2frcb.R
\name{cd.2frcb}
\alias{cd.2frcb}
\title{Randomized Complete Block Design for a two-factor factorial}
\usage{
cd.2frcb(A, B, nb, nc)
}
\arguments{
\item{A}{The levels of factor A.}

\item{B}{The levels of factor B.}

\item{nb}{Number of blocks.}

\item{nc}{Number of columns.}
}
\value{
It returns the fieldbook and fieldplan.
}
\description{
This function creates the fieldbook and fieldplan for a two-factor
factorial with a RCBD.
}
\details{
The treatments are randomly allocated on a field following a RCBD.
The blocks are disposed alongside the rows.
}
\examples{
A <- paste("a", 1:5, sep = "")
B <- paste("b", 1:3, sep = "")
cd.2frcb(A, B, 3, 9)
cd.2frcb(A, B, 2, 7)
}
\author{
Raul Eyzaguirre.
}
