% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_ab.R
\name{cd.ab}
\alias{cd.ab}
\title{Augmented Block Design}
\usage{
cd.ab(geno, checks, nb, nc)
}
\arguments{
\item{geno}{The list of genotypes.}

\item{checks}{The list of checks.}

\item{nb}{Number of blocks.}

\item{nc}{Number of columns.}
}
\value{
It returns the fieldbook and fieldplan.
}
\description{
This function creates the fieldbook and fieldplan for an ABD
}
\details{
The genotypes are randomly allocated on a field following an ABD.
The blocks are disposed alongside the rows.
}
\examples{
checks <- paste("ch", 1:4, sep = "_")
cd.ab(1:40, checks, 4, 10)
cd.ab(1:50, checks, 3, 7)
}
\author{
Raul Eyzaguirre.
}
