% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_crd.R
\name{cd.cr}
\alias{cd.cr}
\title{Completely Randomized Design}
\usage{
cd.cr(geno, nrep, nc)
}
\arguments{
\item{geno}{The list of genotypes.}

\item{nrep}{Number of replications.}

\item{nc}{Number of columns.}
}
\value{
It returns the fieldbook and fieldplan.
}
\description{
This function creates the fieldbook and fieldplan for a CRD.
}
\details{
The genotypes are randomly allocated on a field following a CRD.
}
\examples{
cd.cr(1:20, 3, 12)
cd.cr(1:20, 2, 7)
}
\author{
Raul Eyzaguirre.
}
