% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_str.R
\name{cd.str}
\alias{cd.str}
\title{Strip-Split-Plot Design}
\usage{
cd.str(A, B, nrep)
}
\arguments{
\item{A}{The levels of factor A (row factor).}

\item{B}{The levels of factor B (column factor).}

\item{nrep}{Number of replications (or blocks).}
}
\value{
It returns the fieldbook and fieldplan.
}
\description{
This function creates the fieldbook and fieldplan for a Strip-Split-Plot design.
}
\details{
The levels of the factors are randomly allocated on a field
following a Strip-Split-Plot design. Row and column numbers are specific
to each replication. Each replication is a complete block for factor A
and for factor B.
}
\examples{
A <- paste("a", 1:4, sep = "")
B <- paste("b", 1:3, sep = "")
cd.str(A, B, 3)
}
\author{
Raul Eyzaguirre.
}
