% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_unrep.R
\name{cd.ur}
\alias{cd.ur}
\title{Unreplicated design}
\usage{
cd.ur(geno, nc)
}
\arguments{
\item{geno}{The list of genotypes.}

\item{nc}{Number of columns.}
}
\value{
It returns the fieldbook and fieldplan.
}
\description{
This function creates the fieldbook and fieldplan for an unreplicated design
with genotypes randomly allocated on a field.
}
\details{
The genotypes are randomly allocated on a field.
}
\examples{
cd.ur(1:100, 5)
cd.ur(1:100, 7)
}
\author{
Raul Eyzaguirre.
}
