% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_wescott.R
\name{cd.w}
\alias{cd.w}
\title{Unreplicated design with a grid of checks}
\usage{
cd.w(geno, ch1, ch2, nc, ncb = 10)
}
\arguments{
\item{geno}{The list of genotypes.}

\item{ch1}{Name of check 1.}

\item{ch2}{Name of check 2.}

\item{nc}{Number of columns.}

\item{ncb}{Number of columns between two check columns (default is 10).}
}
\value{
It returns the fieldbook and fieldplan.
}
\description{
This function creates the fieldbook and fieldplan for an unreplicated design
with genotypes randomly allocated on a field with checks following
the method described on Westcott (1981).
}
\details{
The genotypes are randomly allocated on a field between equally spaced
columns of two alternating check varieties. Check columns are planted each
\code{ncb} columns where \code{ncb} must be even. The specified total number of
columns \code{nc} is the maximum available number, the actual number could be less.
}
\examples{
cd.w(1:100, "A", "B", 100)
}
\references{
Westcott, B. (1981). Two methods for early generation yield assessment in winter wheat.
In: Proc. of the 4th meeting of the Biometrics in Plant Breeding Section of Eucarpia.
INRA Poitier, France, pp 91-95.
}
\author{
Raul Eyzaguirre.
}
