% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_ds.R
\name{cds}
\alias{cds}
\title{Collapse design}
\usage{
cds(
  book,
  nc = NULL,
  fillby = c("rows", "columns"),
  serpentine = c("yes", "no")
)
}
\arguments{
\item{book}{The fieldbook design with columns \code{plot}, \code{block}
and \code{geno}.}

\item{nc}{Number of columns in the field.}

\item{fillby}{Allocate the plots by \code{"rows"} or \code{"columns"},
default \code{"rows"}.}

\item{serpentine}{\code{"yes"} or \code{"no"}, default \code{"yes"}.}
}
\value{
It returns the fieldbook and fieldplan.
}
\description{
Collapse a blocks design to remove missing plots.
}
\details{
If \code{"nc"} is \code{"NULL"}, the function will look for the
most squared rectangular field.
}
\examples{
checks <- paste("ch", 1:4, sep = "_")
genos <- paste("g", 1:20, sep = "_")
book <- cr.abd(genos, checks, 4, 5)$book
cds(book, 5, 'rows', 'yes')
}
\author{
Raul Eyzaguirre.
}
