% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdt.R
\name{cdt}
\alias{cdt}
\title{Compute derived traits}
\usage{
cdt(fb, plot.size = NULL)
}
\arguments{
\item{plot.size}{Plot size in square meters.}

\item{data}{The name of the data frame.}
}
\value{
It returns a data frame with the original and derived traits.
}
\description{
Compute derived traits for a given fieldbook.
}
\details{
The data frame must use the labels (lower or upper case) listed in function \code{checknames}.
See \code{?checknames} for details.
}
\examples{
 # The data
 head(pjpz09)
 str(pjpz09)

 # Compute derived traits
 cdt(pjpz09, 4.5)
}
\author{
Raul Eyzaguirre.
}

