% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check.design}
\alias{check.design}
\title{Check design}
\usage{
check.design(trait, geno, env = NULL, rep, data)
}
\arguments{
\item{trait}{The trait to analyze.}

\item{geno}{Genotypes.}

\item{env}{Environments.}

\item{rep}{The replications.}

\item{data}{The name of the data frame.}
}
\value{
Information about the balance, missing values, and replications of the design.
}
\description{
Check frequencies for designs with complete replications and one or two factors.
}
\author{
Raul Eyzaguirre.
}
