% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_freq.R
\name{check.freq}
\alias{check.freq}
\title{Check frequencies for complete replications}
\usage{
check.freq(dfr, y, geno = "geno", rep = "rep", env = NULL)
}
\arguments{
\item{dfr}{The name of the data frame.}

\item{y}{The name of the column for the variable to analyze.}

\item{geno}{The name of the column that identifies the genotypes.}

\item{rep}{The name of the column that identifies the replications.}

\item{env}{The name of the column that identifies the environments,
\code{NULL} if there are no environments.}
}
\value{
Information about the balance, missing values, and replications of the design.
}
\description{
Check frequencies for designs with complete replications and one or several
environments. This is a wrapper for \code{ck.rcbd} and \code{ck.f} functions.
}
\examples{
check.freq(pjpz09, "trw", "geno", "rep")
check.freq(megaclones, "rytha", "geno", "rep", "env")
}
\author{
Raul Eyzaguirre.
}
