% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_genos.R
\name{check.genos}
\alias{check.genos}
\title{Check genotype names for potato and sweetpotato}
\usage{
check.genos(dfr, geno, crop = c("auto", "pt", "sp", "uk"))
}
\arguments{
\item{dfr}{The name of the data frame.}

\item{geno}{The name of the column in \code{dfr} that identifies the genotypes.}

\item{crop}{\code{"auto"} for autodetection or \code{"pt"} for potato,
\code{"sp"} for sweetpotato and \code{"uk"} for unknown.}
}
\value{
It returns a list with all the genotypes with cultivar name or
breeder code.
}
\description{
Check that genotype's names correspond to CIP numbers instead of variety names
or breeder codes.
}
\details{
It checks for genotypes with cultivar names or breeder codes that
have a CIP number.
}
\examples{
check.genos(pjpz09, "geno")
}
\author{
Raul Eyzaguirre.
}
