% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckData.R
\name{check.met}
\alias{check.met}
\title{Check data for a MET in a RCBD}
\usage{
check.met(trait, geno, env, rep, data)
}
\arguments{
\item{trait}{The trait to analyze.}

\item{geno}{The genotypes.}

\item{env}{The environments.}

\item{rep}{The replications.}

\item{data}{The name of the data frame.}
}
\value{
Three control values (\code{c1}, \code{c2}, and \code{c3}), the number of
missing values \code{nmis}, the proportion of missing values (\code{pmis}), the number
of genotypes (\code{ng}), the number of environments (\code{ne}), and the number of
replications (\code{nr}).
}
\description{
This function checks the frequencies of genotypes in each environment in a RCBD.
}
\details{
This function checks if there is more than one replication in a RCBD in
several environments, if there is any genotype without data for some specific environments,
and if the design is balanced.
}
\author{
Raul Eyzaguirre.
}

