% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_names_pt.R
\name{check.names.pt}
\alias{check.names.pt}
\title{Check fieldbook traits names for potato}
\usage{
check.names.pt(dfr, add = NULL)
}
\arguments{
\item{dfr}{The name of the data frame.}

\item{add}{Additional traits.}
}
\value{
It returns a data frame with all traits names in lower case, and a list of the
traits with names not included in the list shown above.
}
\description{
Check that fieldbook traits names correspond with the names defined in 
\url{http://www.cropontology.org/ontology/CO_330/Potato}
}
\examples{
check.names.pt(potatoyield)
}
\author{
Raul Eyzaguirre, Johan Ninanya.
}
