% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_numeric.R
\name{check.numeric}
\alias{check.numeric}
\title{Check fieldbook variables are numeric}
\usage{
check.numeric(dfr, crop = c("auto", "pt", "sp"), add = NULL, checknames = TRUE)
}
\arguments{
\item{dfr}{The name of the data frame.}

\item{crop}{\code{"auto"} for autodetection or \code{"pt"} for potato and
\code{"sp"} for sweetpotato.}

\item{add}{Additional quantitative variables.}

\item{checknames}{Logical indicating if column names should be checked, default \code{TRUE}.}
}
\value{
It returns the data frame with all non-numeric variables transformed
to numeric, and a list of all NAs values introduced by coercion.
}
\description{
Check that fieldbook variables are stored as numeric.
}
\details{
It checks that all variables recognized by \code{check.names}
are stored as numeric. Non-numeric columns are transformed to numeric
and the NAs values introduced by coercion are listed.
}
\examples{
tmp <- check.names(potatoyield)
tmp <- check.numeric(tmp)
tmp <- check.names(pjpz09)
tmp <- check.numeric(tmp)
}
\author{
Raul Eyzaguirre.
}
