% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_source_checks.R
\name{check.rc}
\alias{check.rc}
\title{Check rows and columns}
\usage{
check.rc(row, col, rep = NULL, data)
}
\arguments{
\item{row}{Label for rows.}

\item{col}{Label for columns.}

\item{rep}{Label for replications.}

\item{data}{The name of the data frame.}
}
\value{
For each replication, the number (\code{nplot}) and list (\code{lplot})
of plots (unique row and column position) with more than one genotype.
}
\description{
This function checks that there is only one genotype in each row and column position.
}
\author{
Raul Eyzaguirre.
}
