% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_source_checks.R
\name{check.wd}
\alias{check.wd}
\title{Check data for a Wescott design}
\usage{
check.wd(trait, geno, ch1, ch2, row, col, ncb, data)
}
\arguments{
\item{trait}{The trait to analyze.}

\item{geno}{The genotypes.}

\item{ch1}{Name of check 1.}

\item{ch2}{Name of check 2.}

\item{row}{Label for rows.}

\item{col}{Label for columns.}

\item{ncb}{Number of columns between two check columns.}

\item{data}{The name of the data frame.}
}
\value{
Five control values (\code{c1}, \code{c2}, \code{c3}, \code{c4}, \code{c5})
for the grid of checks, the number of missing values for checks (\code{nmis.check})
and genotypes \code{nmis}, and the proportion of missing values for checks
(\code{pmis.check}) and genotypes (\code{pmis}).
}
\description{
This function checks the grid of checks on the Wescott design and
the number of missing values.
}
\details{
This function checks the grid of checks for the Wescoot design anc
calculates the number of missing values.
}
\author{
Raul Eyzaguirre.
}
