% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/CheckData.R
\name{checkdata02}
\alias{checkdata02}
\title{Check data for a MET in a RCBD}
\usage{
checkdata02(trait, geno, env, data)
}
\arguments{
\item{trait}{The trait to analyze}

\item{geno}{The genotypes}

\item{env}{The environments}

\item{data}{The name of the data frame}
}
\value{
c1, c2, c3, three control values
}
\description{
This function checks the frequencies of genotypes in each environment in a RCBD.
}
\details{
This function checks if there is more than one replication in a RCBD in
several environments, if there is any genotype without data for some specific environments,
and if the design is balanced.
}
\author{
Raul Eyzaguirre
}

