% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckNames.R
\name{checknames}
\alias{checknames}
\title{Check fieldbook traits names}
\usage{
checknames(fb)
}
\arguments{
\item{fb}{The name of the fieldbook data frame.}
}
\value{
It returns a data frame with all traits names in upper case, and a list of the traits
with names not included in the list shown above.
}
\description{
Check that fieldbook traits names correspond with the names defined in the document
"PROCEDURES  FOR THE EVALUATION AND ANALYSIS OF SWEETPOTATO TRIALS".
}
\details{
The data frame must use the following labels (lower or upper case):
\itemize{
 \item L       : Locations (LOC is also valid)
 \item Y       : Years
 \item S       : Seasons
 \item G       : Genotypes (GENO is also valid)
 \item NAME    : Names for genotypes
 \item E       : Environments (ENV is also valid)
 \item R       : Replications (REP is also valid)
 \item NOPS    : Number of plants sowed
 \item NOPE    : Number of plants established
 \item VIR1    : Virus symptoms (1-9), first evaluation, 4-8 weeks after planting (VIR is also valid)
 \item VIR2    : Virus symptoms (1-9), second evaluation, 1 month before harvest
 \item ALT1    : Alternaria symptoms (1-9), first evaluation, 4-8 weeks after planting (ALT is also valid)
 \item ALT2    : Alternaria symptoms (1-9), second evaluation, 1 month before harvest
 \item VV      : Vine vigor (1-9), 1 month before harvest
 \item VW      : Vine weight (kg/plot)
 \item NOPH    : Number of plants harvested
 \item NOPR    : Number of plants with roots
 \item NOCR    : Number of commercial roots
 \item NONC    : Number of non commercial roots
 \item CRW     : Commercial root weight (kg/plot)
 \item NCRW    : Non commercial root weight (kg/plot)
 \item RFCP    : Root primary flesh color using CIP color charts
 \item RFCS    : Root secondary flesh color using CIP color charts
 \item SCOL    : Storage root skin color (1-9)
 \item FCOL    : Storage root flesh color (1-9)
 \item RFCP    : Storage root primary flesh color (1-9)
 \item RFCS    : Storage root secondary flesh color (1-9)
 \item RS      : Root size (1-9)
 \item RF      : Root form (1-9)
 \item DAMR    : Root defects (1-9)
 \item RSPR    : Root sprouting (1-9)
 \item WED     : Weevil damage (1-9)
 \item DMF     : Fresh weight of roots for dry matter assessment (kg)
 \item DMD     : Dry weight of DMF samples (kg)
 \item DM      : Storage root dry matter content (\%)
 \item DMRY    : Dry matter root yield (kg)
 \item DMVF    : Fresh weight vines for dry matter assessment (kg)
 \item DMVD    : Dry weight of DMVF samples (kg)
 \item DMV     : Vines dry matter content (\%)
 \item DMFY    : Dry matter foliage yield (kg)
 \item FRAW1   : Root fiber (1-9), first determination (FRAW is also valid)
 \item SURAW1  : Root sugar (1-9), first determination (SURAW is also valid)
 \item STRAW1  : Root starch (1-9), first determination (STRAW is also valid)
 \item COOF1   : Cooked fiber (1-9), first evaluation (COOF is also valid)
 \item COOSU1  : Cooked sugars (1-9), first evaluation (COOSU is also valid)
 \item COOST1  : Cooked starch (1-9), first evaluation (COOST is also valid)
 \item COOT1   : Cooked taste (1-9), first evaluation (COOT is also valid)
 \item COOAP1  : Cooked appearance (1-9), first evaluation (COOAP is also valid)
 \item FRAW2   : Root fiber (1-9), second determination
 \item SURAW2  : Root sugar (1-9), second determination
 \item STRAW2  : Root starch (1-9), second determination
 \item COOF2   : Cooked fiber (1-9), second evaluation
 \item COOSU2  : Cooked sugars (1-9), second evaluation
 \item COOST2  : Cooked starch (1-9), second evaluation
 \item COOT2   : Cooked taste (1-9), second evaluation
 \item COOAP2  : Cooked appearance (1-9), second evaluation
 \item PROT    : Protein (\%)
 \item FE      : Iron (mg/100 g dry weight)
 \item ZN      : Zinc (mg/100 g dry weight)
 \item CA      : Calcium (mg/100 g dry weight)
 \item MG      : Magnesium (mg/100 g dry weight)
 \item BC      : Beta-carotene (mg/100 g dry weight)
 \item BC.CC   : Beta-carotene with color charts (mg/100 g fresh weight)
 \item TC      : Total carotenoids (mg/100 g dry weight)
 \item STAR    : Starch (\%)
 \item FRUC    : Fructose (\%)
 \item GLUC    : Glucose (\%)
 \item SUCR    : Sucrose (\%)
 \item MALT    : Maltose (\%)
 \item TRW     : Total root weight (kg/plot)
 \item CYTHA   : Commercial root yield (t/ha)
 \item CYTHA.AJ: Commercial root yield (t/ha) adjusted by number of harvested plants
 \item RYTHA   : Total root yield (t/ha)
 \item RYTHA.AJ: Total root yield (t/ha) adjusted by number of harvested plants
 \item ACRW    : Average commercial root weight (kg/root)
 \item NRPP    : Number of roots per plant
 \item YPP     : Yield per plant (kg/plant)
 \item CI      : Percent marketable roots (commercial index)
 \item HI      : Harvest index
 \item SHI     : Harvest sowing index (survival)
 \item BIOM    : Biomass yield (t/ha)
 \item BIOM.AJ : Biomass yield (t/ha) adjusted by number of harvested plants
 \item FYTHA   : Foliage total yield (t/ha)
 \item FYTHA.AJ: Foliage total yield (t/ha) adjusted by number of harvested plants
 \item RFR     : Root foliage ratio
 }
}
\examples{
checknames(pjpz09)
}
\author{
Raul Eyzaguirre.
}

