% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{ck.2f}
\alias{ck.2f}
\title{Check data for a 2-factor factorial}
\usage{
ck.2f(trait, A, B, rep, data)
}
\arguments{
\item{trait}{The trait to analyze.}

\item{A}{Factor A.}

\item{B}{Factor B.}

\item{rep}{The replications.}

\item{data}{The name of the data frame.}
}
\value{
Four control values (\code{c1}, \code{c2}, \code{c3}, and \code{c4}), the number
of missing values \code{nmis}, the proportion of missing values (\code{pmis}), the number
of levels of factor A (\code{na}), the number of levels of factor B (\code{nb}),
the number of replications (\code{nr}), a table with frequencies of valid cases
for each combination of the levels of both factors (\code{tfreq}), and a table with
frequencies of valid cases for each combination of the levels of both factors in each
replication (\code{tfreqr}).
}
\description{
This function checks the frequencies for a 2-factor factorial.
}
\details{
This function checks if there is more than one replication, if there is
any combination of the levels of both factors without data or with more data then
replications, and if the design is balanced.
}
\author{
Raul Eyzaguirre.
}
