% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{ck.abd}
\alias{ck.abd}
\title{Check data for an ABD}
\usage{
ck.abd(trait, geno, rep, dfr)
}
\arguments{
\item{trait}{The trait to analyze.}

\item{geno}{The genotypes including checks.}

\item{rep}{The replications.}

\item{dfr}{The name of the data frame.}
}
\value{
The number of checks \code{ng.check}, the number of no checks \code{ng},
the number of missing values for checks \code{nmis.check}), the number of
missing values for no checks \code{nmis}, the number \code{ncheck.0}
and list \code{check.0} of checks without data, the number \code{ncheck.1}
and list \code{check.1} of checks with only one datum, the number of checks
with at least two data \code{ncheck.2}, the number of replications \code{nr},
and the number of rows in the data frame with missing values for factors
(\code{nmis.fac}).
}
\description{
This function checks the frequencies of genotypes in an ABD.
}
\details{
This function checks the frequencies and number of missing values
in an ABD. For an ANOVA in an ABD it is needed at least two checks with at
least 2 valid cases each.
}
\author{
Raul Eyzaguirre.
}
