% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_f.R
\name{ck.f}
\alias{ck.f}
\title{Check data for a full factorial}
\usage{
ck.f(trait, factors, rep, dfr)
}
\arguments{
\item{trait}{The trait to analyze.}

\item{factors}{The factors.}

\item{rep}{The replications.}

\item{dfr}{The name of the data frame.}
}
\value{
control values (\code{c1}, and \code{c3}, the number
of missing values \code{nmis}, the proportion of missing values (\code{pmis}),
the number of factors (\code{nf}), the number of levels of each factor (\code{nl}),
the number of replications (\code{nr}), a table with frequencies of valid cases
for each combination of the levels of the factors (\code{tfreq}), a table with
frequencies of valid cases for each combination of the levels of the factors in
each replication (\code{tfreqr}), and the number of rows in the data frame with
missing values for factors (\code{nmis.fact}).
}
\description{
This function checks the frequencies for a full factorial.
}
\author{
Raul Eyzaguirre.
}
