% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{ck.fna}
\alias{ck.fna}
\title{Check factors with NA values}
\usage{
ck.fna(factors, dfr)
}
\arguments{
\item{factors}{The factors}

\item{dfr}{The name of the data frame}
}
\value{
The number of rows in the data frame with missing values for factors
(\code{nmis.fac}), and the data frame after removal of all these rows.
}
\description{
This function cheks the factors' structure for missing values.
}
\author{
Raul Eyzaguirre.
}
