% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_gr.R
\name{ck.fq}
\alias{ck.fq}
\title{Check frequencies}
\usage{
ck.fq(trait, factors, rep, dfr)
}
\arguments{
\item{trait}{The trait}

\item{factors}{The factors}

\item{rep}{The replications}

\item{dfr}{The name of the data frame}
}
\value{
A table of frequencies of valid cases for all factors' levels combinations
(\code{tf}) and a table of frequencies of valid cases for all factors' levels and
replications combinations (\code{tfr}).
}
\description{
This function cheks the frequencies of valid cases for treatments and replications.
}
\author{
Raul Eyzaguirre.
}
