% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_fs.R
\name{ck.fs}
\alias{ck.fs}
\title{Check factors structure}
\usage{
ck.fs(dfr, factors, rep = NULL)
}
\arguments{
\item{dfr}{The name of the data frame.}

\item{factors}{The names of the columns that identify the factors.}

\item{rep}{The name of the column that identifies the replications,
default is \code{NULL} for a CRD.}
}
\value{
The number of factors (\code{nf}), the number of levels of the factors
(\code{nl}), the lists of levels of factors (\code{lf}), the number of treatments
(\code{nt}), the list of treatments (\code{lt}), the number of replications
(\code{nrep}), the list of replications (\code{lrep}), the number of rows
with missing values for factors (\code{nmis.fac}), and the data frame after
removal of all these rows.
}
\description{
This function cheks the structure of factors.
}
\examples{
## Example 1
# Create a design
dfr <- cr.rcbd(1:20, 3, 10)
dfr <- dfr$book
# Check the design
ck.fs(dfr, "geno", "block")

## Example 2
# Create a design
A <- paste0("a", 1:5)
B <- paste0("b", 1:3)
dfr <- cr.f(c("A", "B"), list(A, B), "rcbd", 3, 10)
dfr <- dfr$book
# Check the design
ck.fs(dfr, c("A", "B"), "block")
}
\author{
Raul Eyzaguirre.
}
