% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{ck.lna}
\alias{ck.lna}
\title{Check factors' levels without data}
\usage{
ck.lna(trait, factors, dfr)
}
\arguments{
\item{trait}{The trait}

\item{factors}{The factors}

\item{dfr}{The name of the data frame}
}
\value{
The number of rows in the data frame with factor levels without data
(\code{nmis.lev}), and the data frame after removal of all these rows.
}
\description{
This function cheks the factors' structure for  factors' levels without data.
}
\author{
Raul Eyzaguirre.
}
