% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_lxt.R
\name{ck.lxt}
\alias{ck.lxt}
\title{Check data for a line by tester genetic design}
\usage{
ck.lxt(dfr, line, tester)
}
\arguments{
\item{dfr}{The name of the data frame.}

\item{line}{The name of the column that identifies the lines.}

\item{tester}{The name of the column that identifies the testers.}
}
\value{
Three control values, the number of lines (\code{nlin}),
and the number of testers (\code{ntes}). The control values are:
\itemize{
 \item \code{c1}: TRUE if all lines appear as parents.
 \item \code{c2}: TRUE if all testers appear as parents.
 \item \code{c3}: TRUE if all lines x testers are in the crosses.
}
}
\description{
This function checks the data structure for a line by tester genetic design.
}
\examples{
ck.lxt(lxt, "line", "tester")
}
\author{
Raul Eyzaguirre.
}
