% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_pos.R
\name{ck.pos}
\alias{ck.pos}
\title{Check row and column positions}
\usage{
ck.pos(dfr, row, col, rep = NULL)
}
\arguments{
\item{dfr}{The name of the data frame.}

\item{row}{The name of the column that identifies the rows.}

\item{col}{The name of the column that identifies the columns.}

\item{rep}{The name of the column that identifies the replications.}
}
\value{
For each replication, the number (\code{nplot}) and list (\code{lplot})
of plots (unique row and column position) with more than one genotype, and the 
number (\code{nrep}) and list (\code{lrep}) of replications.
}
\description{
This function checks that there is only one genotype in each row and column position
for each replication.
}
\examples{
# Create a design
dfr <- cr.rcbd(1:20, 3, 10)
dfr <- dfr$book
# Check positions
ck.pos(dfr, 'row', 'col', 'block')
}
\author{
Raul Eyzaguirre.
}
