% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_sp.R
\name{clean.data.sp}
\alias{clean.data.sp}
\title{Clean sweetpotato data}
\usage{
clean.data.sp(dfr, f = 10)
}
\arguments{
\item{dfr}{The name of the data frame.}

\item{f}{Factor for extreme values detection.}
}
\value{
It returns the data frame with all impossible values set to \code{NA},
some values set to \code{0} and a list of warnings with all the rows that have
been modified.
}
\description{
This is a wrapper for functions \code{setna.sp} and \code{setzero.sp} and
applies both in that order.
}
\details{
The data frame must use the labels (lower or upper case) listed
in function \code{check.names.sp}. Then functions \code{setna.sp} and
\code{setzero.sp} are applied to the data.
}
\examples{
dfr <- data.frame(crw = c(2.2, 5.0, 3.6, 12, 1600, -4, 0),
                  dm = c(21, 23, 105, 24, -3, 30, NA),
                  nocr = c(1.3, 10, 11, NA, 2, 5, NA),
                  scol = c(1, 0, 15, 5, 4, 7, NA),
                  fcol.cc = c(1, 15, 12, 24, 55, 20, NA))
clean.data.sp(dfr)
}
\author{
Raul Eyzaguirre.
}
