% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_co.R
\name{convert.co.pt}
\alias{convert.co.pt}
\title{Potato CO numbers}
\usage{
convert.co.pt(dfr)
}
\arguments{
\item{dfr}{The name of the data frame.}
}
\value{
It returns a data frame with all short labels for traits converted to
Crop Ontology variable numbers.
}
\description{
Converts potato short labels to Crop Ontology variable numbers.
}
\details{
All labels (lower or upper case) listed in function \code{check.names.pt}
are converted to CO variable numbers.
}
\examples{
convert.co.pt(potatoyield)
}
\author{
Raul Eyzaguirre.
}
