% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_f.R
\name{cr.f}
\alias{cr.f}
\title{Create design for a factorial experiment}
\usage{
cr.f(
  fnames,
  flevels,
  design = c("crd", "rcbd"),
  nrep,
  nc = NULL,
  fillby = c("rows", "columns"),
  serpentine = c("yes", "no"),
  alongside = c("no", "rows", "columns")
)
}
\arguments{
\item{fnames}{Factors' names.}

\item{flevels}{A list with the factors' levels.}

\item{design}{The design, \code{crd} or \code{rcbd}.}

\item{nrep}{Number of replications or blocks.}

\item{nc}{Number of columns in the field for a \code{crd}
or in the field for each block for a \code{rcbd}.}

\item{fillby}{Allocate the plots by \code{"rows"} or \code{"columns"},
default \code{"rows"}.}

\item{serpentine}{\code{"yes"} or \code{"no"}, default \code{"yes"}.}

\item{alongside}{\code{"no"} for independent blocks, or \code{"rows"}
or \code{"columns"} if blocks are together alongside rows or columns.}
}
\value{
It returns the fieldbook and fieldplan.
}
\description{
This function creates the fieldbook and fieldplan for a factorial experiment
following a CRD or a RCBD.
}
\examples{
A <- paste0("a", 1:5)
B <- paste0("b", 1:3)
C <- paste0("c", 1:2)
cr.f(c("A", "B", "C"), list(A, B, C), "rcbd", 3, 12)
}
\author{
Raul Eyzaguirre.
}
