% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_spld.R
\name{cr.spld}
\alias{cr.spld}
\title{Split-Plot Design}
\usage{
cr.spld(A, B, nrep, nc = NULL)
}
\arguments{
\item{A}{The levels of factor A (row factor to whole plots).}

\item{B}{The levels of factor B (column factor to subplots).}

\item{nrep}{Number of replications (or blocks).}

\item{nc}{Number of columns in each replication. Default is the
number of levels of factor B.}
}
\value{
It returns the fieldbook and fieldplan.
}
\description{
This function creates the fieldbook and fieldplan for a Split-Plot design.
}
\examples{
A <- paste0("a", 1:4)
B <- paste0("b", 1:3)
cr.spld(A, B, 3)
}
\author{
Raul Eyzaguirre.
}
