% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_ur.R
\name{cr.ur}
\alias{cr.ur}
\title{Unreplicated experiment}
\usage{
cr.ur(geno, nc)
}
\arguments{
\item{geno}{The list of genotypes.}

\item{nc}{Number of available columns on the field.}
}
\value{
It returns the fieldbook and fieldplan.
}
\description{
This function creates the fieldbook and fieldplan for an unreplicated
experiment with genotypes randomly allocated on a field.
}
\examples{
cr.ur(1:100, 5)
cr.ur(1:100, 7)
}
\author{
Raul Eyzaguirre.
}
