% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bb.R
\name{create.bb.sp}
\alias{create.bb.sp}
\title{Create sweetpotatobase design and data files}
\usage{
create.bb.sp(metadata)
}
\arguments{
\item{metadata}{The name of the metadata template file.}
}
\value{
It returns data.frames with names design.file and data.file ready to
upload into sweetpotatobase. Both data.frames should be saved as xlsx files.
}
\description{
Creates sweetpotatobase design and phenotypic data files for a set of fieldbooks.
}
\details{
The metadata template can be created with function \code{createmd.sp}.
The fieldbooks should be in memory or as csv files in the working directory,
with the same names specified in the column \code{trial_name} of the metadata,
with standard short labels (see \code{?check.names.sp} for the details).
}
\examples{
# Create designs
book1 <- cr.rcbd(1:20, 3, 10)$book
book2 <- cr.rcbd(1:20, 3, 10)$book
# Get fieldbook with minimal set of traits
PEP2023CLM_AT01 <- create.fb.sp(book1)
PEP2023CSR_AT02 <- create.fb.sp(book2)
# Create metadata file
metadata <- create.md.sp(trial_name = c('PEP2023CLM_AT01', 'PEP2023CSR_AT02'),
                         breeding_program = 'Peru-CIP',
                         location = c('CLM', 'CSR'),
                         year = 2023,
                         design_type = 'RCBD',
                         description = '20 genotypes with 3 complete blocks',
                         trial_type = 'Advanced Yield Trial')
# Create design and data files for sweetpotatobase
output <- create.bb.sp(metadata)
}
\author{
Raul Eyzaguirre.
}
