% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dd.R
\name{create.dd}
\alias{create.dd}
\title{Create potatobase and sweetpotatobase design and data files}
\usage{
create.dd(metadata, crop = c("pt", "sp"))
}
\arguments{
\item{metadata}{The name of the metadata template file.}

\item{crop}{\code{pt} for potato or \code{sp} for sweetpotato.}
}
\value{
It returns data.frames with names design.file and data.file ready to
upload into potatobase and sweetpotatobase. Both data.frames should be saved
as xlsx files.
}
\description{
Creates potatobase and sweetpotatobase design and phenotypic data files for
a set of fieldbooks.
}
\details{
The metadata template can be created with function \code{create.md}.
The fieldbooks should be in memory or as csv files in the working directory,
with the same names specified in the column \code{trial_name} of the metadata,
with standard short labels (type \code{pt.ont()} or \code{sp.ont()}
to see the list of variables).
}
\examples{
# Create designs
book1 <- cr.rcbd(1:20, 3, 10)$book
book2 <- cr.rcbd(1:20, 3, 10)$book
# Get fieldbook with minimal set of variables for sweetpotato
PEP2023CLM_AT01 <- create.fb(book1, crop = 'sp')
PEP2023CSR_AT02 <- create.fb(book2, crop = 'sp')
# Create metadata file
metadata <- create.md(trial_name = c('PEP2023CLM_AT01', 'PEP2023CSR_AT02'),
                      breeding_program = 'Peru-CIP',
                      location = c('CLM', 'CSR'),
                      year = 2023,
                      design_type = 'RCBD',
                      description = '20 genotypes with 3 complete blocks',
                      trial_type = 'Advanced Yield Trial')
# Create design and data files for sweetpotatobase
output <- create.dd(metadata, 'sp')
}
\author{
Raul Eyzaguirre.
}
