% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fb.R
\name{create.fb}
\alias{create.fb}
\title{Create potato and sweetpotato fieldbook}
\usage{
create.fb(
  design,
  label = c("standard", "CO"),
  minimal = TRUE,
  add = NULL,
  computation = FALSE,
  crop = c("pt", "sp")
)
}
\arguments{
\item{design}{The name of the design data frame.}

\item{label}{Use \code{standard} for standard sweetpotato short labels or
\code{CO} for CO numbers. Default is \code{standard}.}

\item{minimal}{Logical, if \code{TRUE}, a minimal list of traits is included.
Default is \code{TRUE}.}

\item{add}{Additional traits to include. Only if \code{minimal = TRUE}.}

\item{computation}{Logical, if \code{TRUE}, computed traits are included.
Only if \code{minimal = FALSE}. Default is \code{FALSE}.}

\item{crop}{\code{"pt"} for potato and \code{"sp"} for sweetpotato.}
}
\value{
It returns a data frame with fieldbook design and traits.
}
\description{
Creates a potato and sweetpotato fieldbook with short labels or CO numbers.
}
\details{
Only labels listed in function \code{check.names} are valid.
Uppercase labels are converted to lowercase.
}
\examples{
book <- cr.rcbd(1:20, 3, 10)$book
# Get fieldbook with minimal set of traits for potato
create.fb(book, crop = 'pt')
# Add additional traits
create.fb(book, add = c('fedw', 'zndw'), crop = 'pt')
# Get the fieldbook with CO numbers
create.fb(book, label = 'CO', add = c('dm', 'fedw', 'zndw'), crop = 'pt')
# Get fieldbook with minimal set of traits for sweetpotato
create.fb(book, crop = 'sp')
# Add additional traits
create.fb(book, add = c('bc', 'fe', 'zn'), crop = 'sp')
# Get the fieldbook with CO numbers
create.fb(book, label = 'CO', add = c('bc', 'fe', 'zn'), crop = 'sp')
}
\author{
Raul Eyzaguirre.
}
