% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fb.R
\name{create.fb.sp}
\alias{create.fb.sp}
\title{Create sweetpotato fieldbook}
\usage{
create.fb.sp(
  design,
  label = c("standard", "CO"),
  minimal = TRUE,
  add = NULL,
  computation = FALSE
)
}
\arguments{
\item{design}{The name of the design data frame.}

\item{label}{Use \code{standard} for standard sweetpotato short labels or
\code{CO} for CO numbers. Default is \code{standard}.}

\item{minimal}{Logical, if \code{TRUE}, a minimal list of traits is included.
Default is \code{TRUE}.}

\item{add}{Additional traits to include. Only if \code{minimal = TRUE}.}

\item{computation}{Logical, if \code{TRUE}, computed traits are included.
Only if \code{minimal = FALSE}. Default is \code{FALSE}.}
}
\value{
It returns a data frame with fieldbook design and traits.
}
\description{
Creates a sweetpotato fieldbook with short labels.
}
\details{
Only labels listed in function \code{check.names.sp} are valid.
Uppercase labels are converted to lowercase.
}
\examples{
book <- cr.rcbd(1:20, 3, 10)$book
# Get fieldbook with minimal set of traits
create.fb.sp(book)
# Add additional traits
create.fb.sp(book, add = c('bc', 'fe', 'zn'))
# Get the fieldbook with CO numbers
create.fb.sp(book, label = 'CO', add = c('bc', 'fe', 'zn'))
}
\author{
Raul Eyzaguirre.
}
