% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createfb.R
\name{createfb.pt}
\alias{createfb.pt}
\title{Create potato fieldbook}
\usage{
createfb.pt(
  design,
  label = c("standard", "CO"),
  minimal = TRUE,
  add = NULL,
  computation = FALSE
)
}
\arguments{
\item{design}{The name of the design data frame.}

\item{label}{Use \code{standard} for standard potato short labels or
\code{CO} for CO numbers. Default is \code{standard}.}

\item{minimal}{Logical, if \code{TRUE}, a minimal list of traits is included.
Default is \code{TRUE}.}

\item{add}{Additional traits to include. Only if \code{minimal = TRUE}.}

\item{computation}{Logical, if \code{TRUE}, computed traits are included.
Only if \code{minimal = FALSE}. Default is \code{FALSE}.}
}
\value{
It returns a data frame with fieldbook design and traits.
}
\description{
Creates a potato fieldbook with short labels.
}
\details{
Labels listed in function \code{check.names.pt} are included.
}
\examples{
book <- cr.rcbd(1:20, 3, 10)$book
createfb.pt(book)
}
\author{
Raul Eyzaguirre.
}
