% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docomp.R
\name{docomp}
\alias{docomp}
\title{Do computations over some factors}
\usage{
docomp(do, traits, factors, addcol = NULL, data)
}
\arguments{
\item{do}{The computation to perform. Implemented options are \code{count},
\code{max}, \code{mean}, \code{min}, and \code{sum}.}

\item{traits}{List of traits.}

\item{factors}{List of factors.}

\item{addcol}{Additional columns to keep.}

\item{data}{The name of the data frame containing the data.}
}
\value{
It returns a data frame with the computations.
}
\description{
Do computations for several traits for some specific factors.
}
\details{
This function do a specific computation for all the \code{traits}
for each level's combination of the \code{factors}. Additional columns can be
kept if specified in \code{addcol}.
}
\examples{
# The data
head(spg)
str(spg)

# Compute means for all the traits across the two replications
# for each genotype and location and then for each genotype 
# across the two locations.
traits <- c("rytha", "bc", "dm", "star", "nocr")
factors <- c("geno", "loc")
output1 <- docomp("mean", traits, factors, data = spg)
docomp("mean", traits, "geno", data = output1)

# Compute maxima for all the traits across the two replications
# for each genotype and location.
docomp("max", traits, factors, data = spg)
}
\author{
Raul Eyzaguirre
}

