% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docount.R
\name{docount}
\alias{docount}
\title{Count number of valid values}
\usage{
docount(traits, factors, addcol = NULL, data)
}
\arguments{
\item{traits}{List of traits to count for.}

\item{factors}{List of factors.}

\item{addcol}{Additional columns to keep.}

\item{data}{The name of the data frame containing the data.}
}
\value{
It returns a data frame with the counts.
}
\description{
Count the number of valid values for several traits for some specific factors.
}
\details{
This function counts the number of valid values for all the \code{traits}
for each level's combination of the \code{factors}. Additional columns can be
kept if specified in \code{addcol}.
}
\examples{
# The data
head(spg)
str(spg)

# Count the number of values for all the traits across the two replications
# and the two locations for each genotype.
traits <- c("rytha", "bc", "dm", "star", "nocr")
dosum(traits, "geno", data = spg)
}
\author{
Raul Eyzaguirre
}

