% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/domax.R
\name{domax}
\alias{domax}
\title{Compute maxima over some factors}
\usage{
domax(traits, factors, addcol = NULL, data)
}
\arguments{
\item{traits}{List of traits to compute maxima.}

\item{factors}{List of factors.}

\item{addcol}{Additional columns to keep.}

\item{data}{The name of the data frame containing the data.}
}
\value{
It returns a data frame with the maxima.
}
\description{
Compute maxima for several traits for some specific factors.
}
\details{
This function computes maxima for all the \code{traits}
for each level's combination of the \code{factors}. Additional columns can be
kept if specified in \code{addcol}.
}
\examples{
# The data
head(spg)
str(spg)

# Compute maxima for all the traits across the two replications
# for each genotype and location.
traits <- c('rytha', 'bc', 'dm', 'star', 'nocr')
factors <- c('geno', 'loc')
domax(traits, factors, data=spg)

# Save the output in a data.frame with name 'output1'
# and compute maxima for each genotype across the two locations.
output1 <- domax(traits, factors, data=spg)
domax(traits, 'geno', data=output1)
}
\author{
Raul Eyzaguirre
}

