% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domeans.R
\name{domeans}
\alias{domeans}
\title{Compute means over some factors}
\usage{
domeans(traits, factors, addcol = NULL, data)
}
\arguments{
\item{traits}{List of traits to compute means.}

\item{factors}{List of factors.}

\item{addcol}{Additional columns to keep.}

\item{data}{The name of the data frame containing the data.}
}
\value{
It returns a data frame with the means.
}
\description{
Compute means for several traits for some specific factors.
}
\details{
This function computes means for all the \code{traits}
for each level's combination of the \code{factors}. Additional columns can be
kept if specified in \code{addcol}.
}
\examples{
# The data
head(spg)
str(spg)

# Compute means for all the traits across the two replications
# for each genotype and location.
traits <- c("rytha", "bc", "dm", "star", "nocr")
factors <- c("geno", "loc")
domeans(traits, factors, data = spg)

# Save the output in a data.frame with name "output1"
# and compute means for each genotype across the two locations.
output1 <- domeans(traits, factors, data = spg)
domeans(traits, "geno", data = output1)
}
\author{
Raul Eyzaguirre
}

