% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dosum.R
\name{dosum}
\alias{dosum}
\title{Compute sum over some factors}
\usage{
dosum(traits, factors, addcol = NULL, data)
}
\arguments{
\item{traits}{List of traits to compute the sum.}

\item{factors}{List of factors.}

\item{addcol}{Additional columns to keep.}

\item{data}{The name of the data frame containing the data.}
}
\value{
It returns a data frame with the sum.
}
\description{
Compute the sum for several traits for some specific factors.
}
\details{
This function computes the sum for all the \code{traits}
for each level's combination of the \code{factors}. Additional columns can be
kept if specified in \code{addcol}.
}
\examples{
# The data
head(spg)
str(spg)

# Compute the sum for all the traits across the two replications
# for each genotype and location.
traits <- c('rytha', 'bc', 'dm', 'star', 'nocr')
factors <- c('geno', 'loc')
dosum(traits, factors, data=spg)

# Save the output in a data.frame with name 'output1'
# and compute the sum for each genotype across the two locations.
output1 <- dosum(traits, factors, data=spg)
dosum(traits, 'geno', data=output1)
}
\author{
Raul Eyzaguirre
}

