% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Elston.R
\name{elston}
\alias{elston}
\title{Elston Index}
\usage{
elston(traits, geno, env, rep, data, means = "single", model = "gxe",
  lb = 1)
}
\arguments{
\item{traits}{List of traits.}

\item{geno}{The genotypes.}

\item{env}{The environments.}

\item{rep}{The replications or blocks.}

\item{data}{The name of the data frame containing the data.}

\item{means}{Type of genotypic means to compute the index. Options are \code{"single"}
or \code{"fitted"}. The default is \code{"single"}.}

\item{model}{Type of model to compute the fitted means. Type \code{model = "gxe"} or
\code{model = "g+e"} for a model with or without interaction. The default is \code{"gxe"}.}

\item{lb}{Lower bound. \code{1} for k = min(x) and \code{2} for k = (n*min(x) - max(x))/(n-1)}
}
\value{
It returns a data frame with the genotypic means for each trait and the Elston index.
}
\description{
Function to compute the Elston index (Elston, R. C., 1963).
}
\details{
The Elston index is a weight free index.
If \code{means = "fitted"} and \code{model = "gxe"} then the arguments \code{env} and
\code{rep} must be specified.
If \code{means = "fitted"} and \code{model = "g+e"} then only the argument \code{env}
must be specified.
}
\examples{
# The data
head(spg)
str(spg)

# Run Elston index with all the traits
elston(c("rytha", "bc", "dm", "star", "nocr"), "geno", data = spg)
}
\author{
Raul Eyzaguirre
}
\references{
Elston, R. C. (1963). A weight-free index for the purpose of ranking or selection
with respect to several traits at a time. Biometrics. 19(1): 85-97.
}

