% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{factors}
\alias{factors}
\title{Factorization}
\usage{
factors(n)
}
\arguments{
\item{n}{Number for factorization.}
}
\value{
It returns the prime factors of a number.
}
\description{
Decompose a number in its prime factors.
}
\details{
It decomposes a number in its prime factors.
}
\examples{
factors(18)

}
\author{
Raul Eyzaguirre.
}
