% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friedman.R
\name{friedman.t}
\alias{friedman.t}
\title{Friedman test}
\usage{
friedman.t(trait, treat, block, dfr, alpha = 0.05, print.text = TRUE)
}
\arguments{
\item{trait}{The trait to analyze.}

\item{treat}{The treatments.}

\item{block}{The blocks.}

\item{dfr}{The name of the data frame.}

\item{alpha}{Significant level for comparisons.}

\item{print.text}{Logical, if \code{TRUE}, it prints output text to console.}
}
\value{
It returns the Friedman test and a the multiple comparisons.
}
\description{
Performs a Friedman rank sum test with multiple comparisons.
}
\details{
It tests the null hypothesis that the treatments have identical effects.
}
\examples{
# From Conover (1999)
dfr <- data.frame(y = c(4, 4, 3, 3, 4, 2, 1, 2, 3.5, 4, 4, 3.5,
                        3, 2, 1.5, 1, 2, 2, 3, 4, 1, 1, 2, 1,
                        2, 3.0, 1.5, 2, 1, 2, 2, 1, 2, 3, 3, 2,
                        1, 1, 4, 4, 3, 4, 4, 3, 3.5, 2, 1, 3.5),
                  grass = gl(4, 12),
                  homeowner = rep(1:12, 4))
friedman.t("y", "grass", "homeowner", dfr)
}
\references{
Conover, W. J.(1999). Practical Nonparametric Statistics..
}
\author{
Raul Eyzaguirre.
}
