% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_invalid_names.R
\name{get_invalid_names}
\alias{get_invalid_names}
\title{get column names not defined in crop ontology}
\usage{
get_invalid_names(dfr, add = NULL, crop = c("auto", "pt", "sp"))
}
\arguments{
\item{dfr}{The name of the data frame.}

\item{add}{Additional variables. See details.}

\item{crop}{\code{"auto"} for autodetection or \code{"pt"} for potato and \code{"sp"} for sweetpotato.}
}
\value{
a character vector of invalid column names
}
\description{
run \code{get_invalid_names()} after running \code{check.names()}
}
\details{
Check that fieldbook factors and variables' names correspond with the names defined
in crop ontology \url{https://cropontology.org} and in the potato and sweetpotato
CIP protocols. It also checks that all variables are stored as numeric.

Type \code{pt.ont()} or \code{sp.ont()} to see the list of variables and
corresponding short labels and CO numbers.
Additional variables are checked for extreme values only.
}
\examples{
\dontrun{
check.names(potatoyield) |> get_invalid_names()
check.names(pjpz09) |> get_invalid_names()
}
}
\author{
Raul Eyzaguirre.
}
