% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_tai.R
\name{gg_tai}
\alias{gg_tai}
\title{Tai's stability analysis}
\usage{
gg_tai(trait, geno, env, rep, data, maxp = 0.1, conf = 0.95, title = NULL,
  color = c("darkorange", "black", "gray"), size = c(1, 1),
  scaleSize = TRUE)
}
\arguments{
\item{trait}{The trait to analyze.}

\item{geno}{The genotypes.}

\item{env}{The environments.}

\item{rep}{The replications or blocks. A RCBD is assumed.}

\item{data}{The name of the data frame containing the data.}

\item{maxp}{Maximum allowed proportion of missing values to estimate, default is 10\%.}

\item{conf}{Probability for the Tai limits.}

\item{title}{Main title for plot.}

\item{color}{Color for symbols, labels and lines. Ignored; maintained for compatibility with original tai function.}

\item{size}{Relative size for symbols and labels. Ignored; maintained for compatibility with original tai function.}

\item{scaleSize}{logical. Whether dots should be scaled by trait value or not. Default is TRUE.}
}
\value{
It returns the Tai graph for stability analysis but not anymore the values of alpha
and lambda for each genotype. Use 'tai' function.
}
\description{
This function runs Tai's stability analysis (Tai, G. C. C., 1971).
It assumes a RCBD with fixed effects for genotypes and random effects for environments.
}
\details{
The limits for alpha and lambda are computed using the mean squares from
an ANOVA table for a RCBD with blocks nested into environments. If the data set is
unbalanced, a warning is produced.
}
\author{
Raul Eyzaguirre

Reinhard Simon (transfer to ggplot2)
}
\references{
Tai, G. C. C. (1971). Genotypic Stability Analysis and Its Application to Potato
Regional Trials, Crop Science, Vol 11.
}

