% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggammi.R
\name{ggammi}
\alias{ggammi}
\title{AMMI ggplot}
\usage{
ggammi(x, bp.type = 2, bp1.type = c("effects", "means"))
}
\arguments{
\item{x}{An object of class \code{ammi}.}

\item{bp.type}{Choose 1 for the trait-PC1 biplot and 2 for the PC1-PC2 biplot.}

\item{bp1.type}{Choose "effects" or "means" for biplot-1.}
}
\value{
It returns a dispersion plot of means or effects against the first PC,
or a dispersion plot of PC1 against PC2.
}
\description{
This function produces AMMI (Gollob, H. R., 1968) or GGE (Yan , W. et al., 2000) biplots
using the ggplot2 package.
}
\details{
It produces a ggplot biplot for an object of class \code{ammi}.
See \code{?ammi} for additional details.
}
\examples{
model.ammi <- ammi("y", "geno", "env", "rep", met8x12)
ggammi(model.ammi)
}
\references{
Gollob, H. R. (1968). A Statistical Model which combines Features of Factor Analytic
and Analysis of Variance Techniques, Psychometrika, Vol 33(1): 73-114.

Yan, W. et al. (2000). Cultivar evaluation and mega-environment investigation based
on the GGE biplot, Crop Sci., Vol 40: 597-605.
}
\author{
Raul Eyzaguirre.
}
