% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tai.R
\name{ggtai}
\alias{ggtai}
\title{Tai ggplot}
\usage{
ggtai(x, conf = 0.95, title = NULL)
}
\arguments{
\item{x}{An object of class \code{tai}.}

\item{conf}{Probability for the Tai limits.}

\item{title}{Main title for plot.}
}
\value{
It returns the Tai graph for stability analysis using the ggplot2 package.
}
\description{
This function produces a Tai's stability analysis plot (Tai, G. C. C., 1971)
using the ggplot2 package.
}
\details{
The limits for alpha and lambda are computed using the mean squares from
an ANOVA table for a RCBD with blocks nested into environments.See \code{?tai} for
additional details.
}
\examples{
model.tai <- tai("y", "geno", "env", "rep", met8x12)
ggtai(model.tai)
}
\references{
Tai, G. C. C. (1971). Genotypic Stability Analysis and Its Application to Potato
Regional Trials, Crop Science, Vol 11.
}
\author{
Raul Eyzaguirre.
}
