% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{lbb3c3}
\alias{lbb3c3}
\title{Data for a late blight experiment}
\format{
A data frame with 99 rows and 8 columns.
}
\source{
International Potato Center, potato experimental data.
}
\usage{
lbb3c3
}
\description{
This data set contains late blight evaluations at 6 dates
for a potato experiment with 33 genotypes and 3 replications.
}
