% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madna.R
\name{madna}
\alias{madna}
\title{Set impossible values to \code{NA}}
\usage{
madna(dfr, f = 10, add.con = NULL, add.cat = NULL, k = NULL)
}
\arguments{
\item{dfr}{The name of the data frame.}

\item{f}{Factor for extreme values detection. See details.}

\item{add.con}{Additional continuous positive traits.}

\item{add.cat}{Additional 1 to k categorical traits.}

\item{k}{The \code{k} values.}
}
\value{
It returns the data frame with all impossible values set to \code{NA}
and a list of warnings with all the rows that have been modified.
}
\description{
Detect impossible values for sweetpotato data and set them to missing value
(\code{NA}).
}
\details{
The data frame must use the labels (lower or upper case) listed in
function \code{check.names.sp}; see \code{?check.names.sp} for details.
Rules are:
\itemize{
 \item Continuous positive traits cannot be negative.
 \item 0 to 100 percentage values cannot be less than 0 or greater than 100.
 \item Discrete positive traits cannot be negative and should take integer values.
 \item 1 to 9 categorical data cannot take out of scale values.
 \item 1 to k categorical data cannot take out of scale values for k defined by user.
 \item Beta carotene values determined by RHS color charts cannot take values
 different from the possible values in the RHS color chart.
 \item Extreme low and high values are detected using the interquartile range.
 The rule is to detect any value out of the interval
 \eqn{[Q_1 - m - f \times IQR; Q_3 + m + f \times IQR]}
 where \code{m} is the mean. By default \code{f = 10} and cannot be less than 10.
}
}
\examples{
dfr <- data.frame(trw = c(2.2, 5.0, 3.6, 122, 1600),
                  dm = c(21, 23, 105, 24, -3),
                  tnr = c(1.3, 10, 11, NA, 2),
                  scol = c(1, 0, 15, 5, 4),
                  fcol.cc = c(1, 15, 12, 24, 55))
madna(dfr)
}
\author{
Raul Eyzaguirre.
}
