% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{megaclones}
\alias{megaclones}
\title{Some variables for a multi-environment trial (MET)}
\format{
A data frame with 312 rows and 6 columns.
}
\source{
International Potato Center, sweetpotato experimental data.
}
\usage{
megaclones
}
\description{
This data set has data for root yield in tons per hectare (\code{rytha}),
foliage yield in tons per hectare (\code{fytha}), and dry matter (\code{dm})
for an experiment with 13 sweetpotato genotypes (\code{geno}), 12 environments
(\code{env}), and 2 replications (\code{rep}) in each environment following a RCBD.
}
