% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{met8x12}
\alias{met8x12}
\title{Yields for a multi-environment trial (MET)}
\format{A data frame with 4 columns and 288 rows.}
\source{
International Potato Center, sweetpotato experimental data.
}
\usage{
met8x12
}
\description{
This data set has the yields per plot (\code{y}) for an experiment with
8 sweetpotato genotypes (\code{geno}), 12 environments (\code{env}), and 3 replications
(\code{rep}) in each evironment following a RCBD.
}
