% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/msdplot.r
\name{msdplot}
\alias{msdplot}
\title{Plot means and standard deviations with a dotplot}
\usage{
msdplot(trait, groups, data, conf = 0.95, nmax = 10, dotplot = "TRUE",
  sort.means = "none", main.title = NULL, x.title = "groups",
  y.title = "", col.means = "black", col.lines = "black",
  col.points = "black", jf = 0.1, dist = 0.1, ...)
}
\arguments{
\item{trait}{The trait to plot.}

\item{groups}{The grouping factor.}

\item{data}{The name of the data frame containing the data.}

\item{conf}{Probability for the confidence limits or number of standard deviations.}

\item{nmax}{Maximum number of points for the compulsory dotplot, default is 10.}

\item{dotplot}{Logical. If \code{TRUE}, a dotplot is shown. If \code{FALSE}
it will only suppress the dots if the number of data points is larger than \code{nmax}.}

\item{sort.means}{Sort for means. Options are \code{"none"}, \code{"increasing"},
and \code{"decreasing"}, \code{"none"} by default.}

\item{main.title}{Main title.}

\item{x.title}{Title for x axis.}

\item{y.title}{Title for y axis.}

\item{col.means}{Line color for mean symbols.}

\item{col.lines}{Line color for confidence interval lines.}

\item{col.points}{Color for data points.}

\item{jf}{Jitter factor for dots.}

\item{dist}{Horizontal distance between the means and the dots.}

\item{...}{Additional graphic parameters.}
}
\value{
It returns a plot with the means represented by horizontal lines, a vertical
line representing a confidence limit or a number of standard deviations,
and alternatively the individual data points.
}
\description{
Function to plot means and confidence limits.
}
\details{
An alternative to the controversial dynamite plots.
If \code{conf} is set to a value greater than or equal to 1, then it is interpreted
as number of standard deviations.
}
\examples{
# Simulate some data
mydata <- data.frame(y = rnorm(50, sample(40:60, 5), sample(5:10, 5)),
                     g = rep(1:5, 10))

# Draw the plot
msdplot("y", "g", mydata)
}
\author{
Raul Eyzaguirre
}

