% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mve_met.R
\name{mve.met}
\alias{mve.met}
\title{Estimation of missing values for a MET in a RCBD}
\usage{
mve.met(dfr, y, geno, env, rep, maxp = 0.1, tol = 1e-06)
}
\arguments{
\item{dfr}{The name of the data frame.}

\item{y}{The name of the column for the variable to estimate missing values.}

\item{geno}{The name of the column that identifies the genotypes.}

\item{env}{The name of the column that identifies the environments.}

\item{rep}{The name of the column that identifies the replications.}

\item{maxp}{Maximum allowed proportion of missing values to estimate, default is 10\%.}

\item{tol}{Tolerance for the convergence of the iterative estimation process.}
}
\value{
It returns a data frame with the experimental layout and columns \code{y}
and \code{y.est} with the original data and the original data plus the estimated values.
}
\description{
Function to estimate missing values for a Multi Environment Trial (MET) with a
Randomized Complete Block Design (RCBD) by the least squares method.
}
\examples{
mve.met(met8x12, "y", "geno", "env", "rep")
}
\author{
Raul Eyzaguirre.
}
