% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissValEst.R
\name{mveAB}
\alias{mveAB}
\title{Estimation of missing values for a 2-factor factorial}
\usage{
mveAB(trait, A, B, rep, design = c("crd", "rcbd"), data, maxp = 0.1,
  tol = 1e-06)
}
\arguments{
\item{trait}{The trait to estimate missing values.}

\item{A}{Factor A.}

\item{B}{Factor B.}

\item{rep}{The replications or blocks.}

\item{design}{The statistical design, \code{crd} or \code{rcbd}.}

\item{data}{The name of the data frame.}

\item{maxp}{Maximum allowed proportion of missing values to estimate, default is 10\%.}

\item{tol}{Tolerance for the convergence of the iterative estimation process.}
}
\value{
It returns a data frame with the experimental layout and columns \code{trait}
and \code{trait.est} with the original data and the original data plus the estimated values.
}
\description{
Function to estimate missing values for a 2-factor factorial with a CRD
or a RCBD by the least squares method.
}
\details{
A \code{data.frame} with data for a 2-factor factorial with at least two
replications and at least one datum for each factor levels' combination must be loaded.
Experimental data with only one replication, any factor levels' combination without data,
or more missing values than specified in \code{maxp} will generate an error message.
}
\examples{
# The data
str(asc)

# A copy with some missing values
temp <- asc
temp$dm[c(3, 10, 115)] <- NA

# Estimate the missing values
mveAB("dm", "geno", "treat", "rep", "crd", temp)
}
\author{
Raul Eyzaguirre.
}

