% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_w.R
\name{optim.w}
\alias{optim.w}
\title{Optimize Westcott adjustment}
\usage{
optim.w(trait, geno, replicated, ck1, ck2, row, col, ncb, nrs, method, ind,
  p, opt = c("F.value", "MSE", "CV"), dfr)
}
\arguments{
\item{trait}{The trait to adjust.}

\item{geno}{The genotypes.}

\item{replicated}{A \code{yes/no} column to identify replicated genotypes.}

\item{ck1}{Name of check 1.}

\item{ck2}{Name of check 2.}

\item{row}{Label for rows.}

\item{col}{Label for columns.}

\item{ncb}{Number of columns between two check columns.}

\item{nrs}{Set of values for the \code{nrs} param on \code{aj.w}.}

\item{method}{Set of values for the \code{method} param on \code{aj.w}.}

\item{ind}{Set of values for the \code{ind} param on \code{aj.w}.}

\item{p}{Set of values for the \code{p} param on \code{aj.w}.}

\item{opt}{Criteria for optimization, \code{F.value}, \code{MSE} or \code{CV}.
See details.}

\item{dfr}{The name of the data frame.}
}
\value{
It returns optimal set of values under the selected criteria.
}
\description{
This function optimizes the parameter values of the Westcott adjustment.
}
\details{
If \code{optim = F.value} then the F value for ANOVA is maximized.
If \code{optim = MSE} then the MSE is minimized.
If \code{optim = CV} then the CV is minimized.
}
\author{
Raul Eyzaguirre.
}
