% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesekbaker.R
\name{pesekbaker}
\alias{pesekbaker}
\title{Pesek-Baker Index}
\usage{
pesekbaker(
  geno,
  traits,
  dgg = NULL,
  units = c("sdu", "actual"),
  sf = 0.1,
  dfr,
  G.cov,
  P.cov
)
}
\arguments{
\item{geno}{The name of the column that identifies the genotypes.}

\item{traits}{The names of the columns for the traits.}

\item{dgg}{Desired genetic gains. The default is one standard deviation for each trait.}

\item{units}{Units for dgg, \code{"sdu"} or \code{"actual"}. See details for more information.}

\item{sf}{Selected fraction. The default is 0.1.}

\item{dfr}{The name of the data frame. See details for more information.}

\item{G.cov}{Genotypic covariance matrix.}

\item{P.cov}{Phenotypic covariance matrix.}
}
\value{
It returns:
\itemize{
\item \code{$Desired.Genetic.Gains}, the desired genetic gains in actual units,
\item \code{$Standard.Deviations}, the estimated genotypic standard deviations,
\item \code{$Index.Coefficients}, the index coefficients,
\item \code{$Response.to.Selection}, the response to selection,
\item \code{$Std.Response.to.Selection}, the standardized response to selection, and
\item \code{$Pesek.Baker.Index}, a data frame with the genotypic means for each trait,
the Pesek-Baker index, and the rank for each genotype according to the index.
}
}
\description{
Function to compute the Pesek-Baker index (Pesek, J. and R.J. Baker., 1969).
}
\details{
The Pesek-Baker is an index where relative economic weights have been replaced
by desired gains.
If \code{dgg} is not specified, the desired genetic gains are set to one
standard deviation for each trait. \code{dgg} can be specified in actual units
(\code{units = "actual"}) or in standard deviations (\code{units = "sdu"}),
defaults to \code{"sdu"}. For example, if you have a trait which is expressed
in kilograms and with a standard deviation of 5 kilograms, typing \code{dgg = 2}
means a desired genetic gain of 2 standard deviations that corresponds to 10 kilograms.
If you type \code{dgg = 2} and \code{units = "actual"} then this means a desired
genetic gain of 2 kilograms. If \code{dgg = NULL} then the desired genetic gains
will be one standard deviation, no matter if \code{units} is set as \code{"actual"}
or \code{"sdu"}.
The \code{dfr} must contain the estimations for each genotype with each trait.
}
\examples{
traits <- c("rytha", "bc", "dm", "star", "nocr")
output <- ecm(traits, "geno", "loc", "rep", spg)
colnames(output$blups) <- gsub('blup.', '', colnames(output$blups))
pesekbaker('geno', traits, dfr = output$blups, G.cov = output$G.cov, P.cov = output$P.cov)
# More weight on bc and dm, less on star and nocr
dgg <- c(1, 1.5, 1.5, 0.8, 0.8)
pesekbaker('geno', traits, dgg, dfr = output$blups, G.cov = output$G.cov, P.cov = output$P.cov)
}
\references{
Pesek, J. and R.J. Baker.(1969). Desired improvement in relation to selection indices.
Can. J. Plant. Sci. 9:803-804.
}
\author{
Raul Eyzaguirre.
}
