% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1Edata.R
\docType{data}
\name{pjpz09}
\alias{pjpz09}
\title{Data for a yield trial}
\format{A data frame with 10 columns and 204 rows.}
\source{
International Potato Center, sweetpotato experimental data.
}
\usage{
pjpz09
}
\description{
This data set contains data for several traits for a one
location experiment.
}

