% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ammi.R
\name{plot.ammi}
\alias{plot.ammi}
\title{AMMI or GGE biplots}
\usage{
\method{plot}{ammi}(x, biplot = 2, biplot1 = c("effects", "means"),
  main = NULL, xlab = NULL, color = c("darkorange", "black", "gray"),
  size = c(1, 1), ...)
}
\arguments{
\item{x}{An object of class \code{ammi}.}

\item{biplot}{Choose 1 for the trait-PC1 biplot and 2 for the PC1-PC2 biplot.}

\item{biplot1}{Choose "effects" or "means" for biplot1.}

\item{main}{Main title for biplot1 or biplot2.}

\item{xlab}{Xlab for biplot1.}

\item{color}{Color for lines, symbols and/or labels for environments, genotypes and axes.}

\item{size}{Relative size for symbols and labels.}

\item{...}{Additional plot arguments.}
}
\value{
It returns a dispersion plot of means or effects against the first PC,
or a dispersion plot of PC1 against PC2.
}
\description{
This function produces AMMI (Gollob, H. R., 1968) or GGE (Yan , W. et al., 2000) biplots.
}
\details{
It produces a biplot for an object of class \code{ammi}. See \code{?ammi}
for additional details.
}
\examples{
model.ammi <- ammi("y", "geno", "env", "rep", met8x12)
plot(model.ammi)
plot(model.ammi, biplot = 1)
}
\references{
Gollob, H. R. (1968). A Statistical Model which combines Features of Factor Analytic
and Analysis of Variance Techniques, Psychometrika, Vol 33(1): 73-114.

Yan, W. et al. (2000). Cultivar evaluation and mega-environment investigation based
on the GGE biplot, Crop Sci., Vol 40: 597-605.
}
\author{
Raul Eyzaguirre.
}
