% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ammi.R
\name{plot.st4gi_ammi}
\alias{plot.st4gi_ammi}
\title{AMMI or GGE biplots}
\usage{
\method{plot}{st4gi_ammi}(x, bp.type = 2, bp1.type = c("effects",
  "means"), graph.type = c("base", "ggplot"), color = c("darkorange",
  "black", "gray"), ...)
}
\arguments{
\item{x}{An object of class \code{ammi}.}

\item{bp.type}{Choose 1 for the trait-PC1 biplot and 2 for the PC1-PC2 biplot.}

\item{bp1.type}{Choose "effects" or "means" for biplot-1.}

\item{graph.type}{\code{"base"} or \code{"ggplot"}.}

\item{color}{Color for lines, symbols and/or labels for environments, genotypes and axes
(Only for the base system plot).}

\item{...}{Additional plot arguments (Only for the base system plot).}
}
\value{
It returns a dispersion plot of means or effects against the first PC,
or a dispersion plot of PC1 against PC2.
}
\description{
This function produces AMMI (Gollob, H. R., 1968) or GGE (Yan , W. et al., 2000) biplots
using the base system and ggplot2.
}
\details{
It produces a biplot for an object of class \code{ammi}. See \code{?ammi}
for additional details.
}
\examples{
model.ammi <- ammi("y", "geno", "env", "rep", met8x12)
plot(model.ammi)
plot(model.ammi, bp.type = 1)
plot(model.ammi, graph.type = "ggplot")
}
\references{
Gollob, H. R. (1968). A Statistical Model which combines Features of Factor Analytic
and Analysis of Variance Techniques, Psychometrika, Vol 33(1): 73-114.

Yan, W. et al. (2000). Cultivar evaluation and mega-environment investigation based
on the GGE biplot, Crop Sci., Vol 40: 597-605.
}
\author{
Raul Eyzaguirre.
}
