% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tai.R
\name{plot.st4gi_tai}
\alias{plot.st4gi_tai}
\title{Tai plot}
\usage{
\method{plot}{st4gi_tai}(x, conf = 0.95, graph.type = c("base",
  "ggplot"), color = c("darkorange", "black", "gray"), ...)
}
\arguments{
\item{x}{An object of class \code{tai}.}

\item{conf}{Probability for the Tai limits.}

\item{graph.type}{\code{"base"} or \code{"ggplot"}.}

\item{color}{Color for symbols, labels and lines (Only for the base system plot).}

\item{...}{Additional plot arguments (Only for the base system plot).}
}
\value{
It returns the Tai graph for stability analysis.
}
\description{
This function produces a Tai's stability analysis plot (Tai, G. C. C., 1971)
using the base system and ggplot2.
}
\details{
The limits for alpha and lambda are computed using the mean squares from
an ANOVA table for a RCBD with blocks nested into environments.See \code{?tai} for
additional details.
}
\examples{
model.tai <- tai("y", "geno", "env", "rep", met8x12)
plot(model.tai)
plot(model.tai, graph.type = "ggplot")
}
\references{
Tai, G. C. C. (1971). Genotypic Stability Analysis and Its Application to Potato
Regional Trials, Crop Science, Vol 11.
}
\author{
Raul Eyzaguirre.
}
