% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{potatostress}
\alias{potatostress}
\title{Data for a yield trial}
\format{
A data frame with 24 rows and 7 columns.
}
\source{
International Potato Center, potato experimental data.
}
\usage{
potatostress
}
\description{
This data set contains BLUPs for 24 genotypes and 3 traits in a 
potato trial with two treatments, well water and water stress.
}
