% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{potatoyield}
\alias{potatoyield}
\title{Data for a yield trial}
\format{A data frame with 46 columns and 972 rows.}
\source{
International Potato Center, potato experimental data.
}
\usage{
potatoyield
}
\description{
This data set contains data for 39 traits in a potato trial
with 258 genotypes, 3 irrigation treatments, and 6 blocks.
}
