% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_names.R
\name{pt.ont}
\alias{pt.ont}
\title{Potato ontology}
\usage{
pt.ont()
}
\value{
It returns a data frame.
}
\description{
Lists all variables in potato ontology used by \code{st4gi} package.
}
\examples{
pt.ont()
}
\author{
Raul Eyzaguirre.
}
