% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ptfs}
\alias{ptfs}
\title{Data for 5 potato trials}
\format{
A data frame with 382 rows and 22 columns.
}
\source{
International Potato Center, potato experimental data.
}
\usage{
ptfs
}
\description{
This data set contains data for 13 variables in 5 trials with
different number of genotypes, but some in common.
}
