% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_names.R
\name{ptont}
\alias{ptont}
\title{Potato ontology}
\usage{
ptont()
}
\value{
It returns a data frame.
}
\description{
Lists all variables in potato ontology used by \code{st4gi} package.
}
\examples{
ptont()
}
\author{
Raul Eyzaguirre.
}
