% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rcbd.R
\name{rcbd}
\alias{rcbd}
\title{ANOVA for a RCBD}
\usage{
rcbd(trait, treat, block, data, maxp = 0.1, anova = TRUE)
}
\arguments{
\item{trait}{The trait to analyze.}

\item{treat}{The treatments.}

\item{block}{The blocks.}

\item{data}{The name of the data frame containing the data.}

\item{maxp}{Maximum allowed proportion of missing values to estimate, default is 10\%.}

\item{anova}{Logical, if TRUE the ANOVA table is shown.}
}
\value{
If \code{anova} is {TRUE} it returns and shows the ANOVA table.
If \code{anova} is {FALSE} it returns the ANOVA table and some other components as the
estimated missing values, but nothing is shown.
}
\description{
Fit an analysis of variance model for a RCBD.
}
\details{
If data is unbalanced, missing values are estimated up to an specified maximum
proportion, 10\% by default.
}
\examples{
# The data
head(pjpz09)
str(pjpz09)

# Get a copy with some missing values for trw
temp <- pjpz09
temp[c(10, 20, 30), "trw"] <- NA

# Run ANOVA for trw
rcbd("trw", "geno", "rep", temp)
}
\author{
Raul Eyzaguirre
}

