% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{rm.fna}
\alias{rm.fna}
\title{Remove factors with NA values}
\usage{
rm.fna(factors, dfr)
}
\arguments{
\item{factors}{The factors}

\item{dfr}{The name of the data frame}
}
\value{
The number of rows in the data frame with missing values for factors
(\code{nmis.fac}), and the data frame after removal of all these rows.
}
\description{
This function removes rows for factors with missing values.
}
\author{
Raul Eyzaguirre.
}
