% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsa.R
\name{rsa}
\alias{rsa}
\title{Regression Stability Analysis}
\usage{
rsa(trait, geno, env, rep, dfr, maxp = 0.1)
}
\arguments{
\item{trait}{The trait to analyze.}

\item{geno}{The genotypes.}

\item{env}{The environments.}

\item{rep}{The replications.}

\item{dfr}{The name of the data frame.}

\item{maxp}{Maximum allowed proportion of missing values to estimate, default is 10\%.}
}
\value{
It returns the regression stability analysis decomposition of the GxE interaction
for genotypes and environments (Heterogeneity among regressions and deviation from regression),
the coefficient of variation, and the following regression stability measures for genotypes
and environments:
\itemize{
\item \code{a} the intercept. 
\item \code{b} the slope.
\item \code{se} the standard error for the slope.
\item \code{MSe} the mean square error.
\item \code{MSentry} the variance of the genotype means across environments and the environment
means across genotypes.
\item \code{MSinter} the variance of the genotype interaction effects across environments and the
environment interaction effects across genotypes.
}
}
\description{
Function to run the regression stability analysis (Yates and Cochran, 1938,
Finlay and Wilkinson, 1963).
}
\details{
The regression stability analysis is evaluated with a balanced data set.
If data is unbalanced, missing values are estimated up to an specified maximum proportion,
10\% by default. For the ANOVA table, genotypes and environments are considered as fixed
factors while the blocks are considered as random and nested into the environments.
To run a regression stability analysis you need a set of genotypes evaluated in a set of
environments. At least 3 genotypes or environments are needed. In a regression stability
analysis for genotypes grown at several environments, for each genotype a simple linear
regression of individual yield (Y) on the mean yield of all genotypes for each environment
(X) is fitted. In a similar way, for each environment a simple linear regression of
individual yield (Y) on the mean yield of all environments for each genotype (X) is fitted.
In both cases the X values are centered on zero, so the intercepts of the models correspond to
the means of the genotypes or environments.
}
\examples{
rsa("y", "geno", "env", "rep", met8x12)
}
\references{
Finlay, K. W., and Wilkinson, G. N. (1963). The Analysis of Adaption in a Plant-Breeding Programme.
Aust. J. Agric. Res. 14: 742-754.

Yates, F., and Cochran, W. G. (1938). The Analysis of Group Experiments.
J. Agric. Sci. 28: 556-580.
}
\author{
Raul Eyzaguirre.
}
