% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rts.R
\name{rts1}
\alias{rts1}
\title{Response to selection for a single experiment}
\usage{
rts1()
}
\value{
It returns a plot of response to seleccion versus number of replications
and computes the optimum number of replications and the response to selection
at this optimum value.
}
\description{
It finds the optimum number of replications to get the maximum response
to selection for a single experiment for a given plot capacity,
number of selected genotypes, genotypic variance and error variance.
}
\details{
It uses package \code{shiny} for the web layout.
Type \code{rts1()} in the R console to run the app.
}
\author{
Raul Eyzaguirre.
}
