% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rts.R
\name{rts2}
\alias{rts2}
\title{Response to selection with several locations}
\usage{
rts2()
}
\value{
It returns a plot of response to seleccion versus number of replications
and computes the optimum number of replications and the response to selection
at this optimum value.
}
\description{
It finds the optimum number of replications to get the maximum response
to selection with several locations for a given plot capacity,
number of locations, number of selected genotypes, genotypic variance,
genotypic by location variance, and error variance.
}
\details{
It uses package \code{shiny} for the web layout.
Type \code{rts2()} in the R console to run the app.
}
\author{
Raul Eyzaguirre.
}
