% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rts.R
\name{rts4}
\alias{rts4}
\title{Response to selection with several locations in two steps (two years)}
\usage{
rts4()
}
\value{
It returns the response to selection at step 1 and 2.
}
\description{
It computes the response to selection for each step in a two steps selection
with several locations for a given number of genotypes at step 1,
the number of locations, replications and selected genotypes at step 1,
the number of locations, replications and selected genotypes at step 2,
and the genotypic, genotypic by location, genotypic by year,
genotypic by location by year, and error variances.
}
\details{
It uses package \code{shiny} for the web layout.
Type \code{rts4()} in the R console to run the app.
}
\author{
Raul Eyzaguirre.
}
