% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setna_pt.R
\name{setna.pt}
\alias{setna.pt}
\title{Set values to \code{NA} for potato data.}
\usage{
setna.pt(dfr, f = 10)
}
\arguments{
\item{dfr}{The name of the data frame.}

\item{f}{Factor for extreme values detection. See details.}
}
\value{
It returns the data frame with all impossible values set to \code{NA}
and a list of warnings with all the rows that have been modified.
}
\description{
Detect impossible values for potato data and set them to missing value
(\code{NA}) according to some rules.
}
\details{
The data frame must use the labels (lower or upper case) listed in
function \code{check.names.pt}.
Consider the following groups of traits:
\itemize{
 \item \code{pre} (traits evaluated pre-harvest): \code{ppe}, \code{plant_unif},
 \code{plant_vigor} and \code{se}.
 
 \item \code{cnn} (continuos non-negative traits): \code{tntpl}, \code{nmtpl},
 \code{ttwp}, \code{ttwpl}, \code{mtwp}, \code{mtwpl}, \code{nomtwp}, \code{mtwci},
 \code{mtwcii}, \code{ttya}, \code{ttyna}, \code{mtya}, and \code{mtyna}.
 
 \item \code{cpo} (continuous positive traits): \code{atw}, \code{atmw},  
 \code{fwts1}, \code{fwts2}, \code{dwts1}, and \code{dwts2}.
     
 \item \code{pnn} (percentage non-negative traits): \code{ppe}, \code{pph},  
 \code{fruc}, \code{gluc}, \code{sucr}, and \code{malt}.
 
 \item \code{ppo} (percentage positive traits): \code{dm}, \code{pro},
 \code{star}, and \code{fiber}.

 \item \code{dnn} (discrete non-negative traits): \code{ntp}, \code{npe}, \code{nph},
 \code{tntp}, \code{nmtp}, \code{nnomtp}, \code{nmtci}, and \code{nmtcii}.
 
 \item \code{ctg} (categorical 1 to 9 traits): \code{plant_unif},
 \code{plant_vigor}, \code{se}, \code{tuber_apper}, \code{tub_unif},
 \code{tub_size}, \code{num_stolon}, and \code{leng_stolon}.
}
Values are set to \code{NA} with the following rules:
\itemize{
 \item \code{cnn} traits with negative values are set to \code{NA}.
 \item \code{cpo} traits with non-positive values are set to \code{NA}.
 \item \code{pnn} traits with values out of the [0, 100] interval are set to \code{NA}.  
 \item \code{ppo} with values out of the (0, 100] interval are set to \code{NA}.
 \item \code{dnn} traits with negative and non-integer values are set to \code{NA}.
 \item \code{ctg} traits with out of scale values are set to \code{NA}.
 \item Extreme low and high values are detected using the interquartile range.
 The rule is to detect any value out of the interval
 \eqn{[Q_1 - f \times (m/3 + IQR); Q_3 + f \times (m/3 + IQR)]} where \code{m}
 is the mean. By default \code{f = 10} and if less than 10 a warning is shown.
 Values out of this range are set to \code{NA}.
 \item If \code{npe == 0} and there is some data for any trait,
 then \code{npe} is set to \code{NA}.  
 \item If \code{nph == 0} and there is some data for any non-pre-harvest trait,
 then \code{nph} is set to \code{NA}.
 \item If \code{nmtp == 0} and \code{mtwp > 0}, then \code{nmtp} is set to \code{NA}.
 \item If \code{nmtp > 0} and \code{mtwp == 0}, then \code{mtwp} is set to \code{NA}.
 \item If \code{nnomtp == 0} and \code{nomtwp > 0}, then \code{nnomtp} is set to \code{NA}.
 \item If \code{nnomtp > 0} and \code{nomtwp == 0}, then \code{nomtwp} is set to \code{NA}.
}
}
\examples{
dfr <- data.frame(mtwp = c(2.2, 5.0, 3.6, 12, 1600, -4, 0),
                  dm = c(21, 23, 105, 24, -3, 30, NA),
                  nmtp = c(1.3, 10, 11, NA, 2, 5, NA))
setna.pt(dfr)
}
\author{
Raul Eyzaguirre.
}
