% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setna.R
\name{setna}
\alias{setna}
\title{Set values to \code{NA}.}
\usage{
setna(traits, add = NULL, dfr)
}
\arguments{
\item{traits}{List of traits (at least two).}

\item{add}{List of additional traits.}

\item{dfr}{The name of the data frame.}
}
\value{
It returns a data frame and a list of warnings with all the rows
that have been modified.
}
\description{
Set values to \code{NA} for selected traits.
}
\details{
If all traits in \code{traits} are zero or NA, then all traits
in \code{traits} and \code{add} are set to \code{NA}.
}
\examples{
dfr <- data.frame(noph = c(2, 0, 0, 2, 1),
                  trt1 = c(1, 0, 0, 0, NA),
                  trt2 = c(3, 0, 1, NA, 2),
                  trt3 = c(4, 0, 5, 0, 0),
                  trt4 = c(1, NA, 2, 4, 5))
setna(c('trt1', 'trt2', 'trt3'), 'trt4', dfr)
}
\author{
Raul Eyzaguirre
}
