% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setzero_pt.R
\name{setzero.pt}
\alias{setzero.pt}
\title{Set values to \code{0} for potato data.}
\usage{
setzero.pt(dfr)
}
\arguments{
\item{dfr}{The name of the data frame.}
}
\value{
It returns a data frame and a list of warnings with all the rows
that have been modified.
}
\description{
Set values to \code{0} for harvested traits.
}
\details{
This function sets values to \code{0} for all traits at harvest
(\code{nph}, \code{nmtp}, \code{nnomtp}, \code{mtwp}, and \code{nomtwp})
which are \code{NA} according to the following rules:
\itemize{
 \item If \code{npe > 0} and \code{nph == 0}, then all traits are set to \code{0}.
 \item If \code{nmtp == 0}, then \code{mtwp} is set to \code{0}.
 \item If \code{mtwp == 0}, then \code{nmtp} is set to \code{0}.
 \item If \code{nnomtp == 0}, then \code{nomtwp} is set to \code{0}.
 \item If \code{nomtwp == 0}, then \code{nnomtp} is set to \code{0}.
}
}
\examples{
dfr <- data.frame(nph = c(NA,  0,  3,  3,  3, 3, 3),
                  mtwp  = c(NA, NA, NA, NA,  8, 2, 0),
                  nomtwp = c(NA, NA, NA,  2, NA, 2, 4),
                  nmtp = c(NA, NA, NA,  0,  6, 2, 0),
                  nnomtp = c(NA, NA, NA,  4,  0, 3, 5))
setzero.pt(dfr)
}
\author{
Raul Eyzaguirre
}
