% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setzero.R
\name{setzero}
\alias{setzero}
\title{Set values to \code{0}.}
\usage{
setzero(dfr)
}
\arguments{
\item{dfr}{The name of the data frame.}
}
\value{
It returns a data frame and a list of warnings with all the rows
that have been modified.
}
\description{
Set values to \code{0} for harvested traits.
}
\details{
This function sets values to \code{0} for all traits at harvest
(\code{vw}, \code{nocr}, \code{nonc}, \code{crw}, and \code{ncrw}) which are
\code{NA} if at least one of these traits has a value greater than \code{0}.
}
\examples{
dfr <- data.frame(noph = c(2, 3, 3, 2, 1),
                  crw = c(1, 0, 0, 0, NA),
                  ncrw = c(3, 0, 1, NA, 2),
                  nocr = c(4, 0, 5, 0, 0),
                  nonc = c(1, NA, 2, 4, 5))
setzero(dfr)
}
\author{
Raul Eyzaguirre
}
