% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setzero_sp.R
\name{setzero.sp}
\alias{setzero.sp}
\title{Set values to \code{0} for sweetpotato data.}
\usage{
setzero.sp(dfr)
}
\arguments{
\item{dfr}{The name of the data frame.}
}
\value{
It returns a data frame and a list of warnings with all the rows
that have been modified.
}
\description{
Set values to \code{0} for harvested traits.
}
\details{
This function sets values to \code{0} for all traits at harvest
(\code{noph}, \code{nopr}, \code{vw}, \code{nocr}, \code{nonc}, \code{crw},
and \code{ncrw}) which are \code{NA} according to the following rules:
\itemize{
 \item If \code{nope > 0} and \code{noph == 0}, then all traits are set to \code{0}.
 \item If all traits are \code{0}, then \code{noph} is set to \code{0}.
 \item If \code{nopr == 0}, then all traits with exception of \code{vw}
 are set to \code{0}.
 \item If all traits with exception of \code{vw} are \code{0}, then
 \code{nopr} is set to \code{0}.
 \item If \code{nocr == 0}, then \code{crw} is set to \code{0}.
 \item If \code{crw == 0}, then \code{nocr} is set to \code{0}.
 \item If \code{nonc == 0}, then \code{ncrw} is set to \code{0}.
 \item If \code{ncrw == 0}, then \code{nonc} is set to \code{0}.
}
}
\author{
Raul Eyzaguirre
}
