% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_names.R
\name{sp.ont}
\alias{sp.ont}
\title{Sweetpotato ontology}
\usage{
sp.ont()
}
\value{
It returns a data frame.
}
\description{
Lists all variables in sweetpotato ontology used by \code{st4gi} package.
}
\examples{
sp.ont()
}
\author{
Raul Eyzaguirre.
}
