% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spconsis.R
\name{spconsis}
\alias{spconsis}
\title{Check consistency for sweetpotato experimental data}
\usage{
spconsis(fb, plot.size = NULL, f = 3, out.mod = c("none", "rcbd", "met"),
  out.max = 4, width = 240, file = FALSE)
}
\arguments{
\item{fb}{The name of the fieldbook data frame.}

\item{plot.size}{Plot size in square meters.}

\item{f}{Factor for extreme values detection. See details.}

\item{out.mod}{Statistical model for outliers' detection. See details.}

\item{out.max}{Threshold for outliers' detection.}

\item{width}{Number of columns for the output.}

\item{file}{Logigal, if TRUE the output goes to a file with name output.txt.}
}
\value{
If \code{file = TRUE} it returns a file with name checks.txt with a list of
all rows with some kind of inconsistency and all rows with outliers. If \code{file = FALSE}
the output is shown in the R console.
}
\description{
Set of rules to check for consistency of sweetpotato experimental data.
Data labels must be defined as specified in the PROCEDURES FOR THE EVALUATION
AND ANALYSIS OF SWEETPOTATO TRIALS document.
}
\details{
The data frame must use the labels (lower or upper case) listed in function \code{checknames}.
See \code{?checknames} for details.

\code{Plot.size} must be specified to check if traits expressed in tons per
hectare have been correctly computed from kilograms per plot. If \code{NULL} this
is not verified.

Extreme values are detected using the interquartile range.
The rule is to detect any value out of the interval 
\eqn{[Q_1 - f \times IQR; Q_3 + f \times IQR]}. By default \code{f = 3}.

Outliers are detected based on standardized residuals for some statistical models.
Options are \code{'rcbd'} and \code{'met'} for a randomized complete block design
and a multi environment trial with RCBD in each environment. By default the threshold
value is \code{out.max = 4}.
}
\examples{
spconsis(pjpz09)
}
\author{
Raul Eyzaguirre.
}

