% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{spg}
\alias{spg}
\title{Some traits for a multi-environment trial (MET)}
\format{A data frame with 8 columns and 32 rows.}
\source{
International Potato Center, sweetpotato experimental data.
}
\usage{
spg
}
\description{
This data set has data for root yield in tons per hectare (\code{rytha}),
beta-carotene (\code{bc}), dry matter (\code{dm}), starch (\code{star}) and number of
commercial roots (\code{nocr}) for an experiment with 8 sweetpotato genotypes
(\code{geno}), 2 locations (\code{loc}), and 2 replications (\code{rep}) in each
location following a RCBD.
}
