% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_names.R
\name{spont}
\alias{spont}
\title{Sweetpotato ontology}
\usage{
spont()
}
\value{
It returns a data frame.
}
\description{
Lists all variables in sweetpotato ontology used by \code{st4gi} package.
}
\examples{
spont()
}
\author{
Raul Eyzaguirre.
}
