% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssg1to3.R
\name{ssg1to3}
\alias{ssg1to3}
\title{Stage selection gain from 1 to 3 stages}
\usage{
ssg1to3(nstage, TC, nGF, Vgl, Vgs, Vgls, Ve, TG1, TG2 = NA, TG3 = NA,
  loc1, loc2 = NA, loc3 = NA, rep1, rep2 = NA, rep3 = NA)
}
\arguments{
\item{nstage}{Number of selection stages (1, 2, or 3).}

\item{TC}{Total test capacity as number of plots for the entire multiple environmental trial series.}

\item{nGF}{Number of finally selected clones.}

\item{Vgl}{Genotypic x location variance ratio with respect to genotypic variance.}

\item{Vgs}{Genotypic x season variance ratio with respect to genotypic variance.}

\item{Vgls}{Genotypic x location x season variance ratio with respect to genotypic variance.}

\item{Ve}{Error variance ratio with respect to genotypic variance.}

\item{TG1}{Number of genotypes tested in the first season.}

\item{TG2}{Number of genotypes tested in the second season.}

\item{TG3}{Number of genotypes tested in the third season.}

\item{loc1}{Number of locations in the first season.}

\item{loc2}{Number of locations in the second season.}

\item{loc3}{Number of locations in the third season.}

\item{rep1}{Number of replications in the first season.}

\item{rep2}{Number of replications in the second season.}

\item{rep3}{Number of replications in the third season.}
}
\value{
It returns the response to selection in standardized units.
}
\description{
It computes the response to selection after 1, 2 and 3-stage selection
}
\details{
It computes the response to selection in standardized units
after 1-stage selection with up to three seasons of testing,
2-stage selection with two seasons of testing, and 3-stage
selection with three seasons of testing. If no genotypes are tested at
season 2 or 3, the number of tested genotypes, the number of locations,
and the number of replications must be left as \code{NA}.
}
\examples{
# One stage, three seasons
ssg1to3(1, 2268, 3, 0.189, 0.103, 0.603, 1.162, 42, 42, 42, 9, 9, 9, 2, 2, 2)
# One stage, two seasons
ssg1to3(1, 2268, 3, 0.189, 0.103, 0.603, 1.162, 63, 63, NA, 9, 9, NA, 2, 2, NA)
# One stage, one season
ssg1to3(1, 2268, 3, 0.189, 0.103, 0.603, 1.162, 126, NA, NA, 9, NA, NA, 2, NA, NA)
# Two stages, two seasons
ssg1to3(2, 540, 3, 0.189, 0.103, 0.603, 1.162, 144, 9, NA, 3, 6, NA, 1, 2, NA)
# Three stages, three seasons
ssg1to3(3, 2268, 3, 0.189, 0.103, 0.603, 1.162, 499, 77, 8, 1, 7, 12, 2, 2, 2)
}
\author{
Bert De Boeck, Raul Eyzaguirre
}
