% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Compute sum of two traits}
\usage{
suma(a, b)
}
\arguments{
\item{a}{Name of trait 1 to sum}

\item{b}{Name of trait 2 to sum.}
}
\value{
It returns the sum of the two traits.
}
\description{
Compute the sum of two traits. Missing values do not propagate.
}
\details{
Missing values do not propagate. If NA is present for both traits then NA
is applied to the sum.
}
\examples{
## Compute total biomass as the sum of trw and vw
suma(pjpz09$trw, pjpz09$vw)
}
\author{
Raul Eyzaguirre.
}
