% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aj_westcott.R
\name{aj.w}
\alias{aj.w}
\title{Adjust values with a grid of checks}
\usage{
aj.w(trait, geno, ch1, ch2, row, col, nr = 5, ncb = 10, method = 2,
  p = 0.5, ind = TRUE, data)
}
\arguments{
\item{trait}{The trait to analyze.}

\item{geno}{The genotypes.}

\item{ch1}{Name of check 1.}

\item{ch2}{Name of check 2.}

\item{row}{Label for rows.}

\item{col}{Label for columns.}

\item{nr}{Number of rows used to fit values, 3 or 5.}

\item{ncb}{Number of columns between two check columns.}

\item{method}{The method to fit the values. See details.}

\item{p}{The proportion of the check values differences used for the adjustment.
See details.}

\item{ind}{Logical. If TRUE each check is centered around its own mean.
If FALSE both checks are centered around the overall mean.}

\item{data}{The name of the data frame.}
}
\value{
It returns the adjusted values.
}
\description{
This function adjust the observed values of an experiment planted following
the method described by Westcott (1981) with a grid of checks.
}
\details{
The values of the selected \code{trait} are adjusted following
different methods. For \code{method = 1} each plot is adjusted with the
mean of the six or ten checks located in the three (if \code{nr = 3}) or
five (if \code{nr = 5}) rows spanning it (with \code{nr = 3} it corresponds
tp the method proposed by Westcott). For \code{method = 2} each plot is
adjusted with the value of his position on the surface fitted using the six
or then checks located in the three or five rows spanning it, so the closest
checks receive more weight.

If \code{p = 1} then the values are adjusted in the same
proportion that the checks vary around the field. For values lower than 1
the values are adjusted based on that proportion over the checks variation.
If \code{p = 0} then there is no adjustment.
}
\references{
Westcott, B. (1981). Two methods for early generation yield assessment in winter wheat.
In: Proc. of the 4th meeting of the Biometrics in Plant Breeding Section of Eucarpia.
INRA Poitier, France, pp 91-95.
}
\author{
Raul Eyzaguirre.
}
