% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aj_w.R
\name{aj.w}
\alias{aj.w}
\title{Adjust values following the method of Westcott}
\usage{
aj.w(trait, geno, ck1, ck2, row, col, ncb = 10, nrs = NULL,
  method = c("weighted", "flat"), ind = TRUE, p = 0.5, dfr)
}
\arguments{
\item{trait}{The trait to adjust.}

\item{geno}{The genotypes.}

\item{ck1}{Name of check 1.}

\item{ck2}{Name of check 2.}

\item{row}{Label for rows.}

\item{col}{Label for columns.}

\item{ncb}{Number of columns between two check columns.}

\item{nrs}{Number of rows to span the row of the plot.}

\item{method}{The method to fit the values. See details.}

\item{ind}{Logical. See details.}

\item{p}{The proportion of the check values differences used for the adjustment.
See details.}

\item{dfr}{The name of the data frame.}
}
\value{
It returns the adjusted values.
}
\description{
This function adjust the observed values of an experiment planted following
the method described by Westcott (1981) with a grid of checks.
}
\details{
The values of the selected \code{trait} are adjusted using some mean
of the values of all the checks located on the row of the plot plus the \code{nrs}
rows at each side of the row of the plot. If \code{method = "flat"} the simple mean
of the checks is used for the adjustment, if \code{method = "weighted"} a weighted mean
is used for the adjustmen, where the checks closer to the plot get more weight.
If \code{p = 1} then the values are adjusted in the same proportion that the checks
vary around the field, for values lower than 1 the values are adjusted based on that
proportion over the checks variation, if \code{p = 0} then there is no adjustment.
If \code{ind = TRUE}, each check is centered around its own mean before the adjustment,
if \code{ind = FALSE}, then both checks are centered around the overall mean.
If \code{method = "flat"}, \code{ncb = 5}, \code{nrs = 1}, and \code{ind = FALSE}, then
it corresponds to the method proposed by Westcott.
If not specified, \code{nrs = floor(ncb / 2)}.
If the layout does not correspond with the Westcott method, then the observed values
are adjusted with the values of the checks planted nearby (this is in the rectangular
region definied by \code{floor(ncb / 2)} columns and rows at each side of
the plot) and a warning is issued.
}
\examples{
# Create design
dfr <- cr.w(1:1000, "Dag", "Cem", 40, 10)
dfr <- dfr$book
# Create some random data
dfr$y <- rnorm(dim(dfr)[1])
# Run adjustment
aj.w("y", "geno", "Dag", "Cem", "row", "col", dfr = dfr)
}
\references{
Westcott, B. (1981). Two methods for early generation yield assessment in winter wheat.
In: Proc. of the 4th meeting of the Biometrics in Plant Breeding Section of Eucarpia.
INRA Poitier, France, pp 91-95.
}
\author{
Raul Eyzaguirre.
}
