% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov_rcbd.R
\name{aov.rcbd}
\alias{aov.rcbd}
\title{ANOVA for a RCBD}
\usage{
aov.rcbd(trait, geno, rep, dfr, maxp = 0.1)
}
\arguments{
\item{trait}{The trait to analyze.}

\item{geno}{The genotypes.}

\item{rep}{The replications.}

\item{dfr}{The name of the data frame.}

\item{maxp}{Maximum allowed proportion of missing values to estimate, default is 10\%.}
}
\value{
It returns an ANOVA table.
}
\description{
Fit an analysis of variance model for a RCBD.
}
\details{
If data is unbalanced, missing values are estimated up to an specified maximum
proportion, 10\% by default.
}
\examples{
# Get a copy with some missing values for trw and run ANOVA
temp <- pjpz09
temp[c(10, 20, 30), "trw"] <- NA
aov.rcbd("trw", "geno", "rep", temp)
}
\author{
Raul Eyzaguirre.
}
