% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asl.R
\name{asl}
\alias{asl}
\title{Add non significant lines}
\usage{
asl(dfr, geno, aver, lsd, symbol = NULL)
}
\arguments{
\item{dfr}{The name of the data frame.}

\item{geno}{The name of the column that identifies the genotypes.}

\item{aver}{The name of the column with the averages.}

\item{lsd}{Least significant value.}

\item{symbol}{If \code{NULLL} it uses etters by default. See \code{details}}
}
\value{
It returns a data frame with averages and non significant lines.
}
\description{
This function adds columns to a datra frame with averages to identify
non significant pairwise comparisons.
}
\details{
The data frame is sorted descending and columns are added to
identify non singnificant pairwise comparisons. By default it works with
letters unless a different symbol is specified in \code{symbol}.
}
\examples{
# Adjust model
model <- aov.rcbd("trw", "geno", "rep", pjpz09)
# LSD values
lsd <- sqrt(model[3, 3]) * qt(.975, model[3, 1])
# Compute means
ests <- docomp('mean', 'trw', 'geno', dfr = pjpz09)
# Add non significant lines
ests <- asl(ests, 'geno', 'trw', lsd)
}
\author{
Raul Eyzaguirre.
}
