% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdt.R
\name{cdt}
\alias{cdt}
\title{Compute derived variables for potato and sweetpotato}
\usage{
cdt(
  dfr,
  method = c("none", "ps", "np"),
  value = NULL,
  crop = c("auto", "pt", "sp"),
  checknames = FALSE
)
}
\arguments{
\item{dfr}{The name of the data frame.}

\item{method}{Method to scale data from plot to hectare level. Options are
plot size \code{"ps"} and number of plants for a full hectare \code{"np"}.
See details.}

\item{value}{Value for the method selected in square meters if \code{method = "ps"}
and in number of plants per hectare if \code{method = "np"}. For data frames
with plots with different values, \code{value} must be the name of the column
in the data frame with the corresponding values for each plot.}

\item{crop}{\code{"auto"} for autodetection or \code{"pt"} for potato
and \code{"sp"} for sweetpotato.}

\item{checknames}{Logical indicating if column names should be checked, default \code{FALSE}.}
}
\value{
It returns a data frame with the original and derived variables.
}
\description{
Compute derived variables for a given fieldbook data frame.
}
\details{
The data frame must use the labels (lower or upper case) listed in
functions \code{ptont()} and \code{spont()}. 
Conversion from kilograms per plot to tons per hectare can be done using
\code{ps}, the plot size, or \code{np}, the total number of plants that can be
allocated in a full hectare. In both cases computations can be adjusted by the
number of harvested plants if available in the fieldbook. For \code{method = "np"},
number of plants sowed must be specified to compute non adjusted values.
}
\examples{
cdt(potatoyield)
cdt(pjpz09)
}
\author{
Raul Eyzaguirre.
}
