% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdt.R
\name{cdt}
\alias{cdt}
\title{Compute derived traits for sweetpotato}
\usage{
cdt(dfr, method = c("none", "ps", "np"), value = NULL, nops = NULL)
}
\arguments{
\item{dfr}{The name of the data frame.}

\item{method}{Method to scale data from plot to hectare level. Options are plot size
\code{"ps"} and number of plants for a full hectare \code{"np"}. See details.}

\item{value}{Value for the method selected in square meters if \code{method = "ps"}
and in number of plants per hectare if \code{method = "np"}.}

\item{nops}{Number of plants sowed per plot.}
}
\value{
It returns a data frame with the original and derived traits.
}
\description{
Compute derived traits for a given fieldbook data frame.
}
\details{
The data frame must use the labels (lower or upper case) listed in function
\code{check.names.sp}. See \code{?check.names.sp} for details.

Conversion from kilograms per plot to tons per hectare can be done using \code{ps}, the
plot size, or \code{np}, the total number of plants that are expected to be allocated in a
full hectare. In both cases computations can be adjusted by the number of harvested plants,
if available in the fieldbook. For \code{method = "np"}, \code{nops} must be specified to
compute non adjusted values.
}
\examples{
cdt(pjpz09)
}
\author{
Raul Eyzaguirre.
}
