% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_design.R
\name{check.abd}
\alias{check.abd}
\title{Check data for an ABD}
\usage{
check.abd(trait, treat, rep, data)
}
\arguments{
\item{trait}{The trait to analyze.}

\item{treat}{The treatments.}

\item{rep}{The replications.}

\item{data}{The name of the data frame.}
}
\value{
A list of treatments \code{newmat}, a list of checks \code{checks},
the number of treatments \code{nt.new}, the number of checks \code{nt.check},
the number of missing values for treatments \code{nmis.new}),
the number of missing values for checks \code{nmis.check},
the number \code{nt.check.0} and the list \code{check.0} of checks without data,
the number \code{nt.check.0} and the list \code{check.1} of checks with only one datum,
the number of checks with at least two data \code{nt.check.2},
and the number of replications \code{nr}.
}
\description{
This function checks the frequencies of genotypes in an ABD.
}
\details{
This function checks the frequencies and number of missing values in an ABD.
For an ANOVA in an ABD it is needed at least two checks with at least 2 valid cases each.
}
\author{
Raul Eyzaguirre.
}
