% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check.data}
\alias{check.data}
\title{Check consistency for sweetpotato experimental data}
\usage{
check.data(dfr, f = 3, out.mod = c("none", "rcbd", "met"), out.max = 4,
  aqt = NULL)
}
\arguments{
\item{dfr}{The name of the data frame.}

\item{f}{Factor for extreme values detection. See details.}

\item{out.mod}{Statistical model for outliers' detection. See details.}

\item{out.max}{Threshold for outliers' detection.}

\item{aqt}{Additional quantitative traits.}
}
\value{
It returns all rows with some kind of inconsistency or outliers.
}
\description{
Set of rules to check for consistency of sweetpotato experimental data.
Data labels must be defined as specified in the PROCEDURES FOR THE EVALUATION
AND ANALYSIS OF SWEETPOTATO TRIALS document.
}
\details{
The data frame must use the labels (lower or upper case) listed in function
\code{check.names}. See \code{?check.names} for details.

Extreme low and high values are detected using the interquartile range.
The rule is to detect any value out of the interval 
\eqn{[Q_1 - f \times IQR; Q_3 + f \times IQR]}. By default \code{f = 3}.

Outliers are detected based on standardized residuals for some statistical models.
Options are \code{'rcbd'} and \code{'met'} for a randomized complete block design
and a multi environment trial with RCBD in each environment. By default the threshold
value is \code{out.max = 4}.
}
\examples{
check.data(pjpz09)
}
\author{
Raul Eyzaguirre.
}
