% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_sp.R
\name{check.data.sp}
\alias{check.data.sp}
\title{Check consistency for sweetpotato experimental data}
\usage{
check.data.sp(
  dfr,
  f = 5,
  out.mod = c("none", "rcbd", "met"),
  out.max = 4,
  add = NULL
)
}
\arguments{
\item{dfr}{The name of the data frame.}

\item{f}{Factor for extreme values detection. See details.}

\item{out.mod}{Statistical model for outliers' detection. See details.}

\item{out.max}{Threshold for outliers' detection.}

\item{add}{Additional quantitative traits.}
}
\value{
It returns all rows with some kind of inconsistency or outliers.
}
\description{
Set of rules to check for consistency of sweetpotato experimental data.
Data labels must be defined as specified in the Procedures for the evaluation
and analysis of sweetpotato trials, ISBN 978-92-9060-522-5.
}
\details{
The data frame must use the labels (lower or upper case) listed in
function \code{check.names.sp}.

Extreme low and high values are detected using the interquartile range.
The rule is to detect any value out of the interval 
\eqn{[Q_1 - f \times IQR; Q_3 + f \times IQR]}. By default \code{f = 5}.

Outliers are detected based on standardized residuals for some statistical
models. Options are \code{"rcbd"} and \code{"met"} for a randomized complete
block design and a multi environment trial with RCBD in each environment.
By default the threshold value is \code{out.max = 4}.
}
\examples{
check.data.sp(pjpz09)
}
\author{
Raul Eyzaguirre.
}
