% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_freq.R
\name{check.freq}
\alias{check.freq}
\title{Check frequencies}
\usage{
check.freq(trait, geno, env = NULL, rep, data)
}
\arguments{
\item{trait}{The trait to analyze.}

\item{geno}{Genotypes.}

\item{env}{Environments.}

\item{rep}{The replications.}

\item{data}{The name of the data frame.}
}
\value{
Information about the balance, missing values, and replications of the design.
}
\description{
Check frequencies for designs with complete replications and one or two factors.
This is a wrapper for \code{check.rcbd} and \code{check.2f} functions.
}
\author{
Raul Eyzaguirre.
}
