% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_freq.R
\name{check.freq}
\alias{check.freq}
\title{Check frequencies for complete replications}
\usage{
check.freq(trait, geno, env = NULL, rep, dfr)
}
\arguments{
\item{trait}{The trait to analyze.}

\item{geno}{Genotypes.}

\item{env}{Environments, \code{NULL} if there are no environments.}

\item{rep}{The replications.}

\item{dfr}{The name of the data frame.}
}
\value{
Information about the balance, missing values, and replications of the design.
}
\description{
Check frequencies for designs with complete replications and one or several
environments. This is a wrapper for \code{ck.rcbd} and \code{ck.f} functions.
}
\examples{
check.freq("trw", "geno", NULL, "rep", pjpz09)
check.freq("rytha", "geno", "env", "rep", megaclones)
}
\author{
Raul Eyzaguirre.
}
