% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_names_pt.R
\name{check.names.pt}
\alias{check.names.pt}
\title{Check fieldbook traits names for potato}
\usage{
check.names.pt(dfr, add = NULL)
}
\arguments{
\item{dfr}{The name of the data frame.}

\item{add}{Additional traits.}
}
\value{
It returns a data frame with all traits names in lower case, and a list of the
traits with names not included in the list shown above.
}
\description{
Check that fieldbook traits names correspond with the names defined in 
\url{http://www.cropontology.org/ontology/CO_330/Potato}
}
\details{
The data frame must use the following labels (lower or upper case):

 -------------------- Plot identifiers --------------------
 \itemize{
 \item \code{plot}        : Plot number
 \item \code{row}         : Row number
 \item \code{col}         : Column number
 }
-------------------- Classification factors --------------------
 \itemize{
 \item \code{l}           : Locations (\code{loc} is also valid)
 \item \code{y}           : Years
 \item \code{s}           : Seasons
 \item \code{e}           : Environments (\code{env} is also valid)
 \item \code{g}           : Genotypes (\code{geno} is also valid)
 \item \code{name}        : Names for genotypes
 \item \code{r}           : Replications (\code{rep} is also valid)
 \item \code{b}           : Blocks (\code{block} is also valid)
 \item \code{treatment}   : Irrigation treatments
 \item \code{type}        : Type: early or late
 \item \code{instn}       : Code number
 }
-------------------- Traits N1 group --------------------
 \itemize{
 \item \code{ntp}         : Number of tubers planted
 \item \code{npe}         : Number of plants emerged
 \item \code{nph}         : Number of plants harvested
 \item \code{ppe}         : Proportion of plants emerged
 \item \code{pph}         : Proportion of plants harvested
 }
-------------------- Traits N2 group --------------------
 \itemize{
 \item \code{snpp}        : Stem number per plant
 \item \code{nipp}        : Number of inflorescences per plant
 \item \code{nfwp}        : Number of flowers per main inflorescence
 \item \code{nlpp}        : Number of leaves per plant
 \item \code{Num_Stolon}  : Number of stolons
 \item \code{Leng_Stolon} : Number of stolons
 }
-------------------- Traits N3 group --------------------
 \itemize{
 \item \code{tntp}        : Total number of tuber per plot
 \item \code{tntpl}       : Total number of tuber per plant
 \item \code{nmtp}        : Number of marketable tubers per plot
 \item \code{nmtpl}       : Number of marketable tubers per plant
 \item \code{nnomtp}      : Number of non-marketable tubers per plot
 \item \code{nmtci}       : Number of marketable tubers category I per plot
 \item \code{nmtcii}      : Number of marketable tubers category II per plot
 }
-------------------- Traits N4 group --------------------
 \itemize{
 \item \code{ttwp}        : Total tuber weight per plot
 \item \code{ttwpl}       : Total tuber weight per plant
 \item \code{mtwp}        : Marketable tuber weight per plot
 \item \code{mtwpl}       : Marketable tuber weight per plant
 \item \code{nomtwp}      : Non-marketable tuber weight per plot
 \item \code{mtwci}       : Marketable tuber weight category I per plot
 \item \code{mtwcii}      : Marketable tuber weight category II per plot
 }
-------------------- Traits N5 group--------------------
 \itemize{
 \item \code{ttya}        : Total tuber yield adjusted
 \item \code{ttyna}       : Total tuber yield no adjusted
 \item \code{mtya}        : Marketable tuber yield adjusted
 \item \code{mtyna}       : Marketable tuber yiedl no adjusted
 \item \code{atw}         : Average of tuber weight (\code{mwt} is also valid)
 \item \code{atmw}        : Average of marketable tuber weight (\code{mwmt} is also valid)
 }
-------------------- Traits N6 group--------------------
 \itemize{
 \item \code{stlfw}       : Stolon fresh weight per plant
 \item \code{sfw}         : Stem fresh weight per plant (\code{stfw} is also valid)
 \item \code{lfw}         : Leaf fresh weight per plant
 \item \code{rfw}         : Root fresh weight per plant
 \item \code{tfw}         : Tuber fresh weight per plant
 \item \code{tbfw}        : Total biomass fresh weight per plant
 \item \code{hi_fw}       : Harvest index fresh weight
 \item \code{fwts}        : Fresh weight of tuber sample
 }
-------------------- Traits N7 group--------------------
 \itemize{
 \item \code{stldw}       : Stolon dry weight per plant
 \item \code{sdw}         : Stem dry weight per plant (\code{stdw} is also valid)
 \item \code{ldw}         : Leaf dry weight per plant
 \item \code{rdw}         : Root dry weight per plant
 \item \code{tdw}         : Tuber dry weight per plant
 \item \code{tbdw}        : Total biomass dry weight per plant
 \item \code{hi_dw}       : Harvest index dry weight
 \item \code{dwts}        : Dry weight of tuber sample
 }
-------------------- Traits N8 group--------------------
 \itemize{
 \item \code{ldmcp}       : Leaf dry matter content per plot
 \item \code{sdmcp}       : Stem dry matter content per plot
 \item \code{rdmcp}       : Root dry matter content per plot
 \item \code{tdmcp}       : Tuber dry matter content per plot
 \item \code{pdm}         : Tuber dry matter content (\code{dm} is also valid)
 }
-------------------- Traits N9 group--------------------
 \itemize{
 \item \code{twa}         : Tuber weight in air
 \item \code{tww}         : Tuber weight in water
 \item \code{rsdw}        : Root system dry weight per plant
 \item \code{rd}          : Root density
 \item \code{rl}          : Root length
 \item \code{sg}          : Tuber specific gravity
 \item \code{dsi}         : Drought susceptibility index
 \item \code{dti}         : Drought tolerance index
 }
-------------------- Traits N10 group--------------------
 \itemize{
 \item \code{pro}         : Tuber protein content (\code{protein} is also valid)
 \item \code{star}        : Tuber starch content
 \item \code{fruc}        : Tuber fructose content
 \item \code{gluc}        : Tuber glucose content
 \item \code{sucr}        : Tuber sucrose content
 \item \code{malt}        : Tuber maltose content
 \item \code{fiber}       : Tuber fiber content
 }
-------------------- Traits N11 group--------------------
 \itemize{
 \item \code{plant_unif}  : Plant uniformity
 \item \code{plant_vigor} : Plant vigor
 \item \code{tuber_apper} : Tuber appearance
 \item \code{tub_unif}    : Tuber uniformity
 \item \code{tub_size}    : Tuber size
 \item \code{op}          : Osmotic potential
 \item \code{tlwp}        : Total leaf water potential
 \item \code{ltp}         : Leaf turgor potential
 \item \code{dap_chl}     : Chlorophyll relative content - Days after planting # removed
 \item \code{dap_tc}      : Canopy temperature - Days after planting           # removed
 }
-------------------- Traits N12 group--------------------
 \itemize{
 \item \code{leaflet_twi} : Leaflet turgid weight Evaluation i (i = 1, 2, 3, ...)
 \item \code{insnppi}     : Increase stem number per plant i (i = 1, 2, 3, ...)
 \item \code{insdi}       : Increase stem diameter i (i = 1, 2, 3, ...)
 \item \code{inrwci}      : Increase relative water content i (i = 1, 2, 3, ...)
 }
-------------------- Traits N13 group--------------------
 \itemize{
 \item \code{pw_evi}      : Plant wilting Evaluation i (i = 1, 2, 3, ...)
 \item \code{inplahei}    : Increase plant height i (i = 1, 2, 3, ...)
 \item \code{snpp_evi}    : Stem number per plant Evaluation i (i = 1, 2, 3, ...)
 \item \code{leaflet_fwi} : Leaflet fresh weight Evaluation i (i = 1, 2, 3, ...)
 \item \code{leaflet_dwi} : Leaflet dry weight Evaluation i (i = 1, 2, 3, ...)
 }
-------------------- Traits N14 group--------------------
 \itemize{
 \item \code{chci}        : Chlorophyll content sample i (i = 1, 2, 3, ...)
 \item \code{inchci}      : Increase chlorophyll content i (i = 1, 2, 3, ...)
 \item \code{leafsdi}     : Leaf stomata density sample i (i = 1, 2, 3, ...)
 \item \code{av_leafsd}   : Average leaf stomata density
 }
-------------------- Traits N15 group--------------------
 \itemize{
 \item \code{plahe_evi}   : Plant height Evaluation i (i = 1, 2, 3, ...)
 \item \code{sd_evi}      : Stem diameter Evaluation i (i = 1, 2, 3, ...)
 \item \code{cc_evi}      : Canopy cover Evaluation i (i = 1, 2, 3, ...)
 \item \code{chlspad_evi} : Chlorophyll content index Evaluation i (i = 1, 2, 3, ...)
 \item \code{cr_evi}      : Canopy reflectance Evaluation i (i = 1, 2, 3, ...)
 \item \code{lfa_evi}     : Leaflet area Evaluation i (i = 1, 2, 3, ...)
 \item \code{rwc_evi}     : Relative water content Evaluation i (i = 1, 2, 3, ...)
 \item \code{sla_evi}     : Specific leaf area Evaluation i (i = 1, 2, 3, ...)
 }
-------------------- Traits N16 group--------------------
 \itemize{
 \item \code{plahe_slp}   : Plant height slope
 \item \code{sd_slp}      : Stem diameter slope
 \item \code{cc_slp}      : Canopy cover slope
 \item \code{chlspad_slp} : Chlorophyll content index slope
 \item \code{cr_slp}      : Canopy reflectance slope
 \item \code{lfa_slp}     : Leaflet area slope
 \item \code{sla_slp}     : Specific leaf area slope         
 }
}
\examples{
check.names.pt(potatoyield)
}
\author{
Raul Eyzaguirre, Johan Ninanya.
}
