% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_names.R
\name{check.names}
\alias{check.names}
\title{Check fieldbook factors and variables names for potato and sweetpotato}
\usage{
check.names(dfr, crop = c("auto", "pt", "sp"))
}
\arguments{
\item{dfr}{The name of the data frame.}

\item{crop}{\code{"auto"} for autodetection or \code{"pt"} for potato and
\code{"sp"} for sweetpotato.}
}
\value{
It returns:
\itemize{
\item The fieldbook data frame with all column names in lowercase and
with some possible modifications in the names.
\item A list of warnings for all the column names that have been changed.
\item A list of warnings for all the column names not recognized.
}
}
\description{
Check that fieldbook factors and variables' names correspond with the names defined
in crop ontology \url{https://cropontology.org} and in the potato and sweetpotato
CIP protocols.
}
\details{
The following list of factors are recognized:
\itemize{
\item \code{plot}: The plot number.
\item \code{row}: The row number in the field.
\item \code{col}: The columnt number in the field.
\item \code{rep}: The replication number.
\item \code{block}: The block number.
\item \code{loc}: The location name.
\item \code{year}: The year.
\item \code{season}: The season.
\item \code{env}: The environment name.
\item \code{geno}: The genotype name.
\item \code{type}: The genotype type (clone, entry, parent, control, etc).
\item \code{treat}: The treatment name.
}
Type \code{ptont()} or \code{spont()} to see the list of variables and
corresponding short labels and CO numbers.
}
\examples{
check.names(potatoyield)
check.names(pjpz09)
}
\author{
Raul Eyzaguirre.
}
