% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check.pos}
\alias{check.pos}
\title{Check rows and columns}
\usage{
check.pos(row, col, rep, data)
}
\arguments{
\item{row}{Label for rows.}

\item{col}{Label for columns.}

\item{rep}{Label for replications.}

\item{data}{The name of the data frame.}
}
\value{
For each replication, a list of row and column positions with more than
one genotype.
}
\description{
This function checks that there is only one genotype in each row and column position.
}
\author{
Raul Eyzaguirre.
}
