% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pos.R
\name{check.pos}
\alias{check.pos}
\title{Check row and column positions}
\usage{
check.pos(row, col, rep, dfr)
}
\arguments{
\item{row}{Label for rows.}

\item{col}{Label for columns.}

\item{rep}{Label for replications.}

\item{dfr}{The name of the data frame.}
}
\value{
For each replication a list of plots (unique row and column position)
with more than one genotype.
}
\description{
This function checks that there is only one genotype in each row and column position.
This is a wrapper for \code{ck.pos} function.
}
\examples{
# Create a design
dfr <- cr.rcbd(1:20, 3, 10)
dfr <- dfr$book
# Check positions
check.pos('row', 'col', 'block', dfr)
}
\author{
Raul Eyzaguirre.
}
