% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/CheckData.R
\name{checkdata01}
\alias{checkdata01}
\title{Check data for a RCBD}
\usage{
checkdata01(trait, geno, data)
}
\arguments{
\item{trait}{The trait to analyze.}

\item{geno}{The genotypes.}

\item{data}{The name of the data frame.}
}
\value{
c1, c2, c3, three control values.
}
\description{
This function checks the frequencies of genotypes in a RCBD.
}
\details{
This function checks if there is more than one replication in a RCBD,
if there is any genotype without data, and if the design is balanced.
}
\author{
Raul Eyzaguirre.
}

