% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{ck.crd}
\alias{ck.crd}
\title{Check data for a CRD}
\usage{
ck.crd(trait, geno, dfr)
}
\arguments{
\item{trait}{The trait to analyze.}

\item{geno}{The genotypes.}

\item{dfr}{The name of the data frame.}
}
\value{
The number of genotypes (\code{ng}), the number of genotypes without
data (\code{ng.0}), the number of replications (\code{nr}), a table with
frequencies of valid cases for each genotype (\code{tfreq}), and the number
of rows in the data frame with missing values for factors (\code{nmis.fac}).
}
\description{
This function checks the frequencies of genotypes in a CRD.
}
\author{
Raul Eyzaguirre.
}
