% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{ck.crd}
\alias{ck.crd}
\title{Check data for a CRD}
\usage{
ck.crd(trait, geno, data)
}
\arguments{
\item{trait}{The trait to analyze.}

\item{geno}{The genotypes.}

\item{data}{The name of the data frame.}
}
\value{
Two control values (\code{c1} and \code{c2},
the number of genotypes (\code{ng}), the number of replications (\code{nr}),
a table with frequencies of valid cases for each genotype, and the number of
rows in the data frame with missing values for factors (\code{nmis.fact}).
}
\description{
This function checks the frequencies of genotypes in a CRD.
}
\details{
This function checks if there is more than one replication in a CRD and
if there is any genotype without data.
}
\author{
Raul Eyzaguirre.
}
