% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_crd.R
\name{ck.crd}
\alias{ck.crd}
\title{Check data for a CRD}
\usage{
ck.crd(trait, geno, dfr)
}
\arguments{
\item{trait}{The trait to analyze.}

\item{geno}{The genotypes.}

\item{dfr}{The name of the data frame.}
}
\value{
The number of genotypes (\code{ng}), the number of genotypes without
data (\code{ng.0}), the number of replications (\code{nrep}), and the number
of rows in the data frame with missing values for factors (\code{nmis.fac}).
}
\description{
This function checks the frequencies of genotypes in a CRD.
}
\examples{
# Create design
dfr <- cr.crd(1:50, 3, 10)
dfr <- dfr$book
# Create some random data
dfr$y <- rnorm(150)
# Delete some values
dfr[c(1, 5, 56, 77, 111), 'y'] <- NA
# Delete some values for classification factors
dfr[c(27, 48), 'geno'] <- NA
# Check the design
ck.crd('y', 'geno', dfr)
}
\author{
Raul Eyzaguirre.
}
