% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{ck.f}
\alias{ck.f}
\title{Check data for a full factorial}
\usage{
ck.f(trait, factors, rep, data)
}
\arguments{
\item{trait}{The trait to analyze.}

\item{factors}{The factors.}

\item{rep}{The replications.}

\item{data}{The name of the data frame.}
}
\value{
Four control values (\code{c1}, \code{c2}, \code{c3}, and \code{c4}),
the number of missing values \code{nmis}, the proportion of missing values
(\code{pmis}), the number of factors (\code{nf}), the number of levels of each
factor (\code{nl}), the number of replications (\code{nr}), a table with
frequencies of valid cases for each combination of the levels of the factors
(\code{tfreq}), a table with frequencies of valid cases for each combination
of the levels of the factors in each replication (\code{tfreqr}), and the
number of rows in the data frame with missing values for factors (\code{nmis.fact}).
}
\description{
This function checks the frequencies for a full factorial.
}
\details{
This function checks if there is more than one replication, if there is
any combination of the levels of the factors without data or with more data than
replications, and if the design is balanced.
}
\author{
Raul Eyzaguirre.
}
