% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_f.R
\name{ck.f}
\alias{ck.f}
\title{Check data for a full factorial}
\usage{
ck.f(trait, factors, rep, dfr)
}
\arguments{
\item{trait}{The trait to analyze.}

\item{factors}{The factors.}

\item{rep}{The replications, \code{NULL} for a CRD.}

\item{dfr}{The name of the data frame.}
}
\value{
The number of treatments without data (\code{nt.0}), the number of
treatments that appear more than once in a given replication (\code{nt.mult}),
the number of missing values \code{nmis}, the proportion of missing values
(\code{pmis}), the number of factors (\code{nf}), the number of levels of each
factor (\code{nl}), the number of replications (\code{nrep}), a table with frequencies
of valid cases for each combination of the levels of the factors (\code{tf}), a table
with frequencies of valid cases for each combination of the levels of the factors in
each replication (\code{tfr}), and the number of rows in the data frame with
missing values for factors (\code{nmis.fac}).
}
\description{
This function checks the frequencies for a full factorial.
}
\examples{
# Create a design
A <- paste0("a", 1:5)
B <- paste0("b", 1:3)
dfr <- cr.f(c("A", "B"), list(A, B), "rcbd", 3, 10)
dfr <- dfr$book
# Create some random data
dfr$y <- rnorm(45)
# Delete some values
dfr[c(4, 5, 12), 'y'] <- NA
# Check the design
ck.f("y", c("A", "B"), "block", dfr)
}
\author{
Raul Eyzaguirre.
}
