% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_fq.R
\name{ck.fq}
\alias{ck.fq}
\title{Check frequencies}
\usage{
ck.fq(trait, factors, rep, dfr)
}
\arguments{
\item{trait}{The trait.}

\item{factors}{The factors.}

\item{rep}{The replications, \code{NULL} for a CRD.}

\item{dfr}{The name of the data frame.}
}
\value{
A table of frequencies of valid cases for all factors' levels combinations
(\code{tf}), a table of frequencies of valid cases for all factors' levels and
replications combinations (\code{tfr}), the number of missing values \code{nmis},
and the proportion of missing values (\code{pmis}).
}
\description{
This function cheks the frequencies of valid cases for treatments and replications.
}
\examples{
## Example 1
# Create a design
dfr <- cr.rcbd(1:20, 3, 10)
dfr <- dfr$book
# Create some random data
dfr$y <- rnorm(60)
# Delete some values
dfr[c(1, 5, 16, 17), 'y'] <- NA
# Check the frequencies
ck.fq("y", "geno", "block", dfr)

## Example 2
# Create a design
A <- paste0("a", 1:5)
B <- paste0("b", 1:3)
dfr <- cr.f(c("A", "B"), list(A, B), "rcbd", 3, 10)
dfr <- dfr$book
# Create some random data
dfr$y <- rnorm(45)
# Delete some values
dfr[c(5, 10, 24), 'y'] <- NA
# Check the frequencies
ck.fq("y", c("A", "B"), "block", dfr)
}
\author{
Raul Eyzaguirre.
}
