% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_fs.R
\name{ck.fs}
\alias{ck.fs}
\title{Check number of genotypes and replications}
\usage{
ck.fs(geno, rep = NULL, design = c("crd", "rcbd", "abd"), dfr)
}
\arguments{
\item{geno}{The genotypes}

\item{design}{The statistical design}

\item{dfr}{The name of the data frame}
}
\value{
The number of genotypes (\code{ng}), the number of checks (\code{ng.check}),
the list of genotypes (\code{lg}), the list of checks (\code{lg.check}),
and the number of replications (\code{nr}).
}
\description{
This function cheks the number of genotypes and replications for different designs.
}
\examples{
# Create a design
dfr <- cr.rcbd(1:20, 3, 10)
dfr <- dfr$book
# Check the design
ck.fs('geno', 'block', 'rcbd', dfr = dfr)
}
\author{
Raul Eyzaguirre.
}
