% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_fs.R
\name{ck.fs}
\alias{ck.fs}
\title{Check factors structure}
\usage{
ck.fs(factors, rep = NULL, dfr)
}
\arguments{
\item{factors}{The factors}

\item{rep}{The replications}

\item{dfr}{The name of the data frame}
}
\value{
The number of factors (\code{nf}), the number of levels of the factors
(\code{nl}), the number of treatments (\code{nt}), the list of treatments (\code{lt}),
the number of replications (\code{nrep}), and the list of replications (\code{lrep}).
}
\description{
This function cheks the number of treatments and replications for different designs.
}
\examples{
## Example 1
# Create a design
dfr <- cr.rcbd(1:20, 3, 10)
dfr <- dfr$book
# Check the design
ck.fs("geno", "block", dfr)

## Example 2
# Create a design
A <- paste0("a", 1:5)
B <- paste0("b", 1:3)
dfr <- cr.f(c("A", "B"), list(A, B), "rcbd", 3, 10)
dfr <- dfr$book
# Check the design
ck.fs(c("A", "B"), "block", dfr)
}
\author{
Raul Eyzaguirre.
}
