% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{ck.gr}
\alias{ck.gr}
\title{Check number of genotypes and replications}
\usage{
ck.gr(geno, design = c("crd", "rcbd", "abd"), dfr)
}
\arguments{
\item{geno}{The genotypes}

\item{design}{The statistical design}

\item{dfr}{The name of the data frame}
}
\value{
The number of genotypes (\code{ng}), the number of checks (\code{ng.check}),
and the number of replications (\code{nr}).
}
\description{
This function cheks the number of genotypes and replications for different designs.
}
\author{
Raul Eyzaguirre.
}
