% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_lxt.R
\name{ck.lxt}
\alias{ck.lxt}
\title{Check data for a line by tester genetic design}
\usage{
ck.lxt(line, tester, data)
}
\arguments{
\item{line}{The lines.}

\item{tester}{The testers.}

\item{data}{The name of the data frame.}
}
\value{
Three control values (\code{c1}, \code{c2}, and \code{c3},
the number of lines (\code{nl}), and the number of testers (\code{nt}).
}
\description{
This function checks the data structure for a line by tester genetic design.
}
\details{
The control values are:
\itemize{
 \item \code{c1}: TRUE if all lines appear as parents.
 \item \code{c2}: TRUE if all testers appear as parents.
 \item \code{c3}: TRUE if all lines x testers are in the crosses.
}
}
\author{
Raul Eyzaguirre.
}
