% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_pos.R
\name{ck.pos}
\alias{ck.pos}
\title{Check row and column positions}
\usage{
ck.pos(row, col, rep = NULL, dfr)
}
\arguments{
\item{row}{Label for rows.}

\item{col}{Label for columns.}

\item{rep}{Label for replications.}

\item{dfr}{The name of the data frame.}
}
\value{
For each replication, the number (\code{nplot}) and list (\code{lplot})
of plots (unique row and column position) with more than one genotype, and the 
number (\code{nrep}) and list (\code{lrep}) of replications.
}
\description{
This function checks that there is only one genotype in each row and column position
for each replication.
}
\examples{
# Create a design
dfr <- cr.rcbd(1:20, 3, 10)
dfr <- dfr$book
# Check positions
ck.pos('row', 'col', 'block', dfr)
}
\author{
Raul Eyzaguirre.
}
