% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{ck.rcbd}
\alias{ck.rcbd}
\title{Check data for a RCBD}
\usage{
ck.rcbd(trait, geno, rep, data)
}
\arguments{
\item{trait}{The trait to analyze.}

\item{geno}{The genotypes.}

\item{rep}{The replications.}

\item{data}{The name of the data frame.}
}
\value{
Four control values (\code{c1}, \code{c2}, \code{c3}, and \code{c4}),
the number of missing values \code{nmis}, the proportion of missing values
(\code{pmis}), the number of genotypes (\code{ng}), the number of replications
(\code{nr}), a table with frequencies of valid cases for each genotype, and
the number of rows in the data frame with missing values for factors
(\code{nmis.fact}).
}
\description{
This function checks the frequencies of genotypes in a RCBD.
}
\details{
This function checks if there is more than one replication in a RCBD,
if there is any genotype without data or with more data than replications, and
if the design is balanced.
}
\author{
Raul Eyzaguirre.
}
