% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_rcbd.R
\name{ck.rcbd}
\alias{ck.rcbd}
\title{Check data for a RCBD}
\usage{
ck.rcbd(trait, geno, rep, dfr)
}
\arguments{
\item{trait}{The trait to analyze.}

\item{geno}{The genotypes.}

\item{rep}{The replications.}

\item{dfr}{The name of the data frame.}
}
\value{
The number of genotypes without data (\code{ng.0}), the number of
genotypes with more than one plot in a given block (\code{ng.2}), the number
of missing values \code{nmis}, the proportion of missing values (\code{pmis}),
the number of genotypes (\code{ng}), the number of replications (\code{nrep}),
a table with valid cases for each genotype (\code{tf}), a table with valid 
cases for each genotype in each replication (\code{tfr}), and the number of
rows in the data frame with missing values for factors (\code{nmis.fac}).
}
\description{
This function checks the frequencies of genotypes in a RCBD.
}
\examples{
# Create design
dfr <- cr.rcbd(1:20, 3, 10)
dfr <- dfr$book
# Create some random data
dfr$y <- rnorm(60)
# Delete some values
dfr[c(1, 5, 16, 17), 'y'] <- NA
# Check the design
ck.rcbd('y', 'geno', 'block', dfr)
}
\author{
Raul Eyzaguirre.
}
