% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck_w.R
\name{ck.w}
\alias{ck.w}
\title{Check data for a Wescott layout}
\usage{
ck.w(dfr, y, geno, ck1, ck2, row, col, ncb)
}
\arguments{
\item{dfr}{The name of the data frame.}

\item{y}{The name of the column for the variable to analyze.}

\item{geno}{The name of the column that identifies the genotypes.}

\item{ck1}{Name of check 1.}

\item{ck2}{Name of check 2.}

\item{row}{The name of the column that identifies the rows.}

\item{col}{The name of the column that identifies the columns.}

\item{ncb}{Number of columns between two check columns.}
}
\value{
Four control values (\code{c1}, \code{c2}, \code{c3}, and \code{c4},
for the grid of checks, the number of missing values for checks (\code{nmis.ck})
and genotypes \code{nmis}, the proportion of missing values for checks
(\code{pmis.ck}) and genotypes (\code{pmis}), and the number of rows
in the data frame with missing values for factors (\code{nmis.fac}).
}
\description{
This function checks the grid of checks on the Wescott layout and
the number of missing values.
}
\examples{
# Create a design
dfr <- cr.w(1:1000, "A", "B", 50, 10)
dfr <- dfr$book
# Create some random data
dfr$y <- rnorm(1125)
# Delete some values
dfr[c(11, 165, 569, 914), 'y'] <- NA
# Check the design
ck.w(dfr, "y", "geno", "A", "B", "row", "col", 10)
}
\author{
Raul Eyzaguirre.
}
