% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_pt.R
\name{clean.data.pt}
\alias{clean.data.pt}
\title{Clean potato data}
\usage{
clean.data.pt(dfr, f = 10)
}
\arguments{
\item{dfr}{The name of the data frame.}

\item{f}{Factor for extreme values detection.}
}
\value{
It returns the data frame with all impossible values set to \code{NA},
some values set to \code{0} and a list of warnings with all the rows that have
been modified.
}
\description{
This is a wrapper for functions \code{setna.pt} and \code{setzero.pt} and
applies both in that order.
}
\details{
The data frame must use the labels (lower or upper case) listed
in function \code{check.names.pt}. Then functions \code{setna.pt} and
\code{setzero.pt} are applied to the data.
}
\author{
Raul Eyzaguirre.
}
