% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conplot.r
\name{conplot}
\alias{conplot}
\title{Plot connectivity among trials}
\usage{
conplot(trait, geno, trial, dfr)
}
\arguments{
\item{trait}{The name of the column for the trait to plot.}

\item{geno}{The name of the column that identifies the genotypes.}

\item{trial}{The name of the column that identifies the trials.}

\item{dfr}{The name of the data frame.}
}
\value{
It returns a plot with the number or percentage of genotypes in
common among trials.
}
\description{
Function to plot number of genotypes in common among trials.
}
\examples{
conplot("ttwp", "geno", "trial.name", ptfs)
}
\author{
Raul Eyzaguirre
}
