% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_co.R
\name{convert.co}
\alias{convert.co}
\title{CO numbers}
\usage{
convert.co(
  dfr,
  direction = c("labels.to.co", "co.to.labels"),
  crop = c("auto", "pt", "sp")
)
}
\arguments{
\item{dfr}{The name of the data frame.}

\item{direction}{\code{labels.to.co} or \code{co.to.labels}}

\item{crop}{\code{"auto"} for autodetection or \code{"pt"} for potato and \code{"sp"} for sweetpotato.}
}
\value{
It returns a data frame with all short labels for traits interchanged
with Crop Ontology variable numbers.
}
\description{
Interchanges potato and sweetpotato short labels with Crop Ontology variable numbers.
}
\details{
All labels (lower or upper case) listed in function \code{check.names}
are interchanged with CO variable numbers.
}
\examples{
convert.co(potatoyield)
convert.co(pjpz09)
}
\author{
Raul Eyzaguirre.
}
