% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_a01d.R
\name{cr.a01d}
\alias{cr.a01d}
\title{Alpha (0,1) design}
\usage{
cr.a01d(geno, check, nb, k, nc = NULL, breakib = c("no", "yes"),
  serpentine = c("yes", "no"), alongside = c("no", "rows", "columns"))
}
\arguments{
\item{geno}{The list of genotypes.}

\item{check}{The list of checks to repeat in each incomplete block (optional).}

\item{nb}{Number of complete blocks.}

\item{k}{Size for the incomplete blocks.}

\item{nc}{Number of available columns on the field for each complete block .}

\item{breakib}{If incomplete blocks should be broken in more than one
row, default \code{"no"}.}

\item{serpentine}{\code{"yes"} or \code{"no"}, default \code{"yes"}.}

\item{alongside}{\code{"no"} for independent blocks, or \code{"rows"}
or \code{"columns"} if blocks are together alongside rows or columns.}
}
\value{
It returns the fieldbook and fieldplan.
}
\description{
This function creates the fieldbook and fieldplan for an alpha (0,1) design.
}
\details{
The genotypes are randomly allocated on a field following an alpha
(0,1) design. In this design each block is a complete replication that is
divided into \code{s} incomplete blocks of size \code{k}. For any pair of
genotypes, the number of times they concur into the same incomplete block
is 0 or 1 (hence the name of the design). There are 4 options for this design:
\itemize{
 \item \code{nb = 2} and \code{k <= s}.
 \item \code{nb = 3}, \code{s} odd, and \code{k <= s}.
 \item \code{nb = 3}, \code{s} even, and \code{k <= s-1}.
 \item \code{nb = 4}, \code{s} even and not a multiple of 3, and \code{k <= s}.
}
The blocks and incomplete blocks are disposed alongside the rows.
}
\examples{
cr.a01d(1:100, NULL, 2, 5)
cr.a01d(1:100, NULL, 3, 5, 28)
}
\author{
Raul Eyzaguirre.
}
