% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_abd.R
\name{cr.abd}
\alias{cr.abd}
\title{Augmented Block Design}
\usage{
cr.abd(geno, checks, nb, nc = NULL, serpentine = c("yes", "no"),
  alongside = c("no", "rows", "columns"))
}
\arguments{
\item{geno}{The list of genotypes.}

\item{checks}{The list of checks.}

\item{nb}{Number of blocks.}

\item{nc}{Number of available columns on the field.}

\item{serpentine}{\code{"yes"} or \code{"no"}, default \code{"yes"}.}

\item{alongside}{\code{"no"} for independent blocks, or \code{"rows"}
or \code{"columns"} if blocks are together alongside rows or columns.}
}
\value{
It returns the fieldbook and fieldplan.
}
\description{
This function creates the fieldbook and fieldplan for an ABD.
}
\examples{
checks <- paste("ch", 1:4, sep = "_")
cr.abd(1:40, checks, 4, 10)
cr.abd(1:50, checks, 3, 7)
}
\author{
Raul Eyzaguirre.
}
