% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_crd.R
\name{cr.crd}
\alias{cr.crd}
\title{Completely Randomized Design}
\usage{
cr.crd(geno, nrep, nc)
}
\arguments{
\item{geno}{The list of genotypes.}

\item{nrep}{Number of replications.}

\item{nc}{Number of available columns on the field.}
}
\value{
It returns the fieldbook and fieldplan.
}
\description{
This function creates the fieldbook and fieldplan for a CRD.
}
\examples{
cr.crd(1:20, 3, 12)
cr.crd(1:20, 2, 7)
}
\author{
Raul Eyzaguirre.
}
