% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_factorial.R
\name{cr.f}
\alias{cr.f}
\title{Create design for a factorial experiment}
\usage{
cr.f(A, B, C = NULL, D = NULL, E = NULL, design = c("crd", "rcbd"),
  nrep, nc)
}
\arguments{
\item{A}{The levels of factor A.}

\item{B}{The levels of factor B.}

\item{C}{The levels of factor C.}

\item{D}{The levels of factor D.}

\item{E}{The levels of factor E.}

\item{design}{The design, a crd or a rcbd.}

\item{nrep}{Number of replications or blocks.}

\item{nc}{Number of columns.}
}
\value{
It returns the fieldbook and fieldplan.
}
\description{
This function creates the fieldbook and fieldplan for a factorial experiment
with 2 to 5 factors following a CRD or a RCBD.
}
\details{
The treatments are randomly allocated on a field following a CRD or a RCBD.
}
\examples{
A <- paste("a", 1:5, sep = "")
B <- paste("b", 1:3, sep = "")
cr.f(A, B, design = "rcbd", nrep = 3, nc = 12)
}
\author{
Raul Eyzaguirre.
}
