% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_f.R
\name{cr.f}
\alias{cr.f}
\title{Create design for a factorial experiment}
\usage{
cr.f(fnames, flevels, design = c("crd", "rcbd"), nrep, nc)
}
\arguments{
\item{fnames}{Factor's names.}

\item{flevels}{A list with the factor's levels.}

\item{design}{The design, \code{crd} or \code{rcbd}.}

\item{nrep}{Number of replications or blocks.}

\item{nc}{Number of available columns on the field.}
}
\value{
It returns the fieldbook and fieldplan.
}
\description{
This function creates the fieldbook and fieldplan for a factorial experiment
following a CRD or a RCBD.
}
\examples{
A <- paste0("a", 1:5)
B <- paste0("b", 1:3)
C <- paste0("c", 1:2)
cr.f(c("A", "B", "C"), list(A, B, C), "rcbd", 3, 12)
}
\author{
Raul Eyzaguirre.
}
