% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_lsd.R
\name{cr.lsd}
\alias{cr.lsd}
\title{Latin Square Design}
\usage{
cr.lsd(geno, serpentine = c("yes", "no"))
}
\arguments{
\item{geno}{The list of genotypes.}

\item{serpentine}{\code{"yes"} or \code{"no"}, default \code{"yes"}.}
}
\value{
It returns the fieldbook and fieldplan.
}
\description{
This function creates the fieldbook and fieldplan for a LSD.
}
\examples{
cr.lsd(c("A", "B", "C"))
cr.lsd(c("A", "B", "C", "D", "E"))
}
\author{
Raul Eyzaguirre.
}
