% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_rcbd.R
\name{cr.rcbd}
\alias{cr.rcbd}
\title{Randomized Complete Block Design}
\usage{
cr.rcbd(geno, nb, nc, serpentine = c("yes", "no"))
}
\arguments{
\item{geno}{The list of genotypes.}

\item{nb}{Number of blocks.}

\item{nc}{Number of available columns on the field.}

\item{serpentine}{\code{"yes"} or \code{"no"}, default \code{"yes"}.}
}
\value{
It returns the fieldbook and fieldplan.
}
\description{
This function creates the fieldbook and fieldplan for a RCBD.
}
\examples{
cr.rcbd(1:20, 3, 10)
cr.rcbd(1:20, 2, 7)
}
\author{
Raul Eyzaguirre.
}
