% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_rcbd.R
\name{cr.rcbd}
\alias{cr.rcbd}
\title{Randomized Complete Block Design}
\usage{
cr.rcbd(geno, nb, nc = NULL, serpentine = c("yes", "no"),
  alongside = c("no", "rows", "columns"))
}
\arguments{
\item{geno}{The list of genotypes.}

\item{nb}{Number of blocks.}

\item{nc}{Number of available columns on the field for each block.}

\item{serpentine}{\code{"yes"} or \code{"no"}, default \code{"yes"}.}

\item{alongside}{\code{"no"} for independent blocks, or \code{"rows"}
or \code{"columns"} if blocks are together alongside rows or columns.}
}
\value{
It returns the fieldbook and fieldplan.
}
\description{
This function creates the fieldbook and fieldplan for a RCBD.
}
\examples{
cr.rcbd(1:20, 3, 10)
cr.rcbd(1:20, 2, 7)
}
\author{
Raul Eyzaguirre.
}
