% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_strd.R
\name{cr.strd}
\alias{cr.strd}
\title{Strip-Split-Plot Design}
\usage{
cr.strd(A, B, nrep)
}
\arguments{
\item{A}{The levels of factor A (row factor).}

\item{B}{The levels of factor B (column factor).}

\item{nrep}{Number of replications (or blocks).}
}
\value{
It returns the fieldbook and fieldplan.
}
\description{
This function creates the fieldbook and fieldplan for a Strip-Split-Plot design.
}
\details{
The levels of the factors are randomly allocated on a field
following a Strip-Split-Plot design. Row and column numbers are specific
to each replication. Each replication is a complete block for factor A
and for factor B.
}
\examples{
A <- paste0("a", 1:4)
B <- paste0("b", 1:3)
cr.strd(A, B, 3)
}
\author{
Raul Eyzaguirre.
}
