% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_strd.R
\name{cr.strd}
\alias{cr.strd}
\title{Strip-Split-Plot Design}
\usage{
cr.strd(A, B, nb, serpentine = c("yes", "no"), alongside = c("no",
  "rows", "columns"))
}
\arguments{
\item{A}{The levels of factor A (row factor).}

\item{B}{The levels of factor B (column factor).}

\item{nb}{Number of blocks.}

\item{serpentine}{\code{"yes"} or \code{"no"}, default \code{"yes"}.}

\item{alongside}{\code{"no"} for independent blocks, or \code{"rows"}
or \code{"columns"} if blocks are together alongside rows or columns.}
}
\value{
It returns the fieldbook and fieldplan.
}
\description{
This function creates the fieldbook and fieldplan for a Strip-Split-Plot design.
}
\examples{
A <- paste0("a", 1:4)
B <- paste0("b", 1:3)
cr.strd(A, B, 3)
}
\author{
Raul Eyzaguirre.
}
