% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fb.R
\name{create.fb}
\alias{create.fb}
\title{Create potato and sweetpotato fieldbook}
\usage{
create.fb(
  design,
  crop = c("pt", "sp"),
  label = c("standard", "CO"),
  minimal = TRUE,
  add = NULL,
  computation = FALSE
)
}
\arguments{
\item{design}{The name of the design data frame.}

\item{crop}{\code{"pt"} for potato and \code{"sp"} for sweetpotato.}

\item{label}{Use \code{standard} for standard sweetpotato short labels or
\code{CO} for CO numbers. Default is \code{standard}.}

\item{minimal}{Logical, if \code{TRUE}, a minimal list of variables is included.
Default is \code{TRUE}.}

\item{add}{Additional variables to include. Only if \code{minimal = TRUE}.}

\item{computation}{Logical, if \code{TRUE}, computed variables are included.
Only if \code{minimal = FALSE}. Default is \code{FALSE}.}
}
\value{
It returns a data frame with fieldbook design and variables.
}
\description{
Creates a potato and sweetpotato fieldbook with short labels or CO numbers.
}
\details{
Only labels listed in functions \code{pt.ont()} and \code{sp.ont()}
are valid. Uppercase labels are converted to lowercase.
}
\examples{
book <- cr.rcbd(1:20, 3, 10)$book
# Get fieldbook with minimal set of variables for potato
create.fb(book, 'pt')
# Add additional variables
create.fb(book, 'pt', add = c('fedw', 'zndw'))
# Get the fieldbook with CO numbers
create.fb(book, 'pt', 'CO', add = c('dm', 'fedw', 'zndw'))
# Get fieldbook with minimal set of variables for sweetpotato
create.fb(book, 'sp')
# Add additional variables
create.fb(book, 'sp', add = c('bc', 'fe', 'zn'))
# Get the fieldbook with CO numbers
create.fb(book, 'sp', 'CO', add = c('bc', 'fe', 'zn'))
}
\author{
Raul Eyzaguirre.
}
