% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docomp.R
\name{docomp}
\alias{docomp}
\title{Do computations over some factors}
\usage{
docomp(dfr, do, vars, factors, keep = NULL, method = c("fast", "slow"))
}
\arguments{
\item{dfr}{The name of the data frame.}

\item{do}{The computation to perform. Implemented options are \code{count},
\code{mode}, and standard functions like \code{mean}, \code{median},
\code{min}, \code{max}, \code{sd}, \code{var}, \code{sum}, etc.}

\item{vars}{The names of the columns for the variables.}

\item{factors}{The names of the columns for the factors.}

\item{keep}{The names of additional columns to keep.}

\item{method}{Use \code{fast} or \code{slow} method.}
}
\value{
It returns a data frame with the computations.
}
\description{
Do computations for several variables for some specific factors.
}
\details{
This function do a specific computation for all the \code{vars}
for each level's combination of the \code{factors}. Additional columns can be
kept if specified in \code{keep}. All \code{factors} and \code{keep} values
are converted to character. \code{do = "count"} counts the number
of valid cases (excluding missing values).
}
\examples{
# Compute means across replications and then across locations for each genotype
vars <- c("rytha", "bc", "dm", "star", "nocr")
factors <- c("geno", "loc")
output1 <- docomp(spg, "mean", vars, factors)
docomp("mean", vars, "geno", dfr = output1) 
# Compute maxima across replications for each genotype and location.
docomp(spg, "max", vars, factors)
}
\author{
Raul Eyzaguirre.
}
