% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docomp.R
\name{docomp}
\alias{docomp}
\title{Do computations over some factors}
\usage{
docomp(do, traits, factors, add = NULL, dfr)
}
\arguments{
\item{do}{The computation to perform. Implemented options are \code{count},
and standard functions like \code{mean}, \code{median}, \code{min}, \code{max},
\code{sd}, \code{var}, \code{sum}, etc.}

\item{traits}{List of traits.}

\item{factors}{List of factors.}

\item{add}{Additional columns to keep.}

\item{dfr}{The name of the data frame.}
}
\value{
It returns a data frame with the computations.
}
\description{
Do computations for several traits for some specific factors.
}
\details{
This function do a specific computation for all the \code{traits}
for each level's combination of the \code{factors}. Additional columns can be
kept if specified in \code{add}. All \code{factors} and \code{add} values
are converted to character. \code{do = "count"} counts the number
of valid cases (excluding missing values).
}
\examples{
# Compute means across replications and then across locations for each genotype
traits <- c("rytha", "bc", "dm", "star", "nocr")
factors <- c("geno", "loc")
output1 <- docomp("mean", traits, factors, dfr = spg)
docomp("mean", traits, "geno", dfr = output1) 
# Compute maxima across replications for each genotype and location.
docomp("max", traits, factors, dfr = spg)
}
\author{
Raul Eyzaguirre.
}
