% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcb.R
\name{drcb}
\alias{drcb}
\title{Randomized Complete Block Design}
\usage{
drcb(geno, nb, nr, byrow = TRUE)
}
\arguments{
\item{geno}{The list of genotypes.}

\item{nb}{Number of blocks.}

\item{nr}{Number of rows.}

\item{byrow}{Logical. If \code{"TRUE"} (the default) the blocks are
allocated alongside the rows.}
}
\value{
It returns the fieldbook and fieldplan.
}
\description{
This function creates the fieldbook and fieldplan for a RCBD.
}
\details{
The genotypes are randomly allocated on a field following a RCBD.
}
\examples{
drcb(1:20, 3, 12, TRUE)
drcb(1:20, 2, 7, FALSE)
}
\author{
Raul Eyzaguirre.
}
