% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcb.R, R/dwescott.R
\name{dw}
\alias{dw}
\alias{dw}
\title{Unreplicated design with a grid of checks}
\usage{
dw(geno, ch1, ch2, nr, ncb)

dw(geno, ch1, ch2, nr, ncb)
}
\arguments{
\item{geno}{The list of genotypes.}

\item{ch1}{Name of check 1.}

\item{ch2}{Name of check 2.}

\item{nr}{Number of rows.}

\item{ncb}{Number of columns between two check columns.}

\item{nc}{Number of columns between two check columns.}

\item{geno}{The list of genotypes.}

\item{ch1}{Name of check 1.}

\item{ch2}{Name of check 2.}

\item{nr}{Number of rows.}
}
\value{
It returns the fieldbook and fieldplan.

It returns the fieldbook and fieldplan.
}
\description{
This function creates the fieldbook and fieldplan for an unreplicated design
with genotypes randomly allocated on a field with checks following
the method described on Westcott (1981).

This function creates the fieldbook and fieldplan for an unreplicated design
with genotypes randomly allocated on a field with checks following
the method described on Westcott (1981).
}
\details{
The genotypes are randomly allocated on a field with equally spaced
columns of two alternating check varieties.

The genotypes are randomly allocated on a field with equally spaced
columns of two alternating check varieties.
}
\examples{
dw(1:100, "A", "B", 5, 10)
dw(1:100, "A", "B", 5, 10)
}
\references{
Westcott, B. (1981). Two methods for early generation yield assessment in winter wheat.
In: Proc. of the 4th meeting of the Biometrics in Plant Breeding Section of Eucarpia.
INRA Poitier, France, pp 91-95.

Westcott, B. (1981). Two methods for early generation yield assessment in winter wheat.
In: Proc. of the 4th meeting of the Biometrics in Plant Breeding Section of Eucarpia.
INRA Poitier, France, pp 91-95.
}
\author{
Raul Eyzaguirre.

Raul Eyzaguirre.
}
