% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecm.R
\name{ecm}
\alias{ecm}
\title{Estimate Genotypic and Phenotypic Covariance and Correlation Matrices}
\usage{
ecm(traits, geno, env = NULL, rep, data, method = 1)
}
\arguments{
\item{traits}{The traits to include.}

\item{geno}{The genotypes.}

\item{env}{The environments.}

\item{rep}{The replications or blocks. A RCBD is assumed.}

\item{data}{The name of the data frame containing the data.}

\item{method}{The method to compute genotypic covariances. See details.}
}
\value{
It returns the genotypic and phenotypic covariance and correlation matrices.
}
\description{
This function estimates the genotypic and phenotypic covariance and correlation
matrices with data from a RCBD with one or several environments.
}
\details{
If \code{env = NULL} a RCBD with one environment is considered.
If \code{method = 1} the covariance between each pair of traits is computed using the variance
of each trait and the variance of the sum. If \code{method = 2} the covariance matrix is
approximated using the average of the correlation matrices computed with each replication.
}
\examples{
traits <- c("rytha", "bc", "dm", "star", "nocr")
ecm(traits, "geno", "loc", "rep", spg)
}
\author{
Raul Eyzaguirre.
}

