% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecm.R
\name{ecm}
\alias{ecm}
\title{Estimate Genotypic and Phenotypic Covariance and Correlation Matrices}
\usage{
ecm(traits, geno, env = NULL, rep, dfr, method = 1)
}
\arguments{
\item{traits}{The traits to include.}

\item{geno}{The genotypes.}

\item{env}{The environments.}

\item{rep}{The replications or blocks. A RCBD is assumed.}

\item{dfr}{The name of the data frame.}

\item{method}{The method to compute genotypic covariances. See details.}
}
\value{
It returns the genotypic and phenotypic covariance and correlation matrices.
}
\description{
This function estimates the genotypic and phenotypic covariance and correlation
matrices with data from a RCBD with one or several environments.
}
\details{
If \code{env = NULL} a RCBD with one environment is considered.
If \code{method = 1} the covariances between each pair of traits are computed
using the variances of each trait and the variances of the sum. If \code{method = 2}
the covariance matrices are approximated using the average of the correlation
matrices computed with each replication in each environment.
}
\examples{
traits <- c("rytha", "bc", "dm", "star", "nocr")
ecm(traits, "geno", "loc", "rep", spg)
}
\author{
Raul Eyzaguirre.
}
