% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Elston.R
\name{elston}
\alias{elston}
\title{Elston Index}
\usage{
elston(traits, geno, env = NULL, rep = NULL, data, means = "single",
  model = "gxe", lb = 1)
}
\arguments{
\item{traits}{List of traits.}

\item{geno}{The genotypes.}

\item{env}{The environments.}

\item{rep}{The replications or blocks.}

\item{data}{The name of the data frame containing the data.}

\item{means}{The genotypic means to compute the index, \code{"single"}
or \code{"fitted"}. The default is \code{"single"}. See details for more information.}

\item{model}{Type of model to fit means if \code{means = "fitted"}, \code{"gxe"} for
a model with gxe interaction or \code{"g+e"} for a model without interaction.
The default is \code{"gxe"}. See details for more information.}

\item{lb}{Lower bound. \code{1} for k = min(x) and \code{2} for k = (n*min(x) - max(x))/(n-1)}
}
\value{
It returns a data frame with the genotypic means for each trait, the Elston index,
and the rank for each genotype according to the index.
}
\description{
Function to compute the Elston index (Elston, R. C., 1963).
}
\details{
The Elston index is a weight free index.

If \code{means = "fitted"} and \code{model = "gxe"} then the arguments \code{env} and
\code{rep} must be specified.
If \code{means = "fitted"} and \code{model = "g+e"} then only the argument \code{env}
must be specified.
If \code{means = "single"} and \code{env} and \code{rep} are specified, then
single arithmetic means are computed over the replications for each genotype
at each location and then for each genotype over locations. In any other case
single arithmetic means are computed over all the observations for each genotype.
}
\examples{
# The data
head(spg)
str(spg)

# Run Elston index with all the traits
elston(c("rytha", "bc", "dm", "star", "nocr"), "geno", data = spg)
}
\author{
Raul Eyzaguirre
}
\references{
Elston, R. C. (1963). A weight-free index for the purpose of ranking or selection
with respect to several traits at a time. Biometrics. 19(1): 85-97.
}

