% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elston.R
\name{elston}
\alias{elston}
\title{Elston Index}
\usage{
elston(traits, geno, env = NULL, rep = NULL, dfr, means = c("single",
  "fitted"), lb = 1)
}
\arguments{
\item{traits}{List of traits.}

\item{geno}{The genotypes.}

\item{env}{The environments.}

\item{rep}{The replications.}

\item{dfr}{The name of the data frame.}

\item{means}{The genotypic means to compute the index, \code{"single"} or
\code{"fitted"}. The default is \code{"single"}. See details for more information.}

\item{lb}{Lower bound. \code{1} for \eqn{k = min(x)} and \code{2} for
\eqn{k = (n \times min(x) - max(x)) / (n - 1)}}
}
\value{
It returns a data frame with the genotypic means for each trait,
the Elston index, and the rank for each genotype according to the index.
}
\description{
Function to compute the Elston index (Elston, R. C., 1963).
}
\details{
The Elston index is a weight free index. It is assumed that all the
traits are in the same direction where the highest the value the better the
genotype. To include any trait with an opposite direction it must be transformed
by multiplication by \code{-1} before.

If \code{means = "fitted"} then a linear model is fitted and used to estimate
the means for each genotype.
If \code{means = "single"} and \code{env} is specified, then single arithmetic
means are computed over the replications for each genotype at each environment
and then for each genotype over environments.
If \code{means = "single"} and \code{env} is not specified, then single arithmetic
means are computed over all the observations for each genotype.
}
\examples{
elston(c("rytha", "bc", "dm", "star", "nocr"), "geno", dfr = spg)
}
\references{
Elston, R. C. (1963). A weight-free index for the purpose of ranking or selection
with respect to several traits at a time. Biometrics. 19(1): 85-97.
}
\author{
Raul Eyzaguirre.
}
