% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/Elston.R
\name{elston}
\alias{elston}
\title{Elston Index}
\usage{
elston(traits, geno, data, lb = 1)
}
\arguments{
\item{traits}{List of traits.}

\item{geno}{The genotypes.}

\item{data}{The name of the data frame containing the data.}

\item{lb}{Lower bound. 1 for k = min(x) and 2 for k = (n*min(x) - max(x))/(n-1)}
}
\value{
It returns the Elston index value and the Elston index value
sorted in descending order.
}
\description{
Function to compute the Elston index (Elston, R. C., 1963).
}
\details{
The Elston index is a weight free index.
}
\examples{
# The data
head(spg)
str(spg)

# Run Elston index with all the traits
elston(c("rytha", "bc", "dm", "star", "nocr"), "geno", spg)
}
\author{
Raul Eyzaguirre
}
\references{
Elston, R. C. (1963). A weight-free index for the purpose of ranking or selection
with respect to several traits at a time. Biometrics. 19(1): 85-97.
}

