% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{lxt}
\alias{lxt}
\title{Line by tester data}
\format{A data frame with 4 columns and 92 rows.}
\source{
Singh and Chaudhary, 1979. Biometrical Methods in Quantitative Genetic Analysis.
}
\usage{
lxt
}
\description{
Line by tester data with four replications.
}
