% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/METdata.R
\docType{data}
\name{met8x12}
\alias{met8x12}
\title{Yields for a multi-environment trial (MET)}
\format{A data frame with 4 columns and 288 rows.}
\source{
International Potato Center, sweetpotato experimental data.
}
\usage{
met8x12
}
\description{
This data set gives the yields per plot for 8 genotypes
in 12 environments with a RCBD with 3 blocks in each environment.
}

