% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/msdplot.r
\name{msdplot}
\alias{msdplot}
\title{Plot means and standard deviations with a dotplot}
\usage{
msdplot(trait, groups, data, conf = 0.95, nmax = 10, dotplot = "TRUE",
  sort.means = "none", main = NULL, xlab = "groups", ylab = "",
  colors = c("orange", "orange", "black"), pch = 4, lwd = 2, x.las = 1,
  jf = 0.1, dist = 0.1)
}
\arguments{
\item{trait}{The trait to plot.}

\item{groups}{The grouping factor.}

\item{data}{The name of the data frame containing the data.}

\item{conf}{Probability for the confidence limits or number of standard deviations.}

\item{nmax}{Maximum number of points for the compulsory dotplot, default is 10.}

\item{dotplot}{Logical. If \code{TRUE}, a dotplot is shown. If \code{FALSE}
it will only suppress the dots if the number of data points is larger than \code{nmax}.}

\item{sort.means}{Sort for means. Options are \code{"none"}, \code{"increasing"},
and \code{"decreasing"}, \code{"none"} by default.}

\item{main}{Main title.}

\item{xlab}{Title for x axis.}

\item{ylab}{Title for y axis.}

\item{colors}{Color for mean symbols, confidence interval lines, and data points.}

\item{pch}{Ploting character for means.}

\item{lwd}{Width for ploting characters for means.}

\item{x.las}{x axes labels orientation.}

\item{jf}{Jitter factor for dots.}

\item{dist}{Horizontal distance between the means and the dots.}
}
\value{
It returns a plot with the means represented by horizontal lines, a vertical
line representing a confidence limit or a number of standard deviations,
and alternatively the individual data points.
}
\description{
Function to plot means and confidence limits.
}
\details{
An alternative to the controversial dynamite plots.
If \code{conf} is set to a value greater than or equal to 1, then it is interpreted
as number of standard deviations.
}
\examples{
# Simulate some data
mydata <- data.frame(y = rnorm(50, sample(40:60, 5), sample(5:10, 5)),
                     g = rep(1:5, 10))

# Draw the plot
msdplot("y", "g", mydata)
}
\author{
Raul Eyzaguirre
}

