% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdplot.r
\name{msdplot}
\alias{msdplot}
\title{Plot means and standard deviations with a dotplot}
\usage{
msdplot(trait, groups, dfr, conf = 0.95, dotplot = TRUE,
  sort.means = c("none", "increasing", "decreasing"),
  main.title = NULL, color = c("orange", "orange", "black"),
  x.las = 1, jf = 0.1, hsep = 0.1, ...)
}
\arguments{
\item{trait}{The trait to plot.}

\item{groups}{The grouping factor.}

\item{dfr}{The name of the data frame.}

\item{conf}{Probability for the confidence limits or number of standard deviations.}

\item{dotplot}{Logical. If \code{TRUE}, a dotplot is shown.}

\item{sort.means}{Sort for means. Options are \code{"none"}, \code{"increasing"},
and \code{"decreasing"}, \code{"none"} by default.}

\item{main.title}{Main title.}

\item{color}{Color for mean symbols, confidence interval lines, and data points.}

\item{x.las}{x axes labels orientation.}

\item{jf}{Jitter factor for dots.}

\item{hsep}{Horizontal separation between the means and the dots.}

\item{...}{Additional plot arguments.}
}
\value{
It returns a plot with the means represented by horizontal lines, a vertical
line representing a confidence limit or a number of standard deviations,
and alternatively the individual data points.
}
\description{
Function to plot means and confidence limits.
}
\details{
An alternative to the controversial dynamite plots.
If \code{conf} is set to a value greater than or equal to 1, then it is interpreted
as number of standard deviations.
}
\examples{
# Simulate some data
dfr <- data.frame(y = rnorm(50, sample(40:60, 5), sample(5:10, 5)),
                  g = rep(1:5, 10))
msdplot("y", "g", dfr, lwd = 2, pch = 4)
}
\author{
Raul Eyzaguirre
}
