% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/MeansPlot.r
\name{msdplot}
\alias{msdplot}
\title{Plot means and standard deviations}
\usage{
msdplot(trait, groups, data, conf = 0.95, sort.means = "none",
  main.title = NULL, x.title = "groups", y.title = "", col = "black",
  bg = "darkorange", col.lines = "black")
}
\arguments{
\item{trait}{The trait to analyze.}

\item{groups}{The grouping factor.}

\item{data}{The name of the data frame containing the data.}

\item{conf}{Probability for the confidence limits or number of standard deviations.}

\item{sort.means}{Sort for means, \code{"none"} by default.}

\item{main.title}{Main title.}

\item{x.title}{Title for x axis.}

\item{y.title}{Title for y axis.}

\item{col}{Line color for circles.}

\item{bg}{Background color for circles.}

\item{col.lines}{Line color for confidenced interval lines.}
}
\value{
It returns a plot with the means and confidence limits for each group.
}
\description{
Function to plot means and confidence limits.
}
\details{
An alternative to the controversial dynamite plots.
If \code{conf} is set to a value greater than or equal to 1, then it is interpreted
as number of standard deviations.
}
\examples{
# Simulate some data
mydata <- data.frame(y = rnorm(100, sample(80:120, 10), sample(10:20, 10)), g = rep(1:10, 10))

# Draw the plot
msdplot("y", "g", mydata)
}
\author{
Raul Eyzaguirre
}

