% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mve_f.R
\name{mve.f}
\alias{mve.f}
\title{Estimation of missing values for a factorial experiment}
\usage{
mve.f(trait, factors, rep, dfr, maxp = 0.1, tol = 1e-06)
}
\arguments{
\item{trait}{The trait to estimate missing values.}

\item{factors}{The factors.}

\item{rep}{The replications or blocks, \code{NULL} for a CRD.}

\item{dfr}{The name of the data frame.}

\item{maxp}{Maximum allowed proportion of missing values to estimate, default is 10\%.}

\item{tol}{Tolerance for the convergence of the iterative estimation process.}
}
\value{
It returns a data frame with the experimental layout and columns \code{trait}
and \code{trait.est} with the original data and the original data plus the estimated
values.
}
\description{
Function to estimate missing values for factorial experiment with a CRD
or a RCBD by the least squares method.
}
\examples{
# A data frame with some missing values
temp <- asc
temp$dm[c(3, 10, 115)] <- NA
# Estimate the missing values
mve.f("dm", c("geno", "treat"), NULL, temp)
}
\author{
Raul Eyzaguirre.
}
