% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss_val_est.R
\name{mve.met}
\alias{mve.met}
\title{Estimation of missing values for a MET in a RCBD}
\usage{
mve.met(trait, geno, env, rep, data, maxp = 0.1, tol = 1e-06)
}
\arguments{
\item{trait}{The trait to estimate missing values.}

\item{geno}{The genotypes.}

\item{env}{The environments.}

\item{rep}{The replications.}

\item{data}{The name of the data frame.}

\item{maxp}{Maximum allowed proportion of missing values to estimate, default is 10\%.}

\item{tol}{Tolerance for the convergence of the iterative estimation process.}
}
\value{
It returns a data frame with the experimental layout and columns \code{trait}
and \code{trait.est} with the original data and the original data plus the estimated values.
}
\description{
Function to estimate missing values for a Multi Environment Trial (MET) with a
Randomized Complete Block Design (RCBD) by the least squares method.
}
\details{
A \code{data.frame} with data for a MET in a RCBD with at least two replications
and at least one datum for each genotype must be loaded. Experimental data
with only one replication, any genotype without data, or more missing values than
specified in \code{maxp} will generate an error message.
}
\examples{
# The data
str(met8x12)

# Estimate the missing values
mve.met("y", "geno", "env", "rep", met8x12)
}
\author{
Raul Eyzaguirre.
}
