% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss_val_est.R
\name{mve.rcbd}
\alias{mve.rcbd}
\title{Estimation of missing values for a RCBD}
\usage{
mve.rcbd(trait, treat, rep, data, maxp = 0.1, tol = 1e-06)
}
\arguments{
\item{trait}{The trait to estimate missing values.}

\item{treat}{The treatments.}

\item{rep}{The replications.}

\item{data}{The name of the data frame.}

\item{maxp}{Maximum allowed proportion of missing values to estimate, defaults to 10\%.}

\item{tol}{Tolerance for the convergence of the iterative estimation process.}
}
\value{
It returns a data frame with the experimental layout and columns \code{trait}
and \code{trait.est} with the original data and the original data plus the estimated values.
}
\description{
Function to estimate missing values for a Randomized Complete Block Design (RCBD) by
the least squares method.
}
\details{
A \code{data.frame} with data for a RCBD with at least two replications
and at least one datum for each treatment must be loaded. Experimental data
with only one replication, any treatment without data, or more missing values than
specified in \code{maxp} will generate an error message.
}
\examples{
temp <- subset(met8x12, env == "TM80N")
mve.rcbd("y", "geno", "rep", temp)
}
\author{
Raul Eyzaguirre.
}
