% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mve_rcbd.R
\name{mve.rcbd}
\alias{mve.rcbd}
\title{Estimation of missing values for a RCBD}
\usage{
mve.rcbd(trait, geno, rep, dfr, maxp = 0.1, tol = 1e-06)
}
\arguments{
\item{trait}{The trait to estimate missing values.}

\item{geno}{The genotypes.}

\item{rep}{The replications.}

\item{dfr}{The name of the data frame.}

\item{maxp}{Maximum allowed proportion of missing values to estimate, defaults to 10\%.}

\item{tol}{Tolerance for the convergence of the iterative estimation process.}
}
\value{
It returns a data frame with the experimental layout and columns \code{trait}
and \code{trait.est} with the original data and the original data plus the estimated values.
}
\description{
Function to estimate missing values for a Randomized Complete Block Design (RCBD) by
the least squares method.
}
\examples{
temp <- met8x12[met8x12$env == "TM80N", ]
mve.rcbd("y", "geno", "rep", temp)
}
\author{
Raul Eyzaguirre.
}
