% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PesekBaker.R
\name{pesekbaker}
\alias{pesekbaker}
\title{Pesek-Baker Index}
\usage{
pesekbaker(traits, geno, env, rep = NULL, data, means = "single",
  model = "gxe", dgg = NULL, units = "sdu", sf = 0.1)
}
\arguments{
\item{traits}{List of traits.}

\item{geno}{The genotypes.}

\item{env}{The environments.}

\item{rep}{The replications or blocks. Must be defined if \code{model = "gxe"}.}

\item{data}{The name of the data frame containing the data.}

\item{means}{The genotypic means to compute the index, \code{"single"}
or \code{"fitted"}. The default is \code{"single"}. See details for more information.}

\item{model}{Type of model, \code{"gxe"} for a model with gxe interaction or \code{"g+e"}
for a model without interaction. The default is \code{"gxe"}. See details for more information.}

\item{dgg}{Desired genetic gains. The default is one standard deviation for each trait.}

\item{units}{Units for dgg, \code{"actual"} or \code{"sdu"}. See details for more information.}

\item{sf}{Selected fraction. The default is 0.1.}
}
\value{
It returns:
\itemize{
\item \code{$Desired.Genetic.Gains}, the desired genetic gains in actual units,
\item \code{$Standard.Deviations}, the estimated standard deviations,
\item \code{$Genetic.Variances}, the estimated genetic variances,
\item \code{$Correlation.Matrix}, the estimated correlation matrix,
\item \code{$Index.Coefficients}, the index coefficients,
\item \code{$Response.to.Selection}, the response to selection,
\item \code{$Std.Response.to.Selection}, the standardized response to selection, and
\item \code{$Pesek.Baker.Index}, a data frame with the genotypic means for each trait,
the Pesek-Baker index, and the rank for each genotype according to the index.
}
}
\description{
Function to compute the Pesek-Baker index (Pesek, J. and R.J. Baker., 1969).
}
\details{
The Pesek-Baker is an index where relative economic weights have been replaced
by desired gains.

By default a model with components for genotypes, environments, genotypes by environments
interaction and replications nested into environments is fitted (\code{model = "gxe"}).
If \code{model = "g+e"} then a model with components for genotypes
and environments is fitted, and in this case the gxe variance includes the gxe plus the
error variance. Response to selection is only computed when \code{model = "gxe"}.

If \code{means = "fitted"} then the model specified in \code{model} is used to fit
the means of the genotypes. Otherwise single arithmetic means are computed over the
replications for each genotype at each location and then for each genotype over locations.

If \code{dgg} is not specified, the standard deviations of the traits are used.
It means that the desired genetic gains are equal to one standard deviation for
each trait. \code{dgg} can be specified in actual units (\code{units = "actual"}) or in
standard deviations (\code{units = "sdu"}), defaults to \code{"sdu"}. For example,
if you have a trait which is expressed in kilograms and with a standard deviation of
5 kilograms, typing \code{dgg = 2} means a desired genetic gain of 2 standard deviations
that corresponds to 10 kilograms. If you type \code{dgg = 2} and \code{units = "actual"}
then this means a desired genetic gain of 2 kilograms. If \code{dgg = NULL} then the
desired genetic gain will be one standard deviation, no matter if \code{units} is set
as \code{"actual"} or \code{"sdu"}.

To compute the index the package \code{lme4} is needed.
}
\examples{
# The data
head(spg)
str(spg)

# Run Pesek-Baker index with all the traits
pesekbaker(c("rytha", "bc", "dm", "star", "nocr"), "geno", "loc", "rep", spg)

# Use different desired genetic gains for each trait,
# more weight on bc and dm, less on star and nocr.
pesekbaker(c("rytha", "bc", "dm", "star", "nocr"), "geno", "loc", "rep", spg,
           dgg = c(1, 1.5, 1.5, 0.8, 0.8))
}
\author{
Raul Eyzaguirre
}
\references{
Pesek, J. and R.J. Baker.(1969). Desired improvement in relation to selection indices.
Can. J. Plant. Sci. 9:803-804.
}

