% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesekbaker.R
\name{pesekbaker}
\alias{pesekbaker}
\title{Pesek-Baker Index}
\usage{
pesekbaker(traits, geno, env = NULL, rep, dfr, means = c("single",
  "fitted"), dgg = NULL, units = c("sdu", "actual"), sf = 0.1,
  method = 1)
}
\arguments{
\item{traits}{List of traits.}

\item{geno}{The genotypes.}

\item{env}{The environments.}

\item{rep}{The replications.}

\item{dfr}{The name of the data frame.}

\item{means}{The genotypic means to compute the index, \code{"single"}
or \code{"fitted"}. The default is \code{"single"}. See details for more information.}

\item{dgg}{Desired genetic gains. The default is one standard deviation for each trait.}

\item{units}{Units for dgg, \code{"actual"} or \code{"sdu"}. See details for more information.}

\item{sf}{Selected fraction. The default is 0.1.}

\item{method}{The method to compute genotypic covariances. See ?ecm for details.}
}
\value{
It returns:
\itemize{
\item \code{$Desired.Genetic.Gains}, the desired genetic gains in actual units,
\item \code{$Standard.Deviations}, the estimated genotypic standard deviations,
\item \code{$Index.Coefficients}, the index coefficients,
\item \code{$Response.to.Selection}, the response to selection,
\item \code{$Std.Response.to.Selection}, the standardized response to selection, and
\item \code{$Pesek.Baker.Index}, a data frame with the genotypic means for each trait,
the Pesek-Baker index, and the rank for each genotype according to the index.
}
}
\description{
Function to compute the Pesek-Baker index (Pesek, J. and R.J. Baker., 1969).
}
\details{
The Pesek-Baker is an index where relative economic weights have been replaced
by desired gains.

If \code{means = "fitted"} then fitted means under the model are used for each genotype.
Otherwise single arithmetic means are computed over the replications for each genotype
at each environment and then for each genotype over environments.

If \code{dgg} is not specified, the genotypic standard deviations of the traits are used.
It means that the desired genetic gains are equal to one standard deviation for
each trait. \code{dgg} can be specified in actual units (\code{units = "actual"}) or in
standard deviations (\code{units = "sdu"}), defaults to \code{"sdu"}. For example,
if you have a trait which is expressed in kilograms and with a standard deviation of
5 kilograms, typing \code{dgg = 2} means a desired genetic gain of 2 standard deviations
that corresponds to 10 kilograms. If you type \code{dgg = 2} and \code{units = "actual"}
then this means a desired genetic gain of 2 kilograms. If \code{dgg = NULL} then the
desired genetic gains will be one standard deviation, no matter if \code{units} is set
as \code{"actual"} or \code{"sdu"}.
}
\examples{
traits <- c("rytha", "bc", "dm", "star", "nocr")
pesekbaker(traits, "geno", "loc", "rep", spg)
# More weight on bc and dm, less on star and nocr
pesekbaker(traits, "geno", "loc", "rep", spg, dgg = c(1, 1.5, 1.5, 0.8, 0.8))
}
\references{
Pesek, J. and R.J. Baker.(1969). Desired improvement in relation to selection indices.
Can. J. Plant. Sci. 9:803-804.
}
\author{
Raul Eyzaguirre.
}
