% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/PesekBaker.R
\name{pesekbaker}
\alias{pesekbaker}
\title{Pesek-Baker Index}
\usage{
pesekbaker(traits, geno, env, rep, data, dgg = NULL, units = "sdu",
  sf = 0.1)
}
\arguments{
\item{traits}{List of traits.}

\item{geno}{The genotypes.}

\item{env}{The environments.}

\item{rep}{The replications or blocks.}

\item{data}{The name of the data frame containing the data.}

\item{dgg}{Desired genetic gains, defaults to one standard deviation.}

\item{units}{Units for dgg, \code{"actual"} or \code{"sdu"}. See details for more information.}

\item{sf}{Selected fraction, defaults to 0.1.}
}
\value{
It returns the desired genetic gains in actual units,
the estimated standard deviations,
the estimated genetic variances,
the estimated correlation matrix,
the index coefficients,
the response to selection,
the standardized response to selection,
the Pesek-Baker index value,
and the Pesek-Baker index value sorted in descending order.
}
\description{
Function to compute the Pesek-Baker index (Pesek, J. and R.J. Baker., 1969).
}
\details{
The Pesek-Baker is an index where relative economic weights have been replaced
by desired gains. If \code{dgg} is not specified, the standard deviations of the traits
are used. It means that the desired genetic gains are equal to one standard deviation for
each trait. \code{dgg} can be specified in actual units (\code{units = "actual"}) or in
standard deviations (\code{units = "sdu"}), defaults to \code{"sdu"}. For example,
if you have a trait which is expressed in kilograms and with a standard deviation of
5 kilograms, typing \code{dgg = 2} means a desired genetic gain of 2 standard deviations
that corresponds to 10 kilograms. If you type \code{dgg = 2} and \code{units = "actual"}
then this means a desired genetic gain of 2 kilograms. If \code{dgg = NULL} then the
desired genetic gain will be one standard deviation, no matter if \code{units} is set
as \code{"actual"} or \code{"sdu"}.
To compute the index the package \code{lme4} is needed.
}
\examples{
# The data
head(spg)
str(spg)

# Run Pesek-Baker index with all the traits
pesekbaker(c("rytha", "bc", "dm", "star", "nocr"), "geno", "loc", "rep", spg)

# Use 2 standard deviations as desired genetic gains for each trait
pesekbaker(c("rytha", "bc", "dm", "star", "nocr"), "geno", "loc", "rep", dgg = 2, spg)
}
\author{
Raul Eyzaguirre
}
\references{
Pesek, J. and R.J. Baker.(1969). Desired improvement in relation to selection indices.
Can. J. Plant. Sci. 9:803-804.
}

