% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pjpz09}
\alias{pjpz09}
\title{Data for a yield trial}
\format{A data frame with 10 columns and 170 rows.}
\source{
International Potato Center, sweetpotato experimental data.
}
\usage{
pjpz09
}
\description{
This data set contains data for number of plants sowed (\code{nops}),
number of plants harvested (\code{noph}), number of commercial roots (\code{nocr}),
number of non commercial roots (\code{nonc}), commercial root weight (\code{crw}),
non commercial root weght (\code{ncrw}), total root weight (\code{trw}), and vine
weight (\code{vw}) for an experiment with 85 sweetpotato genotypes (\code{geno})
and 2 replications (\code{rep}) with a RCBD.
}
