% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tai.R
\name{plot.tai}
\alias{plot.tai}
\title{Tai plot.}
\usage{
\method{plot}{tai}(x, title = NULL, color = c("darkorange", "black",
  "gray"), size = c(1, 1), ...)
}
\arguments{
\item{x}{An object of class \code{tai}.}

\item{title}{Main title for plot.}

\item{color}{Color for symbols, labels and lines.}

\item{size}{Relative size for symbols and labels.}

\item{...}{Additional plot arguments.}
}
\value{
It returns the Tai graph for stability analysis.
}
\description{
This function produces a Tai's stability analysis plot (Tai, G. C. C., 1971).
}
\details{
It produces a Tai's stability analysis plot for an object of class
\code{tai}. See \code{?tai} for additional details.
}
\examples{
model.tai <- tai("y", "geno", "env", "rep", met8x12)
plot(model.tai)
}
\author{
Raul Eyzaguirre.
}
\references{
Tai, G. C. C. (1971). Genotypic Stability Analysis and Its Application to Potato
Regional Trials, Crop Science, Vol 11.
}

