% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{rm.fna}
\alias{rm.fna}
\title{Remove factors with NA values}
\usage{
rm.fna(factors, dfr)
}
\arguments{
\item{factors}{The factors}

\item{dfr}{The name of the data frame}
}
\value{
The number of rows in the data frame with missing values for factors
(\code{nmis.fac}), and the data frame after removal of all these rows.
}
\description{
This function removes rows for factors with missing values.
}
\examples{
# Create a design
dfr <- cr.crd(1:10, 3, 10)
dfr <- dfr$book
# Delete some values for classification factors
dfr[5, 2] <- NA
dfr[7, 3] <- NA
dfr[11, 4] <- NA
# Check missing values
rm.fna(c('row', 'col', 'geno'), dfr)
}
\author{
Raul Eyzaguirre.
}
