% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_lna.R
\name{rm.lna}
\alias{rm.lna}
\title{Remove factors' levels without data}
\usage{
rm.lna(trait, factors, dfr)
}
\arguments{
\item{trait}{The trait.}

\item{factors}{The factors.}

\item{dfr}{The name of the data frame.}
}
\value{
The number of rows in the data frame with factor levels without data
(\code{nmis.lev}), and the data frame after removal of all these rows.
}
\description{
This function removes rows for factors' levels without data.
}
\examples{
# Create a design
dfr <- cr.crd(1:10, 3, 10)
dfr <- dfr$book
# Create some random data
dfr$y <- rnorm(30)
# Delete values for some factor levels
dfr[dfr$geno == 5, 'y'] <- NA
# Check missing values
rm.lna('y', 'geno', dfr)
}
\author{
Raul Eyzaguirre.
}
