% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setna.R
\name{setna}
\alias{setna}
\title{Set values to NA or zero.}
\usage{
setna(traits, vines = NULL, roots = NULL, fb)
}
\arguments{
\item{traits}{List of traits.}

\item{vines}{Trait for vines.}

\item{roots}{Trait for roots.}

\item{fb}{The name of the fieldbook data frame.}
}
\value{
It returns a data frame.
}
\description{
Set values to NA or zero for selected traits.
}
\details{
This function sets values to NA or zero with the following rules:
\itemize{
 \item If \code{noph = 0} then all traits are set to \code{NA}.
 \item If \code{noph > 0} then all traits with \code{NA} are set to zero.
 \item If \code{is.na(noph)} and \code{vines = 0} and \code{roots = 0} then set
 \code{noph} to zero and all traits to \code{NA}.
 \item If \code{is.na(noph)} and \code{vines > 0} or \code{roots > 0} then set
 all traits to zero.
 }
}
\examples{
traits <- c('nocr', 'nonc', 'crw', 'ncrw', 'trw', 'vw')
setna(traits, 'vw', 'trw', pjpz09)
}
\author{
Raul Eyzaguirre
}
