% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setna.R
\name{setna}
\alias{setna}
\title{Set values to \code{NA} or \code{0}}
\usage{
setna(dfr)
}
\arguments{
\item{dfr}{The name of the data frame.}
}
\value{
It returns a data frame.
}
\description{
Set values to \code{NA} or \code{0} according to some rules.
}
\details{
Consider the following groups of traits:
\itemize{
 \item \code{rhar} (traits evaluated with roots on harvest): \code{nopr},
 \code{nocr}, \code{nonc}, \code{crw}, \code{ncrw}, \code{tnr}, and \code{trw}.

 \item \code{rpos} (traits evaluated with roots post-harvest): \code{scol},
 \code{fcol}, \code{fcol.cc}, \code{rs}, \code{rf}, \code{damr}, \code{rspr},
 \code{wed}, \code{dmf}, \code{dmd}, \code{dm}, \code{fraw}, \code{fraw1},
 \code{suraw}, \code{suraw1}, \code{straw}, \code{straw1}, \code{coof},
 \code{coof1}, \code{coosu}, \code{coosu1}, \code{coost}, \code{coost1},
 \code{coot}, \code{coot1}, \code{cooap}, \code{cooap1}, \code{fraw2},
 \code{suraw2}, \code{straw2}, \code{coof2}, \code{coosu2}, \code{coost2},
 \code{coot2}, \code{cooap2}, \code{prot}, \code{fe}, \code{zn}, \code{ca},
 \code{mg}, \code{bc}, \code{bc.cc}, \code{tc}, \code{star}, \code{fruc},
 \code{gluc}, \code{sucr}, and \code{malt}.
 
 \item \code{vpre} (traits evaluated with vines pre-harvest): \code{vir},
 \code{vir1}, \code{vir2}, \code{alt}, \code{alt1}, \code{alt2}, and \code{vv}.
 
 \item \code{vpos} (traits evaluated with vines post-harvest): \code{dmvf},
 \code{dmvd}, and \code{dmv}.
}
Values are set to \code{NA} or \code{0} with the following rules:
\itemize{

 \item If \code{nope == 0} and there is some data for \code{vpre}, then
 \code{nope} is set to \code{NA}.
 
 \item If \code{noph == 0} and there is some data for \code{rhar}, then
 \code{noph} is set to \code{NA}.
}
}
\examples{
dfr <- data.frame(nope = c(2, 0, 2, 2, 1),
                  noph = c(1, 0, 2, 0, NA),
                  vv = c(3, 2, 1, NA, 2),
                  crw = c(4, 0, 5, 3, 0),
                  trt4 = c(1, NA, 2, 4, 5))
setna(dfr)
}
\author{
Raul Eyzaguirre
}
