% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setna.R
\name{setna}
\alias{setna}
\title{Set values to NA or zero.}
\usage{
setna(traits, fb)
}
\arguments{
\item{traits}{List of traits.}

\item{fb}{The name of the fieldbook data frame.}
}
\value{
It returns a data frame.
}
\description{
Set values to NA or zero for selected traits.
}
\details{
This function sets values to NA or zero with the following rules:
\itemize{
 \item If all traits are zero or NA, then all traits are set to NA, and if \code{noph}
 exists and it is zero or NA, then it is set to zero.
 \item If there is at least one trait with data and \code{noph} is zero, then
 it is set to NA.
 }
}
\examples{
ddd <- data.frame(noph = c(2, 0, 0, 2, 1),
                  trt1 = c(1, 0, 0, 0, NA),
                  trt2 = c(3, 0, 1, NA, 2),
                  trt3 = c(4, 0, 5, 0, 0))
setna(c('trt1', 'trt2', 'trt3'), ddd)
}
\author{
Raul Eyzaguirre
}
