% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setna.R
\name{setna}
\alias{setna}
\title{Set values to \code{NA} for potato and sweetpotato data.}
\usage{
setna(dfr, f = 10, crop = c("auto", "pt", "sp"))
}
\arguments{
\item{dfr}{The name of the data frame.}

\item{f}{Factor for extreme values detection. See details.}

\item{crop}{\code{"auto"} for autodetection or \code{"pt"} for potato and \code{"sp"} for sweetpotato.}
}
\value{
It returns the data frame with all impossible values set to \code{NA}
and a list of warnings with all the rows that have been modified.
}
\description{
Detect impossible values for potato and sweetpotato data and set them to
missing value (\code{NA}) according to some rules.
}
\details{
The data frame must use the labels (lower or upper case) listed in
functions \code{pt.ont()} and \code{sp.ont()}.

Consider the following groups of variables for potato:
\itemize{
 \item \code{pre} (variables evaluated pre-harvest): \code{ppe}, \code{plant_unif},
 \code{plant_vigor} and \code{se}.
 \item \code{cnn} (continuos non-negative variables): \code{tntpl}, \code{nmtpl},
 \code{ttwp}, \code{ttwpl}, \code{mtwp}, \code{mtwpl}, \code{nomtwp}, \code{mtwci},
 \code{mtwcii}, \code{ttya}, \code{ttyna}, \code{mtya}, and \code{mtyna}.
 \item \code{cpo} (continuous positive variables): \code{atw}, \code{atmw},  
 \code{fwts1}, \code{fwts2}, \code{dwts1}, and \code{dwts2}.
 \item \code{pnn} (percentage non-negative variables): \code{ppe}, \code{pph},  
 \code{fruc}, \code{gluc}, \code{sucr}, and \code{malt}.
 \item \code{ppo} (percentage positive variables): \code{dm}, \code{pro},
 \code{star}, and \code{fiber}.
 \item \code{dnn} (discrete non-negative variables): \code{ntp}, \code{npe}, \code{nph},
 \code{tntp}, \code{nmtp}, \code{nnomtp}, \code{nmtci}, and \code{nmtcii}.
 \item \code{ctg} (categorical variables): \code{plant_unif},
 \code{plant_vigor}, \code{flowering}, \code{rlb}, \code{se}, \code{tuber_apper},
 \code{tub_unif}, \code{tub_size}, \code{chip_color}, \code{num_stolon},
 and \code{leng_stolon}.
}
Values are set to \code{NA} with the following rules:
\itemize{
 \item \code{cnn} variables with negative values are set to \code{NA}.
 \item \code{cpo} variables with non-positive values are set to \code{NA}.
 \item \code{pnn} variables with values out of the [0, 100] interval are set to \code{NA}.  
 \item \code{ppo} with values out of the (0, 100] interval are set to \code{NA}.
 \item \code{dnn} variables with negative and non-integer values are set to \code{NA}.
 \item \code{ctg} variables with out of scale values are set to \code{NA}.
 \item Extreme low and high values are detected using the interquartile range.
 The rule is to detect any value out of the interval
 \eqn{[Q_1 - f \times (m/3 + IQR); Q_3 + f \times (m/3 + IQR)]} where \code{m}
 is the mean. By default \code{f = 10} and if less than 10 a warning is shown.
 Values out of this range are set to \code{NA}.
 \item If \code{npe == 0} and there is some data for any variable,
 then \code{npe} is set to \code{NA}.  
 \item If \code{npe == 0} and there is no data but some variables that are \code{0},
 then all those variables are set to \code{NA}.  
 \item If \code{nph == 0} and there is some data for any non-pre-harvest variable,
 then \code{nph} is set to \code{NA}.
 \item If \code{nmtp == 0} and \code{mtwp > 0}, then \code{nmtp} is set to \code{NA}.
 \item If \code{nmtp > 0} and \code{mtwp == 0}, then \code{mtwp} is set to \code{NA}.
 \item If \code{nnomtp == 0} and \code{nomtwp > 0}, then \code{nnomtp} is set to \code{NA}.
 \item If \code{nnomtp > 0} and \code{nomtwp == 0}, then \code{nomtwp} is set to \code{NA}.
}
Consider the following groups of variables for sweetpotato:
\itemize{
 \item Variables evaluated pre-harvest: \code{vir}, \code{vir1}, \code{vir2},
 \code{alt}, \code{alt1}, \code{alt2}, and \code{vv}.
 \item Variables evaluated with vines non-pre-harvest: \code{vw}, \code{biom},
 \code{biom.d}, \code{vw.d}, \code{fytha}, \code{fytha.aj}, \code{dmvy},
 \code{dmvy.aj}, \code{bytha}, \code{bytha.aj}, \code{dmby}, \code{dmby.aj},
 \code{vpp}, \code{vpsp}, \code{dmvf}, \code{dmvd}, \code{hi}, \code{shi},
 and \code{dmv}.
 \item Variables evaluated only with roots non-pre-harvest: \code{crw},
 \code{ncrw}, \code{trw}, \code{trw.d}, \code{cytha}, \code{cytha.aj},
 \code{rytha}, \code{rytha.aj}, \code{dmry}, \code{dmry.aj}, \code{nrpp},
 \code{nrpsp}, \code{ncrpp}, \code{ncrpsp}, \code{ypp}, \code{ypsp},
 \code{rtyldpct}, \code{rfr}, \code{bc.cc}, \code{fcol.cc}, \code{bc},
 \code{tc}, \code{fe}, \code{zn}, \code{ca}, \code{mg}, \code{dmf},
 \code{dmd}, \code{acrw}, \code{ancrw}, \code{atrw}, \code{ci}, \code{fruc},
 \code{gluc}, \code{sucr}, \code{malt}, \code{dm}, \code{prot}, \code{star},
 \code{nocr}, \code{nonc}, \code{tnr}, \code{scol}, \code{fcol}, \code{fcol2},
 \code{rs}, \code{rf}, \code{rtshp}, \code{damr}, \code{rspr}, \code{alcdam},
 \code{wed}, \code{stspwv}, \code{milldam}, \code{fraw}, \code{suraw},
 \code{straw}, \code{coof}, \code{coosu}, \code{coost}, \code{coot}, and
 \code{cooap} 
 \item \code{cnn} (continuos non-negative variables): \code{vw}, \code{crw},
 \code{ncrw}, \code{trw}, \code{trw.d}, \code{biom}, \code{biom.d},
 \code{cytha}, \code{cytha.aj}, \code{rytha}, \code{rytha.aj}, \code{dmry},
 \code{dmry.aj}, \code{vw.d}, \code{fytha}, \code{fytha.aj}, \code{dmvy},
 \code{dmvy.aj}, \code{bytha}, \code{bytha.aj}, \code{dmby}, \code{dmby.aj},
 \code{nrpp}, \code{nrpsp}, \code{ncrpp}, \code{ncrpsp}, \code{ypp},
 \code{ypsp}, \code{vpp}, \code{vpsp}, \code{rtyldpct}, \code{rfr},
 \code{bc}, \code{tc}, \code{fe}, \code{zn}, \code{ca}, and \code{mg}.
 \item \code{cpo} (continuous positive variables): \code{dmf}, \code{dmd},
 \code{dmvf}, \code{dmvd}, \code{acrw}, \code{ancrw}, and \code{atrw}.
 \item \code{pnn} (percentage non-negative variables): \code{ci}, \code{hi},
 \code{shi}, \code{fruc}, \code{gluc}, \code{sucr}, and \code{malt}.
 \item \code{ppo} (percentage positive variables): \code{dm}, \code{dmv},
 \code{prot}, and \code{star}.
 \item \code{dnn} (discrete non-negative variables): \code{nops}, \code{nope},
 \code{noph}, \code{nopr}, \code{nocr}, \code{nonc}, and \code{tnr}.
 \item \code{ctg} (categorical 1 to 9 variables): \code{vir}, \code{vir1},
 \code{vir2}, \code{alt}, \code{alt1}, \code{alt2}, \code{vv}, \code{scol},
 \code{fcol}, \code{fcol2}, \code{rs}, \code{rf}, \code{rtshp}, \code{damr},
 \code{rspr}, \code{alcdam}, \code{wed}, \code{stspwv}, \code{milldam},
 \code{fraw}, \code{suraw}, \code{straw}, \code{coof}, \code{coosu},
 \code{coost}, \code{coot}, and \code{cooap}.
}
Values are set to \code{NA} with the following rules:
\itemize{
 \item \code{cnn} variables with negative values are set to \code{NA}.
 \item \code{cpo} variables with non-positive values are set to \code{NA}.
 \item \code{pnn} variables with values out of the [0, 100] interval are set to \code{NA}.  
 \item \code{ppo} with values out of the (0, 100] interval are set to \code{NA}.
 \item \code{dnn} variables with negative and non-integer values are set to \code{NA}.
 \item \code{ctg} variables with out of scale values are set to \code{NA}.
 \item Beta carotene values determined by RHS color charts with values different from
 the possible values in the RHS color chart are set to \code{NA}.
 \item Extreme low and high values are detected using the interquartile range.
 The rule is to detect any value out of the interval
 \eqn{[Q_1 - f \times (m/3 + IQR); Q_3 + f \times (m/3 + IQR)]} where \code{m}
 is the mean. By default \code{f = 10} and if less than 10 a warning is shown.
 Values out of this range are set to \code{NA}.
 \item If \code{nope == 0} and there is some data for any variable,
 then \code{nope} is set to \code{NA}.
 \item If \code{nope == 0} and there is no data but some variables that are \code{0},
 then all those variables are set to \code{NA}.
 \item If \code{noph == 0} and there is some data for any non-pre-harvest variable,
 then \code{noph} is set to \code{NA}.
 \item If \code{nopr == 0} and there is some data for any variable evaluated with roots,
 then \code{nopr} is set to \code{NA}.
 \item If \code{noph > 0} and \code{nocr}, \code{nonc}, \code{crw}, \code{ncrw}, 
 and \code{vw} are all 0, then \code{vw} is set to \code{NA}.
 \item If \code{nopr > 0} and \code{nocr}, \code{nonc}, \code{crw}, and \code{ncrw}
 are all 0, then \code{ncrw} and \code{nonc} are both set to \code{NA}.
 \item If \code{nocr == 0} and \code{crw > 0}, then \code{nocr} is set to \code{NA}.
 \item If \code{nocr > 0} and \code{crw == 0}, then \code{crw} is set to \code{NA}.
 \item If \code{nonc == 0} and \code{ncrw > 0}, then \code{nonc} is set to \code{NA}.
 \item If \code{nonc > 0} and \code{ncrw == 0}, then \code{ncrw} is set to \code{NA}.
}
}
\examples{
dfr <- data.frame(mtwp = c(2.2, 5.0, 3.6, 12, 1600, -4, 0),
                  dm = c(21, 23, 105, 24, -3, 30, NA),
                  nmtp = c(1.3, 10, 11, NA, 2, 5, NA))
setna(dfr)
dfr <- data.frame(trw = c(2.2, 5.0, 3.6, 12, 1600, -4),
                  dm = c(21, 23, 105, 24, -3, 30),
                  tnr = c(1.3, 10, 11, NA, 2, 5),
                  scol = c(1, 0, 15, 5, 4, 7),
                  fcol.cc = c(1, 15, 12, 24, 55, 20))
setna(dfr)
}
\author{
Raul Eyzaguirre.
}
