% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setna.R
\name{setna}
\alias{setna}
\title{Set values to \code{NA} for potato and sweetpotato data.}
\usage{
setna(dfr, f = 10, crop = c("auto", "pt", "sp"), checknames = TRUE)
}
\arguments{
\item{dfr}{The name of the data frame.}

\item{f}{Factor for extreme values detection. See details.}

\item{crop}{\code{"auto"} for autodetection or \code{"pt"} for potato and \code{"sp"} for sweetpotato.}

\item{checknames}{Logical indicating if column names should be checked, default \code{TRUE}.}
}
\value{
It returns the data frame with all impossible values set to \code{NA}
and a list of warnings with all the rows that have been modified.
}
\description{
Detect impossible values for potato and sweetpotato data and set them to
missing value (\code{NA}) according to some rules.
}
\details{
The data frame must use the labels (lower or upper case) listed in
functions \code{ptont()} and \code{spont()}.

Values are set to \code{NA} with the following rules:
\itemize{
 \item All values out of the variable scale are set to \code{NA}.
 \item Extreme low and high values are detected using the interquartile range.
 The rule is to detect any value out of the interval
 \eqn{[Q_1 - f \times (m/3 + IQR); Q_3 + f \times (m/3 + IQR)]} where \code{m}
 is the mean. By default \code{f = 10} and if less than 10 a warning is shown.
 Values out of this range are set to \code{NA}.
 \item If the number of plants established is \code{0} and there is some data for
 any variable, then the number of plants established is set to \code{NA}.
 \item If the number of plants established is \code{0} and there is no data but
 some variables that are \code{0}, then all those variables are set to \code{NA}.
 \item If the number of plants harvested is \code{0} and there is some data for
 any non-pre-harvest variable, then the number of plants harvested is set to \code{NA}.
 \item If the number of plants with roots or tubers is \code{0} and there is
 some data for any variable evaluated with roots or tubers, then the number of
 plants with roots or tubers is set to \code{NA}.
 \item If the number of plants with roots or tubers is \code{> 0} and all
 variables at harvest with roots are 0, then the non-commercial yields are
 set to \code{NA}.  
 \item If the number of commercial roots or tubers is \code{0} and the
 commercial root or tuber weight is \code{> 0}, then the number of commercial
 roots or tubers is set to \code{NA}.
 \item If the number of commercial roots or tubers is \code{> 0} and the
 commercial root or tuber weight is \code{0}, then the commercial root or 
 tuber weight is set to \code{NA}.
 \item If the number of non-commercial roots or tubers is \code{0} and the
 non-commercial root or tuber weight is \code{> 0}, then the number of commercial
 roots or tubers is set to \code{NA}.
 \item If the number of non-commercial roots or tubers is \code{> 0} and the
 non-commercial root or tuber weight is \code{0}, then the non-commercial root
 or tuber weight is set to \code{NA}.
}
}
\examples{
dfr <- data.frame(mtwp = c(2.2, 5.0, 3.6, 12, 1600, -4, 0),
                  dm = c(21, 23, 105, 24, -3, 30, NA),
                  nmtp = c(1.3, 10, 11, NA, 2, 5, NA))
setna(dfr)
dfr <- data.frame(trw = c(2.2, 5.0, 3.6, 12, 1600, -4),
                  dm = c(21, 23, 105, 24, -3, 30),
                  tnr = c(1.3, 10, 11, NA, 2, 5),
                  scol = c(1, 0, 15, 5, 4, 7),
                  fcol.cc = c(1, 15, 12, 24, 55, 20))
setna(dfr)
}
\author{
Raul Eyzaguirre.
}
