% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setzero.R
\name{setzero}
\alias{setzero}
\title{Set values to \code{0}.}
\usage{
setzero(dfr)
}
\arguments{
\item{dfr}{The name of the data frame.}
}
\value{
It returns a data frame and a list of warnings with all the rows
that have been modified.
}
\description{
Set values to \code{0} for harvested traits.
}
\details{
This function sets values to \code{0} for all traits at harvest
(\code{noph}, \code{nopr}, \code{vw}, \code{nocr}, \code{nonc}, \code{crw},
and \code{ncrw}) which are \code{NA} according to the following rules:
\itemize{
 \item If \code{noph == 0}, then all traits are set to \code{0}.
 \item If all traits are \code{0}, then \code{noph} is set to \code{0}.
 \item If \code{nopr == 0}, then all traits with exception of \code{vw}
 are set to \code{0}.
 \item If all traits with exception of \code{vw} are \code{0}, then
 \code{nopr} is set to \code{0}.
 \item If \code{nocr == 0}, then \code{crw} is set to \code{0}.
 \item If \code{crw == 0}, then \code{nocr} is set to \code{0}.
 \item If \code{nonc == 0}, then \code{ncrw} is set to \code{0}.
 \item If \code{ncrw == 0}, then \code{nonc} is set to \code{0}.
}
}
\examples{
dfr <- data.frame(noph = c(NA,  0,  3,  3,  3, 3, 3),
                  nopr = c(NA,  0,  0,  2,  3, 3, 3),
                  vw   = c(NA, NA,  3,  2,  3, 3, 6),
                  crw  = c(NA, NA, NA, NA,  8, 2, 0),
                  ncrw = c(NA, NA, NA,  2, NA, 2, 4),
                  nocr = c(NA, NA, NA,  0,  6, 2, 0),
                  nonc = c(NA, NA, NA,  4,  0, 3, 5))
setzero(dfr)
}
\author{
Raul Eyzaguirre
}
