% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spconsis.R
\name{spconsis}
\alias{spconsis}
\title{Check consistency for sweetpotato experimental data}
\usage{
spconsis(data, plot.size, f = 3, width = 240)
}
\arguments{
\item{data}{The name of the data frame.}

\item{plot.size}{Plot size in square meters.}

\item{f}{Factor for extreme values detection. See details.}

\item{width}{Number of columns for the output file.}
}
\value{
It returns a file with name checks.txt with a list of
all rows with some kind of inconsistency and all rows with outliers.
}
\description{
Set of rules to check for consistency of sweetpotato experimental data.
Data labels must be defined as specified in the PROCEDURES FOR THE EVALUATION
AND ANALYSIS OF SWEETPOTATO TRIALS document.
}
\details{
The data frame must use the following labels (lower or upper case):
\itemize{
 \item L       : Locations (LOC is also valid)
 \item Y       : Years
 \item S       : Seasons
 \item G       : Genotypes (GENO is also valid)
 \item NAME    : Names for genotypes
 \item E       : Environments (ENV is also valid)
 \item R       : Replications (REP is also valid)
 \item NOPS    : Number of plants sowed
 \item NOPE    : Number of plants established
 \item VIR1    : Virus symptoms (1-9), first evaluation
 \item VIR2    : Virus symptoms (1-9), second evaluation
 \item VIR3    : Virus symptoms (1-9), third evaluation
 \item ALT1    : Alternaria symptoms (1-9), first evaluation
 \item ALT2    : Alternaria symptoms (1-9), second evaluation
 \item VV1     : Vine vigor (1-9), first evaluation
 \item VV2     : Vine vigor2 (1-9), second evaluation
 \item VW      : Vine weight
 \item NOPH    : Number of plants harvested
 \item NOPR    : Number of plants with roots
 \item NOCR    : Number of commercial roots
 \item NONC    : Number of non commercial roots
 \item CRW     : Commercial root weight
 \item NCRW    : Non commercial root weight
 \item RFCP    : Root primary flesh color using CIP color charts
 \item RFCS    : Root secondary flesh color using CIP color charts
 \item SCOL    : Storage root skin color (1-9)
 \item FCOL    : Storage root flesh color (1-9)
 \item RFCP    : Storage root primary flesh color (1-9)
 \item RFCS    : Storage root secondary flesh color (1-9)
 \item RS      : Root size (1-9)
 \item RF      : Root form (1-9)
 \item DAMR    : Root defects (1-9)
 \item RSPR    : Root sprouting (1-9)
 \item WED1    : Weevil damage (1-9), first evaluation
 \item WED2    : Weevil damage2 (1-9), second evaluation
 \item DMF     : Fresh weight of roots for dry matter assessment
 \item DMD     : Dry weight of DMF samples
 \item DM      : Storage root dry matter content (\%)
 \item DMRY    : Dry matter root yield
 \item DMVF    : Fresh weight vines for dry matter assessment
 \item DMVD    : Dry weight of DMVF samples
 \item DMV     : Vines dry matter content (\%)
 \item DMFY    : Dry matter foliage yield
 \item FRAW1   : Root fiber (1-9), first determination
 \item SURAW1  : Root sugar (1-9), first determination
 \item STRAW1  : Root starch (1-9), first determination
 \item COOF1   : Cooked fiber (1-9), first evaluation
 \item COOSU1  : Cooked sugars (1-9), first evaluation
 \item COOST1  : Cooked starch (1-9), first evaluation
 \item COOT1   : Cooked taste (1-9), first evaluation
 \item COOAP1  : Cooked appearance (1-9), first evaluation
 \item FRAW2   : Root fiber (1-9), second determination
 \item SURAW2  : Root sugar (1-9), second determination
 \item STRAW2  : Root starch (1-9), second determination
 \item COOF2   : Cooked fiber (1-9), second evaluation
 \item COOSU2  : Cooked sugars (1-9), second evaluation
 \item COOST2  : Cooked starch (1-9), second evaluation
 \item COOT2   : Cooked taste (1-9), second evaluation
 \item COOAP2  : Cooked appearance (1-9), second evaluation
 \item PROT    : Protein (\%)
 \item FE      : Iron (mg/100 g dry weight)
 \item ZN      : Zinc (mg/100 g dry weight)
 \item CA      : Calcium (mg/100 g dry weight)
 \item MG      : Magnesium (mg/100 g dry weight)
 \item BC      : Beta-carotene (mg/100 g dry weight)
 \item BC.CC   : Beta-carotene with color charts
 \item TC      : Total carotenoids (mg/100 g dry weight)
 \item STAR    : Starch (\%)
 \item FRUC    : Fructose (\%)
 \item GLUC    : Glucose (\%)
 \item SUCR    : Sucrose (\%)
 \item MALT    : Maltose (\%)
 \item TRW     : Total root weight
 \item CYTHA   : Commercial root yield t/ha
 \item RYTHA   : Total root yield t/ha
 \item ACRW    : Average commercial root weight = CRW / NOCR
 \item NRPP    : Number of roots per plant
 \item YPP     : Yield per plant Kg
 \item CI      : Percent marketable roots (commercial index)
 \item HI      : Harvest index
 \item SHI     : Harvest sowing index  (survival)
 \item BIOM    : Biomass yield
 \item FYTHA   : Foliage total yield t/ha
 \item RFR     : Root foliage ratio
 }
 Extreme values are detected using the interquartile range.
 The rule is to detect any value out of the interval
 \eqn{[Q_1 - f \times IQR; Q_3 + f \times IQR]}. By default \code{f = 3}.
}
\examples{
# The data
 head(pjpz09)
 str(pjpz09)

 # Check the data
 spconsis(pjpz09, 4.5)
}
\author{
Raul Eyzaguirre.
}

