% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/CheckConsis.R
\name{spconsis}
\alias{spconsis}
\title{Check consistency for sweetpotato experimental data}
\usage{
spconsis(data, plot.size, width = 240)
}
\arguments{
\item{data}{The name of the data frame.}

\item{plot.size}{Plot size in square meters.}

\item{width}{Number of columns for the output file.}
}
\value{
It returns a file with name checks.txt with a list of
all rows with some kind of inconsistency and all rows with outliers.
}
\description{
Set of rules to check for consistency of sweetpotato experimental data.
Data labels must be defined as specified in the PROCEDURES FOR THE EVALUATION
AND ANALYSIS OF SWEETPOTATO TRIALS document.
}
\details{
The data frame must use the following labels:
\itemize{
 \item L       : Locations
 \item Y       : Year
 \item S       : Season
 \item G       : Genotypes
 \item NAME    : Names for genotypes
 \item E       : Environments
 \item R       : Replications or blocks
 \item NOPS    : Number of plants planted
 \item NOPE    : Number of plants established
 \item VIR1    : Virus symptoms (1-9), first evaluation
 \item VIR2    : Virus symptoms (1-9), second evaluation
 \item VIR3    : Virus symptoms (1-9), third evaluation
 \item ALT1    : Alternaria symptoms (1-9), first evaluation
 \item ALT2    : Alternaria symptoms (1-9), second evaluation
 \item VV1     : Vine vigor (1-9), first evaluation
 \item VV2     : Vine vigor2 (1-9), second evaluation
 \item VW      : Vine weight
 \item NOPH    : Number of plants harvested
 \item NOPR    : Number of plants with roots
 \item NOCR    : Number of commercial roots
 \item NONC    : Number of non commercial roots
 \item CRW     : Commercial root weight
 \item NCRW    : Non commercial root weight
 \item RFCP    : Root primary flesh color
 \item RFCS    : Root secondary flesh color
 \item SCOL    : Storage root skin color
 \item FCOL    : Storage root flesh color
 \item RS      : Root size (1-9)
 \item RF      : Root form (1-9)
 \item DAMR    : Root defects (1-9)
 \item RSPR    : Root sprouting (1-9)
 \item WED1    : Weevil damage (1-9), first evaluation
 \item WED2    : Weevil damage2 (1-9), second evaluation
 \item DMF     : Fresh weight of roots for dry matter assessment
 \item DMD     : Dry weight of DMF samples
 \item DMVF    : Fresh weight vines for dry matter assessment
 \item DMVD    : Dry weight of DMVF samples
 \item DM      : Storage root dry matter content (%)
 \item DMFY    : Dry matter foliage yield
 \item DMRY    : Dry matter root yield
 \item FRAW1   : Root fiber (1-9), first determination
 \item SURAW1  : Root sugar (1-9), first determination
 \item STRAW1  : Root starch (1-9), first determination
 \item COOF1   : Cooked fiber (1-9), first evaluation
 \item COOSU1  : Cooked sugars (1-9), first evaluation
 \item COOST1  : Cooked starch (1-9), first evaluation
 \item COOT1   : Cooked taste (1-9), first evaluation
 \item COOAP1  : Cooked appearance (1-9), first evaluation
 \item FRAW2   : Root fiber (1-9), second determination
 \item SURAW2  : Root sugar (1-9), second determination
 \item STRAW2  : Root starch (1-9), second determination
 \item COOF2   : Cooked fiber (1-9), second evaluation
 \item COOSU2  : Cooked sugars (1-9), second evaluation
 \item COOST2  : Cooked starch (1-9), second evaluation
 \item COOT2   : Cooked taste (1-9), second evaluation
 \item COOAP2  : Cooked appearance (1-9), second evaluation
 \item PROT    : Protein
 \item FE      : Iron in dry weight
 \item ZN      : Zinc in dry weight
 \item CA      : Calcium in dry weight
 \item MG      : Magnesium in dry weight
 \item BC      : Beta-carotene in dry weight (NIRS)
 \item BC.CC   : Beta-carotene with color charts
 \item TC      : Total carotenoids  in dry weight (NIRS)
 \item STAR    : Starch
 \item FRUC    : Fructose
 \item GLUC    : Glucose
 \item SUCR    : Sucrose
 \item MALT    : Maltose
 \item TRW     : Total root weight
 \item CYTHA   : Commercial root yield t/ ha
 \item RYTHA   : Total root yield t/ha
 \item ACRW    : Average commercial root weight = CRW/NOCR
 \item NRPP    : Number of roots per plant
 \item YPP     : Yield per plant Kg
 \item CI      : Percent marketable roots (commercial index)
 \item HI      : Harvest index
 \item SHI     : Harvest sowing index  (survival)
 \item BIOM    : Biomass yield
 \item FYTHA   : Foliage total yield t/ha
 \item RFR     : Root foliage ratio
 }
}
\examples{
# The data
 head(pjpz09)
 str(pjpz09)

 # Check the data
 spconsis(pjpz09, 4.5)
}
\author{
Raul Eyzaguirre.
}

