% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/METdata.R
\docType{data}
\name{spg}
\alias{spg}
\title{Some traits for a multi-environment trial (MET)}
\format{A data frame with 8 columns and 32 rows.}
\source{
International Potato Center, sweetpotato experimental data.
}
\usage{
spg
}
\description{
This data set has data for root yield in tons per hectare (rytha),
beta-carotene (bc), dry matter (dm), starch (star) and number of commercial roots
(nocr) for 8 genotypes in 2 locations with a RCBD with 2 blocks in each location.
}

