% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sti.R
\name{sti}
\alias{sti}
\title{Compute stress tolerance indices}
\usage{
sti(traits, geno, normal, stress, dfr)
}
\arguments{
\item{traits}{The list of traits.}

\item{geno}{The name of the column that identifies the genotypes.}

\item{normal}{The identification for the normal group.}

\item{stress}{The identification for the stress group.}

\item{dfr}{The name of the data frame.}
}
\value{
It returns a data frame with the indices and a data frame with
the stress intensity values for each trait. The names for the indices
follow the convention \code{trait.index}.
}
\description{
Compute several stress tolerance indices.
}
\details{
The names of the columns for the traits must follow the convention
\code{trait.normal} and \code{trait.stress}, so the data frame must be in
wide format for the factor with levels \code{normal} and \code{stress}.
For a trait \code{y} with values \code{yn} and \code{ys} under normal and
stress conditions, the computed indices are:
\itemize{
 \item \code{tol} : Tolerance.
 \item \code{yrr} : Yield reduction ratio.
 \item \code{ssi} : Stress susceptibility index.
 \item \code{sti} : Stress tolerance index.
 \item \code{mp}  : Mean productivity.
 \item \code{gmp} : Geometric mean productivity.
 }
}
\examples{
traits <- c("nmtp", "mtwp", "nnomtp")
sti(traits, 'genotype', 'DTWW', 'DTWS', potatostress)
}
\author{
Raul Eyzaguirre.
}
