% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Compute sum of two vectors}
\usage{
suma(v1, v2)
}
\arguments{
\item{v1}{Name of vector 1.}

\item{v2}{Name of vector 2.}
}
\value{
It returns the sum of the two vectors elementwise.
}
\description{
Compute the sum of two vectors elementwise. Missing values do not propagate.
}
\details{
Missing values do not propagate. If NA is present for both vectors then NA
is applied to the sum.
}
\examples{
suma(pjpz09$trw, pjpz09$vw)
}
\author{
Raul Eyzaguirre.
}
