% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tai.R
\name{tai}
\alias{tai}
\title{Tai's stability analysis}
\usage{
tai(trait, geno, env, rep, dfr, maxp = 0.1)
}
\arguments{
\item{trait}{The trait to analyze.}

\item{geno}{The genotypes.}

\item{env}{The environments.}

\item{rep}{The replications.}

\item{dfr}{The name of the data frame.}

\item{maxp}{Maximum allowed proportion of missing values to estimate, default is 10\%.}
}
\value{
It returns the alpha and lambda values for each genotype for the Tai
stability analysis.
}
\description{
This function runs Tai's stability analysis (Tai, G. C. C., 1971).
It assumes a RCBD with fixed effects for genotypes and random effects for environments.
}
\details{
If the data set is unbalanced, a warning is produced.
}
\examples{
model.tai <- tai("y", "geno", "env", "rep", met8x12)
model.tai$Tai_values
}
\references{
Tai, G. C. C. (1971). Genotypic Stability Analysis and Its Application to Potato
Regional Trials, Crop Science, Vol 11.
}
\author{
Raul Eyzaguirre.
}
