% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/Tai.R
\name{tai}
\alias{tai}
\title{Tai's stability analysis}
\usage{
tai(trait, geno, env, rep, data, conf = 0.95, title = NULL,
  color = c("darkorange", "black", "gray"), ...)
}
\arguments{
\item{trait}{The trait to analyze.}

\item{geno}{The genotypes.}

\item{env}{The environments.}

\item{rep}{The replications or blocks. A RCBD is assumed.}

\item{data}{The name of the data frame containing the data.}

\item{conf}{Probability for the Tai limits.}

\item{title}{Main title for plot.}

\item{color}{Color for symbols, labels and lines.}

\item{...}{Additional graphic parameters.}
}
\value{
It returns the Tai graph for stability analysis and the values of alpha
and lambda for each genotype.
}
\description{
This function runs Tai's stability analysis (Tai, G. C. C., 1971).
It assumes a RCBD with fixed effects for genotypes and random effects for environments.
}
\details{
The limits for alpha and lambda are computed using the mean squares from
an ANOVA table for a RCBD with blocks nested into environments. If the data set is
unbalanced, a warning is produced.
}
\examples{
# The data
head(met8x12)
str(met8x12)

# Run Tai for trait y
tai("y", "geno", "env", "rep", met8x12)
}
\author{
Raul Eyzaguirre
}
\references{
Tai, G. C. C. (1971). Genotypic Stability Analysis and Its Application to Potato
Regional Trials, Crop Science, Vol 11.
}

